% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_method_dispatch.R
\name{AAFE.default}
\alias{AAFE.default}
\title{Default method for AAFE()}
\usage{
\method{AAFE}{default}(obj, ...)
}
\arguments{
\item{obj}{An object}

\item{...}{Additional arguments currently not in use.}
}
\value{
An error, when a non-pk object is used for the first argument.
}
\description{
Default method for AAFE()
}
