% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_method_dispatch.R
\name{AAFE}
\alias{AAFE}
\title{AAFE()}
\usage{
AAFE(obj, ...)
}
\arguments{
\item{obj}{An object.}

\item{...}{Additional arguments currently not in use.}
}
\value{
A dataframe with one row for each `data_group`, `model` and `method`.
  The final column contains the AAFE of the model fitted by the corresponding
  method, using the data in `newdata`.
}
\description{
This is the S3 method generic for AAFE()
}
\seealso{
[AAFE.pk()] for the method for class [pk()]
}
