% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc_1comp.R
\name{auc_1comp}
\alias{auc_1comp}
\title{Analytic AUC for 1-compartment model}
\usage{
auc_1comp(params, time, dose, route, medium)
}
\arguments{
\item{params}{A named list of model parameter values. For either IV or oral
dosing, this must include `kelim` (elimination rate, 1/h). For IV dosing, it
must also include `Vdist`. For oral dosing, it must include `kgutabs` (oral
absorption rate, 1/h), and either `Fgutabs` and `Vdist` (respectively,
unitless fraction of oral dose that is absorbed, and volume of
distribution), or `Fgutabs_Vdist` (the ratio of Fgutabs to Vdist).
`Fgutabs_Vdist` is an alternate parameterization useful for parameter
estimation when only oral data are available with no IV data, in which case
only the ratio of `Fgutabs` to `Vdist` is identifiable. if `Fgutabs` and
`Vdist` are provided along with `Fgutabs_Vdist`, then `Fgutabs_Vdist` will
not be used.}

\item{time}{A numeric vector of times in hours.}

\item{dose}{A numeric vector of doses in mg/kg}

\item{route}{A logical vector: TRUE for single IV bolus dose; FALSE for single oral
dose}
}
\value{
A vector of plasma AUC values (mg/L*time) corresponding to `time`.
}
\description{
Calculate area under the plasma concentration vs. time curve for the
1-compartment model, using an analytical equation (the integral of the
1-compartment model equation with respect to time).
}
\author{
Caroline Ring, John Wambaugh
}
