% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_faceted_methods_BIC.R, R/pk_methods_BIC.R
\name{BIC.pk}
\alias{BIC.pk}
\title{BIC}
\usage{
\method{BIC}{pk}(obj, newdata = NULL, model = NULL, method = NULL, exclude = TRUE)

\method{BIC}{pk}(obj, newdata = NULL, model = NULL, method = NULL, exclude = TRUE)
}
\arguments{
\item{obj}{A `pk` object}

\item{newdata}{Optional: A `data.frame` with new data for which to compute
log-likelihood. If NULL (the default), then BICs will be
computed for the data in `obj$data`. `newdata` is required to contain at
least the following variables: `Time`, `Time.Units`, `Dose`, `Route`,`Media`, `Conc`,
`Detect`, `N_Subjects`. Before log-likelihood is calculated, `Time` will be
transformed according to the transformation in `obj$scales$time` and `Conc`
will be transformed according to the transformation in `obj$scales$conc`.}

\item{model}{Optional: Specify one or more of the fitted models for which to
calculate BIC. If NULL (the default), log-likelihoods will be
returned for all of the models in `obj$stat_model`.}

\item{method}{Optional: Specify one or more of the [optimx::optimx()] methods
for which to calculate BICs. If NULL (the default),
log-likelihoods will be returned for all of the methods in
`obj$settings_optimx$method`.}

\item{exclude}{Logical: `TRUE` to compute the AIC after removing any
observations in the data marked for exclusion (if there is a variable
`exclude` in the data, an observation is marked for exclusion when `exclude
%in% TRUE`). `FALSE` to include all observations, regardless of exclusion
status. Default `TRUE`.}
}
\value{
A [tibble::tibble()] grouped by the faceting variables with a new variable ` BIC ` containing the results of applying [BIC.pk()] to each item in the list column of [pk()] objects, `obj$pk_object`

A named list of numeric vectors. There is one list element
  named for each model in `obj`'s [stat_model()] element, i.e. each PK model
  that was fitted to the data. Each list element is a numeric vector with as
  many elements as there were [optimx::optimx()] methods (specified in
  [settings_optimx()]). The vector names are the method names.  Each vector
  element contains the BIC of the model fitted by the corresponding method,
  using the data in `newdata`.
}
\description{
BIC for pk_faceted objects

Get the Bayesian information criterion (AIC) for a fitted `pk` object
}
\details{
The BIC is calculated from the log-likelihood (LL) as follows:
\deqn{\textrm{BIC} = -2\textrm{LL} + \log(n_{obs}) n_{par}}

where \eqn{n_{par}} is the number of parameters in the fitted model.

Note that the BIC is just the AIC with \eqn{k = \(n_{obs})}.
}
\seealso{
Other methods for pk_faceted objects: 
\code{\link{AIC.pk}()},
\code{\link{check_needed_status.pk}()},
\code{\link{coef.pk}()},
\code{\link{coef_sd.pk}()},
\code{\link{compare_models.pk}()},
\code{\link{data_summary.pk}()},
\code{\link{eval_tkstats.pk}()},
\code{\link{fold_error.pk}()},
\code{\link{get_data.pk}()},
\code{\link{get_data_group.pk}()},
\code{\link{get_data_info.pk}()},
\code{\link{get_data_original.pk}()},
\code{\link{get_data_sigma_group.pk}()},
\code{\link{get_error_group.pk}()},
\code{\link{get_fit.pk}()},
\code{\link{get_mapping.pk}()},
\code{\link{get_nca.pk}()},
\code{\link{get_prefit.pk}()},
\code{\link{get_scale_conc.pk}()},
\code{\link{get_scale_time.pk}()},
\code{\link{get_settings_data_info.pk}()},
\code{\link{get_settings_optimx.pk}()},
\code{\link{get_settings_preprocess.pk}()},
\code{\link{get_stat_error_model.pk}()},
\code{\link{get_stat_model.pk}()},
\code{\link{get_status.pk}()},
\code{\link{get_tkstats.pk}()},
\code{\link{get_winning_model.pk}()},
\code{\link{is.pk}()},
\code{\link{logLik.pk}()},
\code{\link{nca.pk}()},
\code{\link{plot.pk}()},
\code{\link{predict.pk}()},
\code{\link{preprocess_data.pk_faceted}()},
\code{\link{residuals.pk}()},
\code{\link{rmse.pk}()},
\code{\link{rsq.pk}()},
\code{\link{subtract_pkproto.pk}()},
\code{\link{summary.pk}()}

Other fit evaluation metrics: 
\code{\link{AIC.pk}()},
\code{\link{logLik.pk}()},
\code{\link{rmse.pk}()},
\code{\link{rsq.pk}()}

Other log likelihood functions: 
\code{\link{AIC.pk}()},
\code{\link{logLik.pk}()}

Other methods for fitted pk objects: 
\code{\link{AIC.pk}()},
\code{\link{coef.pk}()},
\code{\link{coef_sd.pk}()},
\code{\link{eval_tkstats.pk}()},
\code{\link{get_fit.pk}()},
\code{\link{get_tkstats.pk}()},
\code{\link{logLik.pk}()},
\code{\link{predict.pk}()},
\code{\link{residuals.pk}()},
\code{\link{rmse.pk}()},
\code{\link{rsq.pk}()}
}
\author{
Caroline Ring
}
\concept{fit evaluation metrics}
\concept{log likelihood functions}
\concept{methods for fitted pk objects}
\concept{methods for pk_faceted objects}
