% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{case_match_lst}
\alias{case_match_lst}
\title{Wrapper for [dplyr::case_match()] that accepts named lists as arguments}
\usage{
case_match_lst(.x, .lst, .default = NULL)
}
\arguments{
\item{.x}{A vector to match against. In a data context, a column.}

\item{.lst}{A named list in the form `old_value = new_value`.}

\item{.default}{Value to be used when there are no values. Suggest setting it to `.x`}
}
\value{
A vector of replaced values
}
\description{
Wrapper for [dplyr::case_match()] that accepts named lists as arguments
}
