% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_group_hierarchy.R
\name{check_group_hierarchy}
\alias{check_group_hierarchy}
\title{Checking data, error, and summary group hierarchical structure}
\usage{
check_group_hierarchy(obj)
}
\arguments{
\item{obj}{An object created by [pk()].}
}
\value{
Prints a tree of the summary hierarchies and an error when
  the hierarchical structure expectation is not met.
}
\description{
Checking data, error, and summary group hierarchical structure
}
