% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_method.R
\name{check_method}
\alias{check_method}
\title{Check methods}
\usage{
check_method(obj, method)
}
\arguments{
\item{obj}{A [pk()] object}

\item{method}{A user-supplied `character` vector of method names}
}
\value{
`TRUE` if all `method %in% obj$settings_optimx$method`; otherwise stops with an error
}
\description{
Check methods for validity
}
\details{
Helper function to ensure that a list of methods specified by the user matches the methods available in the fitted [pk()] object.
}
\author{
Caroline Ring
}
