% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_method.R
\name{check_model}
\alias{check_model}
\title{Check models}
\usage{
check_model(obj, model)
}
\arguments{
\item{obj}{A [pk()] object}

\item{model}{A user-supplied `character` vector of model names}
}
\value{
`TRUE` if all `model %in% names(obj$stat_model)`; otherwise stops with an error
}
\description{
Check models for validity
}
\details{
Helper function to ensure that a list of models specified by the user matches the models available in the fitted [pk()] object.
}
\author{
Caroline Ring
}
