% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_params_flat.R
\name{check_params_flat}
\alias{check_params_flat}
\title{Check flat model parameters}
\usage{
check_params_flat(params, route, medium, ...)
}
\arguments{
\item{params}{A named numeric vector of parameters for the flat model.}

\item{route}{A character vector of routes: "iv" and/or "oral".}

\item{medium}{A character vector of tissue media: "plasma" and/or "blood".}

\item{...}{Additional arguments (not currently used)}
}
\value{
Character: A message. If all required parameters are present for the
  given media & routes, the message is "Parameters OK". If required
  parameters for the oral route are missing, the message is "Error: For
  flat oral model, missing parameters (comma-separated list of
  parameter names)".  If required
  parameters for the IV route are missing, the message is "Error: For
  flat oral model, missing parameters (comma-separated list of
  parameter names)".
}
\description{
Check to make sure required parameters are present to evaluate flat
model for a given route and medium
}
\author{
Caroline Ring
}
