% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_method_dispatch.R
\name{check_required_status.default}
\alias{check_required_status.default}
\title{Default method for checking required status}
\usage{
\method{check_required_status}{default}(obj, ...)
}
\description{
Default method for checking required status
}
