% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_method_dispatch.R
\name{check_required_status}
\alias{check_required_status}
\title{Check required status}
\usage{
check_required_status(obj, ...)
}
\arguments{
\item{obj}{An object.}
}
\description{
This is the S3 method generic.
}
\seealso{
[check_required_status.pk()] for the method for class [pk()]
}
