% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_faceted_methods_coef.R, R/pk_methods_coef.R
\name{coef.pk}
\alias{coef.pk}
\title{coef}
\usage{
\method{coef}{pk}(obj, model = NULL, method = NULL, ...)

\method{coef}{pk}(obj, model = NULL, method = NULL, ...)
}
\arguments{
\item{obj}{A [pk()] object}

\item{model}{Optional: Specify (as a `character` vector) one or more of the
fitted models whose coefficients to return. If `NULL` (the default),
coefficients will be returned for all of the models in `obj$stat_model`.}

\item{method}{Optional: Specify (as a `character` vector)one or more of the
[optimx::optimx()] methods whose coefficients to return. If `NULL` (the
default), coefficients will be returned for all of the models in
`obj$settings_optimx$method`.}
}
\value{
A [tibble::tibble()] grouped by the faceting variables with a new variable ` coef ` containing the results of applying [coef.pk()] to each item in the list column of [pk()] objects, `obj$pk_object`

A named list of numeric matrixes. There is one list element named for
  each model in `model`. Each list element is a matrix with as many rows as
  items in `method`. The row names are the method names. The matrix column
  names are the names of the fitted parameters, including any error standard
  deviation hyperparameters (whose names begin with "sigma"). The matrix
  elements are the values of the corresponding model parameters as fitted by
  the corresponding method.
}
\description{
coef for pk_faceted objects

Extract coefficients from a fitted [pk()] object
}
\details{
This function extracts fitted model parameter values from a fitted [pk()]
object.
}
\seealso{
Other methods for pk_faceted objects: 
\code{\link{AIC.pk}()},
\code{\link{BIC.pk}()},
\code{\link{check_needed_status.pk}()},
\code{\link{coef_sd.pk}()},
\code{\link{compare_models.pk}()},
\code{\link{data_summary.pk}()},
\code{\link{eval_tkstats.pk}()},
\code{\link{fold_error.pk}()},
\code{\link{get_data.pk}()},
\code{\link{get_data_group.pk}()},
\code{\link{get_data_info.pk}()},
\code{\link{get_data_original.pk}()},
\code{\link{get_data_sigma_group.pk}()},
\code{\link{get_error_group.pk}()},
\code{\link{get_fit.pk}()},
\code{\link{get_mapping.pk}()},
\code{\link{get_nca.pk}()},
\code{\link{get_prefit.pk}()},
\code{\link{get_scale_conc.pk}()},
\code{\link{get_scale_time.pk}()},
\code{\link{get_settings_data_info.pk}()},
\code{\link{get_settings_optimx.pk}()},
\code{\link{get_settings_preprocess.pk}()},
\code{\link{get_stat_error_model.pk}()},
\code{\link{get_stat_model.pk}()},
\code{\link{get_status.pk}()},
\code{\link{get_tkstats.pk}()},
\code{\link{get_winning_model.pk}()},
\code{\link{is.pk}()},
\code{\link{logLik.pk}()},
\code{\link{nca.pk}()},
\code{\link{plot.pk}()},
\code{\link{predict.pk}()},
\code{\link{preprocess_data.pk_faceted}()},
\code{\link{residuals.pk}()},
\code{\link{rmse.pk}()},
\code{\link{rsq.pk}()},
\code{\link{subtract_pkproto.pk}()},
\code{\link{summary.pk}()}

Other methods for fitted pk objects: 
\code{\link{AIC.pk}()},
\code{\link{BIC.pk}()},
\code{\link{coef_sd.pk}()},
\code{\link{eval_tkstats.pk}()},
\code{\link{get_fit.pk}()},
\code{\link{get_tkstats.pk}()},
\code{\link{logLik.pk}()},
\code{\link{predict.pk}()},
\code{\link{residuals.pk}()},
\code{\link{rmse.pk}()},
\code{\link{rsq.pk}()}
}
\author{
Caroline Ring
}
\concept{methods for fitted pk objects}
\concept{methods for pk_faceted objects}
