% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_method_dispatch.R
\name{coef_sd.default}
\alias{coef_sd.default}
\title{Coefficient standard deviation default}
\usage{
\method{coef_sd}{default}(obj, ...)
}
\arguments{
\item{obj}{An object}
}
\description{
Coefficient standard deviation default
}
