% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_method_dispatch.R
\name{coef_sd}
\alias{coef_sd}
\title{Coefficient standard deviations}
\usage{
coef_sd(obj, model, method, suppress.messages, ...)
}
\arguments{
\item{obj}{An object}

\item{model}{The TK model used.}

\item{method}{Optimizer method used.}

\item{suppress.messages}{Boolean. Whether messages will be printed.}

\item{...}{Additional arguments currently not in use.}
}
\description{
This is the S3 method generic for `coef_sd`.
}
\seealso{
[coef_sd.pk()] for the `coef_sd` method for class [pk()]
}
