% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combined_mean.R
\name{combined_mean}
\alias{combined_mean}
\title{Combined mean}
\usage{
combined_mean(group_mean, group_n, na.rm = TRUE, log = FALSE)
}
\arguments{
\item{group_mean}{Numeric vector: Observed sample means for summary data, or
observed values for non-summary data. Censored observations should *not* be
NA; they should be substituted with some value at or below the
corresponding LOQ (e.g. LOQ or LOQ/2). Even if `log %in% TRUE`, these
should *not* be log-transformed.}

\item{group_n}{Numeric vector: Observed sample number of subjects for summary
data. For non-summary data (individual-subject observations), `group_n`
should be set to 1.}

\item{na.rm}{Logical. If TRUE (default), then any groups where mean *or*
N were NA will be dropped. If FALSE, they will be retained (and the result
will be NA).}
}
\value{
Numeric: the mean of the combined population (i.e. if
  all the groups were concatenated into one large group).
}
\description{
Given mean, standard deviation, and N for some set of groups, calculate the
combined mean. Note that the groups may not overlap.
}
\author{
Caroline Ring
}
