% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_faceted_methods_compare_models.R
\name{compare_models.pk_faceted}
\alias{compare_models.pk_faceted}
\title{compare_models}
\usage{
\method{compare_models}{pk_faceted}(obj, ...)
}
\arguments{
\item{obj}{An object of class `pk_faceted`}
}
\value{
A [tibble::tibble()] grouped by the faceting variables with a new variable ` compare_models ` containing the results of applying [compare_models.pk()] to each item in the list column of [pk()] objects, `obj$pk_object`
}
\description{
compare_models for pk_faceted objects
}
\seealso{
Other methods for pk_faceted objects: 
\code{\link{AIC.pk_faceted}()},
\code{\link{BIC.pk_faceted}()},
\code{\link{check_needed_status.pk_faceted}()},
\code{\link{coef.pk_faceted}()},
\code{\link{coef_sd.pk_faceted}()},
\code{\link{data_summary.pk_faceted}()},
\code{\link{eval_tkstats.pk_faceted}()},
\code{\link{fold_error.pk_faceted}()},
\code{\link{get_data.pk_faceted}()},
\code{\link{get_data_group.pk_faceted}()},
\code{\link{get_data_info.pk_faceted}()},
\code{\link{get_data_original.pk_faceted}()},
\code{\link{get_data_sigma_group.pk_faceted}()},
\code{\link{get_error_group.pk_faceted}()},
\code{\link{get_fit.pk_faceted}()},
\code{\link{get_mapping.pk_faceted}()},
\code{\link{get_nca.pk_faceted}()},
\code{\link{get_prefit.pk_faceted}()},
\code{\link{get_scale_conc.pk_faceted}()},
\code{\link{get_scale_time.pk_faceted}()},
\code{\link{get_settings_data_info.pk_faceted}()},
\code{\link{get_settings_optimx.pk_faceted}()},
\code{\link{get_settings_preprocess.pk_faceted}()},
\code{\link{get_stat_error_model.pk_faceted}()},
\code{\link{get_stat_model.pk_faceted}()},
\code{\link{get_status.pk_faceted}()},
\code{\link{get_tkstats.pk_faceted}()},
\code{\link{get_winning_model.pk_faceted}()},
\code{\link{is.pk_faceted}()},
\code{\link{logLik.pk_faceted}()},
\code{\link{nca.pk_faceted}()},
\code{\link{plot.pk_faceted}()},
\code{\link{predict.pk_faceted}()},
\code{\link{preprocess_data.pk_faceted}()},
\code{\link{residuals.pk_faceted}()},
\code{\link{rmse.pk_faceted}()},
\code{\link{rsq.pk_faceted}()},
\code{\link{subtract_pkproto.pk_faceted}()},
\code{\link{summary.pk_faceted}()}
}
\author{
Caroline Ring
}
\concept{methods for pk_faceted objects}
