% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_method_dispatch.R
\name{compare_models}
\alias{compare_models}
\title{Model comparison}
\usage{
compare_models(obj, ...)
}
\arguments{
\item{obj}{An object.}

\item{...}{Additional arguments currently not in use.}
}
\value{
A `data.frame` with variables
- `model`: The name of each model
- `method`: The name of each method
- A variable named for `criterion` (e.g. if `criterion = "AIC"` then the result will have a
  variable named `AIC`): The criterion value for each model/method
}
\description{
This is the S3 method generic for compare_models()
}
\seealso{
[compare_models.pk()] for the method for class [pk()]
}
