% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cp_1comp_rad.R
\name{cp_1comp_rad}
\alias{cp_1comp_rad}
\title{Analytical 1-compartment model for total radiolabeled material}
\usage{
cp_1comp_rad(params, time, dose, route, medium = "plasma", restrictive = TRUE)
}
\arguments{
\item{params}{A named numeric vector of model parameter values. See Details for
requirements.}

\item{time}{A numeric vector of times, reflecting the time point when
concentration is measured after the corresponding single bolus dose. Must be
same length as `dose` and `iv.dose`, or length 1.}

\item{dose}{A numeric vector of doses, reflecting single bolus doses
administered at time 0. Must be same length as `time` and `iv.dose`, or
length 1. In this model, it is expected that this value represents a measurement
of radioactive particles from a radiolabeling experiment.}

\item{route}{A character vector, reflecting the route of administration of
each single bolus dose: `'oral'` or `'iv'`.  Must be same length as `time`
and `dose`, or length 1.}

\item{medium}{A character vector reflecting the medium in which each resulting
concentration is to be calculated: "blood" or "plasma". Default is "plasma".
Must be same length as `time` and `dose`, or length 1.}

\item{restrictive}{A logical value (TRUE or FALSE. Default: FALSE) that says whether the
assumption is that the clearance is restrictive or non-restrictive}
}
\value{
A vector of blood or plasma concentration values  corresponding
 to `time`.
}
\description{
Calculates plasma concentrations vs. time according to the analytical solution
for the 1-compartment model, for single bolus doses (IV and/or oral).
}
\section{Required parameters}{


`params` must include the following named items:
  \describe{
  \item{Fup}{Fraction of compound unbound in plasma. Unitless.}
  \item{Clint}{Intrinsic clearance by hepatocytes. Units: 1/hr}
  \item{Q_totli}{Total blood flow through the liver. Units: L/h/kg body weight ^ (3/4)}
  \item{Q_gfr}{Glomerular filtration rate, how quickly do kidneys filter. Units: L/h/kg body weight ^ (3/4)}
  \item{Vdist}{Apparent volume of central compartment, volume/unit BW.}
  }

For oral administration (if any `route %in% "oral"`), `params` must also
include:
  \describe{
  \item{Fgutabs}{Oral bioavailability, unitless fraction.}
  \item{kgutabs}{Rate of absorption from gut, 1/time.}
  }

For excreta (if any `media %in% "excreta"`), `params` must also include:
  \describe{
  \item{Frec}{Fraction recovered in excreta. This is a precalculated value.}
  }

If `any(medium %in% 'blood')`, then `params` must also include
`Rblood2plasma`, the ratio of chemical concentration in whole blood to the
chemical concentration in blood plasma.
}

\seealso{
Other built-in model functions: 
\code{\link{auc_1comp}()},
\code{\link{auc_1comp_cl}()},
\code{\link{auc_1comp_rad}()},
\code{\link{auc_2comp}()},
\code{\link{auc_flat}()},
\code{\link{cp_1comp}()},
\code{\link{cp_1comp_cl}()},
\code{\link{cp_2comp}()},
\code{\link{cp_2comp_dt}()},
\code{\link{cp_flat}()},
\code{\link{cp_httk_3comp2}()},
\code{\link{get_params_1comp}()},
\code{\link{get_params_1comp_cl}()},
\code{\link{get_params_1comp_fup}()},
\code{\link{get_params_1comp_rad}()},
\code{\link{get_params_2comp}()},
\code{\link{get_params_flat}()},
\code{\link{get_params_httk_3comp2}()},
\code{\link{get_starts_1comp}()},
\code{\link{get_starts_1comp_cl}()},
\code{\link{get_starts_1comp_rad}()},
\code{\link{get_starts_2comp}()},
\code{\link{get_starts_flat}()},
\code{\link{get_starts_httk_3comp2}()},
\code{\link{tkstats_2comp}()},
\code{\link{transformed_params_2comp}()}

Other 1-compartment radiation model functions: 
\code{\link{get_params_1comp_rad}()},
\code{\link{get_starts_1comp_rad}()}

Other model concentration functions: 
\code{\link{cp_1comp}()},
\code{\link{cp_1comp_cl}()},
\code{\link{cp_2comp}()},
\code{\link{cp_flat}()},
\code{\link{cp_httk_3comp2}()},
\code{\link{get_params_httk_3comp2}()}
}
\author{
Caroline Ring, John Wambaugh, Gilberto Padilla Mercado
}
\concept{1-compartment radiation model functions}
\concept{built-in model functions}
\concept{model concentration functions}
