% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cp_httk_3comp2.R
\name{cp_httk_3comp2}
\alias{cp_httk_3comp2}
\title{Wrapper for `httk`'s `3compartment2` PBPK model}
\usage{
cp_httk_3comp2(
  params,
  time,
  dose,
  route,
  medium = "plasma",
  this_chem = NULL,
  this_species = NULL,
  restrictive = TRUE
)
}
\arguments{
\item{params}{A named numeric vector of model parameter values.}

\item{time}{A numeric vector of times, reflecting the time point when
concentration is measured after the corresponding single bolus dose. Must be
same length as `dose` and `iv.dose`, or length 1.}

\item{dose}{A numeric vector of doses, reflecting single bolus doses
administered at time 0. Must be same length as `time` and `iv.dose`, or
length 1. In this model, it is expected that this value represents a measurement
of radioactive particles from a radiolabeling experiment.}

\item{route}{A character vector, reflecting the route of administration of
each single bolus dose: `'oral'` or `'iv'`.  Must be same length as `time`
and `dose`, or length 1.}

\item{medium}{A character vector reflecting the medium in which each resulting
concentration is to be calculated: "blood" or "plasma". Default is "plasma".
Must be same length as `time` and `dose`, or length 1.}

\item{this_chem}{A character vector naming the chemical for calculations in `httk`.}

\item{this_species}{A character vector naming the species for calculations in `httk`.}

\item{restrictive}{A logical value (TRUE or FALSE. Default: FALSE) that says whether the
assumption is that the clearance is restrictive or non-restrictive}
}
\value{
A vector of blood or plasma concentration values  corresponding
 to `time`
 .
}
\description{
Calculated plasma concentrations vs time according to the `3compartment2`
model (Wambaugh, Schacht, and Ring. 2025).
}
\section{Required parameters}{

These are given by the \link[httk]{parameterize_3comp2} function in `httk`.
Furthermore, they are transformed to a vector during the prefitting process.
The optimized parameters are `Clint` and `Funbound.plasma`.
}

\section{Process of calculation}{

The calculated plasma concentrations
}

\seealso{
Other built-in model functions: 
\code{\link{auc_1comp}()},
\code{\link{auc_1comp_cl}()},
\code{\link{auc_1comp_rad}()},
\code{\link{auc_2comp}()},
\code{\link{auc_flat}()},
\code{\link{cp_1comp}()},
\code{\link{cp_1comp_cl}()},
\code{\link{cp_1comp_rad}()},
\code{\link{cp_2comp}()},
\code{\link{cp_2comp_dt}()},
\code{\link{cp_flat}()},
\code{\link{get_params_1comp}()},
\code{\link{get_params_1comp_cl}()},
\code{\link{get_params_1comp_fup}()},
\code{\link{get_params_1comp_rad}()},
\code{\link{get_params_2comp}()},
\code{\link{get_params_flat}()},
\code{\link{get_params_httk_3comp2}()},
\code{\link{get_starts_1comp}()},
\code{\link{get_starts_1comp_cl}()},
\code{\link{get_starts_1comp_fup}()},
\code{\link{get_starts_1comp_rad}()},
\code{\link{get_starts_2comp}()},
\code{\link{get_starts_flat}()},
\code{\link{get_starts_httk_3comp2}()},
\code{\link{tkstats_2comp}()},
\code{\link{transformed_params_2comp}()}

Other httk model functions: 
\code{\link{get_params_httk_3comp2}()},
\code{\link{get_starts_httk_3comp2}()}

Other model concentration functions: 
\code{\link{cp_1comp}()},
\code{\link{cp_1comp_cl}()},
\code{\link{cp_1comp_rad}()},
\code{\link{cp_2comp}()},
\code{\link{cp_flat}()},
\code{\link{get_params_httk_3comp2}()}
}
\author{
Gilberto Padilla Mercado
}
\concept{built-in model functions}
\concept{httk model functions}
\concept{model concentration functions}
