% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cvt}
\alias{cvt}
\title{CvTdb data}
\format{
A `data.frame` with 13937 rows and 61 variables:
\describe{
  \item{conc_time_id}{Unique database identifier for each CvT observation.}
  \item{fk_series_id}{Unique database identifier for experimental series.}
  \item{time_original}{Timepoint in original units.}
  \item{time_hr}{Timepoint in hours.}
  \item{conc_original}{Concentration in original units.}
  \item{conc_sd_original}{Standard deviation of concentration in original units.}
  \item{conc}{Concentration in normalized units.}
  \item{conc_sd}{Standard deviation of concentration in normalized units.}
  \item{fk_analyzed_chemical_id}{Unique database identifier for analyte.}
  \item{analyzed_chem_dtxsid}{DTXSID of chemical analyte.}
  \item{analyzed_chem_name_original}{Original analyte name.}
  \item{analyzed_chem_casrn}{CASRN of chemical analyte.}
  \item{analyzed_chem_name}{Preferred name for analyte.}
  \item{time_units_original}{Original time units.}
  \item{conc_units_original}{Original concentration units.}
  \item{conc_units_normalized}{Normalized concentration units.}
  \item{conc_unit_norm_factor}{Ratio of conc/conc_original}
  \item{loq}{Level of quantification.}
  \item{loq_units}{Units for loq.}
  \item{n_subjects_in_series}{Number of subjects in each series.}
  \item{radiolabeled}{Answers whether this observation comes from a radiolabelling or isotope tracing experiment.}
  \item{fk_study_id}{Unique database identifier for each study.}
  \item{administration_route_normalized}{Route of exposure/administration, either oral or iv.}
  \item{fk_dosed_chemical_id}{Unique database identifier for dosed chemical.}
  \item{dosed_chem_dtxsid}{DTXSID of dosed chemical.}
  \item{dosed_chem_name_original}{Original dosed chemical name.}
  \item{dosed_chem_casrn}{CASRN of dosed chemical.}
  \item{dosed_chem_name}{Preferred name for dosed chemical.}
  \item{dose_volume}{Volume of dose.}
  \item{dose_volume_units}{Units for dose_volume.}
  \item{dose_vehicle}{If available, specifies what vehicle was used CvT experiment.}
  \item{dose_duration}{Duration of the dose, if available.}
  \item{dose_duration_units}{Units for dose_duration.}
  \item{dose_frequency}{Frequency of dosing, these should all be 1 for a single bolus.}
  \item{fasting_period}{If available, describes the fasting period for subjects.}
  \item{dose_level_normalized}{Dose levels in normalized units.}
  \item{dose_level_original}{Dose levels in original units.}
  \item{dose_level_units_original}{Units for dose_level_original.}
  \item{conc_medium_normalized}{Standardized media names, blood or plasma.}
  \item{conc_medium_original}{Original media names.}
  \item{fk_subject_id}{Unique database identifier for each subject.}
  \item{weight_kg}{Subject weight, in kilograms.}
  \item{species}{Subject species.}
  \item{sex}{Subject sex, if available.}
  \item{age}{Subject age, if available.}
  \item{age_units}{Units for age.}
  \item{age_category}{Categories for age.}
  \item{fk_extraction_document_id}{Unique database identifier for documents that were curated.}
  \item{pmid}{Document PubMed ID.}
  \item{year}{Year of publication.}
  \item{other_study_identifier}{Alternative identifier for documents, used for NTP studies.}
  \item{url}{Document URL address.}
  \item{doi}{Document DOI.}
  \item{extracted}{Curation level of document.}
  \item{curation_set_tag}{A grouping tag for specific document extraction and curation efforts.}
  \item{n_subjects_normalized}{Normalized subject number.}
  \item{invivPK_dose_level_units}{dose_level_units used in this package, mg/kg.}
  \item{invivPK_conc_units}{conc_units used in this package, ug/mL}
  \item{invivPK_conc}{Concentrations normalized to ug/mL}
  \item{invivPK_dose_level}{Dose normalized to mg/kg}
  \item{invivPK_loq}{Level of quantification in ug/mL}
  \item{invivPK_loq_units}{LOQ units used in this package, ug/mL.}
  \item{invivPK_conc_sd}{Standard deviation of concentrations, in ug/mL.}
}
}
\usage{
cvt
}
\description{
Concentration vs. time data from CvTdb
}
\details{
This is concentration vs. time data from CvTdb, most recently downloaded as
of the date in [`cvt_date`].

These data have been filtered to retain only oral and intravenous
administration, and only measurements in blood and plasma. They have also
been filtered to retain only observations where the same chemical was both
administered and measured in blood/plasma (i.e., excluding observations where
a metabolite was measured).
}
\keyword{datasets}
