% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cvt_2.0.0}
\alias{cvt_2.0.0}
\title{CvTdb data for invivoPKfit 2.0.0 release (old)}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 7918 rows and 62 columns.
}
\usage{
cvt_2.0.0
}
\description{
The CvTdb data released for the manuscript Informatics for toxicokinetics (2025).
}
\details{
A data.frame with similar data to [`cvt`]
}
\keyword{datasets}
