% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cvt_date}
\alias{cvt_date}
\title{CvTdb download date}
\format{
An object of class \code{Date} of length 1.
}
\usage{
cvt_date
}
\description{
The most recent download date of [`cvt`] data
}
\details{
A character scalar giving the date in "YYYY-MM-DD" format of the download
date of the data in [`cvt`] from the CvTdb database.
}
\keyword{datasets}
