% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_methods_data_info.R
\name{data_info.pk}
\alias{data_info.pk}
\title{Get summary data info and NCA}
\usage{
\method{data_info}{pk}(obj)
}
\arguments{
\item{obj}{An object of class [pk()]}
}
\description{
Get summary data info and NCA
}
\author{
Caroline Ring
}
