% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_faceted_methods_data_info.R
\name{data_info.pk_faceted}
\alias{data_info.pk_faceted}
\title{calculate summary data info for a `pk_faceted` object}
\usage{
\method{data_info}{pk_faceted}(obj)
}
\arguments{
\item{obj}{An object of class `pk_faceted``}
}
\description{
Calculate summary data information, including non-compartmental analysis.
}
\author{
Caroline Ring
}
