% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_method_dispatch.R
\name{data_summary.default}
\alias{data_summary.default}
\title{Default method for data_summary()}
\usage{
\method{data_summary}{default}(obj, ...)
}
\arguments{
\item{obj}{An object}
}
\description{
Default method for data_summary()
}
