% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_faceted_methods_data_summary.R,
%   R/pk_methods_data_summary.R
\name{data_summary.pk}
\alias{data_summary.pk}
\title{data_summary}
\usage{
\method{data_summary}{pk}(obj, newdata = NULL, summary_group = NULL)

\method{data_summary}{pk}(obj, newdata = NULL, summary_group = NULL)
}
\arguments{
\item{obj}{A [pk()] model object. Must be fitted, or the function will exit
with an error.}

\item{newdata}{Optional: A `data.frame` containing new data for which to
compute the TK stats. Must contain at least variables `Chemical`,
`Species`, `Route`, `Dose`, `Conc`, `Dose.Units`, `Conc.Units`, either
`Time_trans.Units` or `Time.Units`, and any other variables named in
`tk_grouping`. Default `NULL`, to use the data in `get_data(obj)`.}

\item{summary_group}{A list of variables provided using a `dplyr::vars()`
call. The data (either `newdata` or `obj$data`) will be grouped according
to the unique combinations of these variables. For each unique combination
of these variables in the data, a set of summary statistics will be
computed. The default is `NULL`, to use the same data grouping that was set
in [stat_nca()] for the `pk` object. However, you may specify a different
data grouping if you wish.}
}
\value{
A [tibble::tibble()] grouped by the faceting variables with a new variable ` data_summary ` containing the results of applying [data_summary.pk()] to each item in the list column of [pk()] objects, `obj$pk_object`

A `data.frame` with variables including all the grouping variables in
  `summary_group`, `group_id`; `param_name` (the name of the summary
  statistic; see Details); `param_value` (the summary statistic value);  `param_units`
  (the units of the summary statistic, derived from the units of the data).
}
\description{
data_summary for pk_faceted objects

Calculate data summary statistics for a `pk` object
}
\details{
Get summary statistics for data in a `pk` object (or optionally, new data),
using data groupings defined by `get_nca_group()` for the `pk` object (or
optionally, new groupings). If you provide both `newdata` and
`summary_group`, then everything in the `pk` object will be ignored and you
will simply be doing data summary *de novo* (which may be what you want).

Summary statistics include, for each group:

 -`n_obs`: the number of observations
- `n_exclude`: The number of excluded observations
- `n_detect`: The number of non-excluded detected observations
- `n_series_id`: The number of unique series IDs
- `n_timepts`: The number of unique time points
- `n_ref`: The number of unique reference IDs
- `tlast`: The time of the latest non-excluded observation
- `tlast_detect`: The time of the latest non-excluded detected observation
- `tfirst`: The time of the earliest non-excluded observation
- `tfirst_detect`: The time of the earliest non-excluded detected observation
}
\seealso{
Other methods for pk_faceted objects: 
\code{\link{AIC.pk}()},
\code{\link{BIC.pk}()},
\code{\link{check_needed_status.pk}()},
\code{\link{coef.pk}()},
\code{\link{coef_sd.pk}()},
\code{\link{compare_models.pk}()},
\code{\link{eval_tkstats.pk}()},
\code{\link{fold_error.pk}()},
\code{\link{get_data.pk}()},
\code{\link{get_data_group.pk}()},
\code{\link{get_data_info.pk}()},
\code{\link{get_data_original.pk}()},
\code{\link{get_data_sigma_group.pk}()},
\code{\link{get_error_group.pk}()},
\code{\link{get_fit.pk}()},
\code{\link{get_mapping.pk}()},
\code{\link{get_nca.pk}()},
\code{\link{get_prefit.pk}()},
\code{\link{get_scale_conc.pk}()},
\code{\link{get_scale_time.pk}()},
\code{\link{get_settings_data_info.pk}()},
\code{\link{get_settings_optimx.pk}()},
\code{\link{get_settings_preprocess.pk}()},
\code{\link{get_stat_error_model.pk}()},
\code{\link{get_stat_model.pk}()},
\code{\link{get_status.pk}()},
\code{\link{get_tkstats.pk}()},
\code{\link{get_winning_model.pk}()},
\code{\link{is.pk}()},
\code{\link{logLik.pk}()},
\code{\link{nca.pk}()},
\code{\link{plot.pk}()},
\code{\link{predict.pk}()},
\code{\link{preprocess_data.pk_faceted}()},
\code{\link{residuals.pk}()},
\code{\link{rmse.pk}()},
\code{\link{rsq.pk}()},
\code{\link{subtract_pkproto.pk}()},
\code{\link{summary.pk}()}
}
\author{
Caroline Ring
}
\concept{methods for pk_faceted objects}
