% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_method_dispatch.R
\name{data_summary}
\alias{data_summary}
\title{data_summary()}
\usage{
data_summary(obj, ...)
}
\arguments{
\item{obj}{An object.}
}
\description{
This is the S3 method generic for data_summary()
}
\seealso{
[data_summary.pk()] for the method for class [pk()]
}
