% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_method_dispatch.R
\name{do_data_info.default}
\alias{do_data_info.default}
\title{do_data_info default method}
\usage{
\method{do_data_info}{default}(obj, ...)
}
\arguments{
\item{obj}{an object}

\item{...}{Additional arguments currently not in use.}
}
\description{
do_data_info default method
}
