% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_methods_do_data_info.R
\name{do_data_info.pk}
\alias{do_data_info.pk}
\title{calculate summary data info}
\usage{
\method{do_data_info}{pk}(obj, ...)
}
\arguments{
\item{obj}{A `pk` object}

\item{...}{Additional arguments. Not in use currently.}
}
\value{
Object of class [pk()] with an added `$data_info` list containing
non-compartmental analysis results.
}
\description{
Calculate summary data information, including non-compartmental analysis.
}
\author{
Caroline Ring
}
