% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_method_dispatch.R
\name{do_data_info}
\alias{do_data_info}
\title{do_data_info generic}
\usage{
do_data_info(obj, ...)
}
\description{
do_data_info generic
}
\seealso{
[do_data_info.pk()] for the `do_data_info` method for class [pk()]
}
