% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_method_dispatch.R
\name{do_fit.default}
\alias{do_fit.default}
\title{do_fit default method}
\usage{
\method{do_fit}{default}(obj, ...)
}
\arguments{
\item{obj}{an object}

\item{...}{Additional arguments currently not in use.}
}
\description{
do_fit default method
}
