% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_method_dispatch.R
\name{do_fit}
\alias{do_fit}
\title{Fitting}
\usage{
do_fit(obj, ...)
}
\arguments{
\item{obj}{the pk object}

\item{...}{Additional arguments currently not in use.}
}
\value{
The same [pk] object, with element `fit` containing the fitted
  results for each model in `stat_model`.
}
\description{
This is the S3 generic method for `do_fit`.
}
\seealso{
[do_fit.pk()] for the `do_fit` method for class [pk()]
}
