% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_linreg.R
\name{do_linreg}
\alias{do_linreg}
\title{Fit a line to data}
\usage{
do_linreg(x, y, intercept_exp = TRUE, slope_neg = TRUE)
}
\arguments{
\item{x}{A numeric vector of `x` values}

\item{y}{A numeric vector of `y` values}

\item{intercept_exp}{Logical: Whether to exp-transform the intercept before
returning. Default TRUE. For example, this is useful when `y` is
log-transformed and you ultimately want the intercept on the natural scale.}

\item{slope_neg}{Logical: Whether to take the negative of the slope before
returning. Default TRUE. For example, this is useful when a time constant
is the negative of the slope, and you ultimately desire the time constant.}
}
\value{
A list with two named numeric scalar elements: `intercept` and `slope`,
  containing the intercept and slope of the fitted line, respectively
}
\description{
Helper function to do linear regression for method of residuals
}
\details{
If there are two or more unique `x` values, this function calls [stats::lm()]
and takes the resulting coefficients. If there are fewer than two unique `x`
values, this function returns `NA_real_` for both intercept and slope,
because no line fit can be done in this case.
}
\author{
Caroline Ring
}
