% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_nca.R
\name{do_nca}
\alias{do_nca}
\title{Get non-compartmental analysis (NCA) results}
\usage{
do_nca(obs_data, dose_norm = TRUE)
}
\arguments{
\item{obs_data}{A `data.table`: a set of concentration vs. time data (one
chemical, species, route of dose administration, and medium of
concentration). Must have variables named `Time` (times of observations),
`Subject` (subject ID for each observation), `Conc_Dose` (dose-normalized
concentration, with non-detects substituted by LOQ), and `Route` ("po" for
oral administration, "iv" for IV administration).}
}
\value{
A named list of NCA-estimated parameters as from [PK::nca()]. These
  will be "AUC_tlast_1mgkg" (AUC to the last time point), "AUC_inf_1mgkg"
  (AUC extrapolated to infinity), "AUMC_inf_1mgkg" (AUMC to infinity), "MRT"
  (mean residence time), "halflife" (non-compartmental halfife), "Clearance",
  (clearance), and "Vss" (volume of distribution at steady-state). Note that
  if route is oral, halflife and Vss are not valid estimates; Clearance is
  actually Clearance divided by bioavailability (Fgutabs); and MRT is
  actually MTT.
}
\description{
Do NCA on one set of dose-normalized concentration vs. time data
}
\details{
This function calls [PK::nca()] after automatically determining the correct
"design" expected by [PK::nca()] (depending on whether all subjects have
measurements at all time points or not).
}
