% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_noncomp_fit.R
\name{do_noncomp_fit}
\alias{do_noncomp_fit}
\title{Helper function to perform noncomp fits}
\usage{
do_noncomp_fit(data.set)
}
\arguments{
\item{data.set}{A table of concentration-time data.}
}
\value{
A data.table with average concentration values
}
\description{
Helper function to perform noncomp fits
}
\author{
John Wambaugh, Caroline Ring, Christopher Cook
}
