% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_method_dispatch.R
\name{do_prefit.default}
\alias{do_prefit.default}
\title{do_prefit default method}
\usage{
\method{do_prefit}{default}(obj, ...)
}
\arguments{
\item{obj}{an object}

\item{...}{Additional arguments currently not in use.}
}
\description{
do_prefit default method
}
