% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_method_dispatch.R
\name{do_preprocess.default}
\alias{do_preprocess.default}
\title{do_preprocess default method}
\usage{
\method{do_preprocess}{default}(obj, ...)
}
\arguments{
\item{obj}{an object}

\item{...}{Additional arguments currently not in use.}
}
\value{
An error, when a non-pk object is used for the first argument.
}
\description{
do_preprocess default method
}
