% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_tkstats.R
\name{evaluate_tkstats}
\alias{evaluate_tkstats}
\title{Evaluate model predictions of toxicokinetic (TK) summary statistics}
\usage{
evaluate_tkstats(
  cvt_pre,
  fit_flat,
  fit_1comp,
  fit_2comp,
  group_by = c("DTXSID", "Species", "Route", "Media"),
  dose_norm = TRUE
)
}
\arguments{
\item{fit_flat}{A `data.table` of fitting output for the flat model: the output of `fit_all()`
with `model = "flat"`}

\item{fit_1comp}{A `data.table` of fitting output for the 1-compartment model: the output of
`fit_all()` with `model = "1compartment"`}

\item{fit_2comp}{A `data.table` of fitting output for the 2-compartment model: the output of
`fit_all()` with `model = "2compartment"`}

\item{group_by}{A character vector of variables to in `cvt_pre` to group data
by. NCA stats will be computed for each group defined by a unique
combination of the values of these variables. Default is `c("DTXSID",
"Species", "Route", "Media")`. You could also use `c("DTXSID", "Species",
"Route", "Media", "Dose")` to check dose-specific parameters.}

\item{dose_norm}{Logical: TRUE (the default) means to do NCA on dose-normalized
concentrations (`Conc_Dose` in `cvt_pre`), and FALSE means to do NCA on
non-dose-normalized concentrations (`Conc` in `cvt_pre`).}

\item{obs_data}{A `data.table` of observed concentration vs. time data, e.g.
as produced by [preprocess_data()]. Must include variables `Time`, `Dose`,
`Conc`, `Conc_SD`, `LOQ`, `N_Subjects`, `Route`, and `Media`.}
}
\value{
A `data.table` of goodness-of-fit measures for each dataset, each
  model, and each analysis (joint, separate, and pooled)
}
\description{
Evaluate how well a fitted toxicokinetic (TK) model predicts quantities that
summarize kinetics.
}
\details{
Evaluate how well a fitted TK model predicts summary statistics such as:

\eqn{t_{\textrm{max}}}: the time of peak concentration for a single oral
bolus dose

\eqn{C_{\textrm{max}}(1 \textrm{mg/kg})}: the peak concentration for an oral
bolus dose of 1 mg/kg

\eqn{AUC_{tlast}(1 \textrm{mg/kg})}: the area under the concentration-time
curve for a bolus dose of 1 mg/kg (IV or oral), evaluated at the last
observed time point

These statistics are selected because they can be estimated from observed TK
data (observations of concentration vs. time in plasma and/or blood after a
single bolus dose, IV and/or oral).
}
\author{
Caroline Ring
}
