% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_params_1comp.R
\name{fill_params_1comp}
\alias{fill_params_1comp}
\title{Fill parameters for 1-compartment model}
\usage{
fill_params_1comp(params)
}
\arguments{
\item{params}{Named list of parameters for the 1-compartment model.}
}
\value{
A named numeric vector of parameters, with any 1-compartment model
  parameters not present in `params` filled with `NA_real_`. If any two of
  `Fgutabs`, `Vdist`, and `Fgutabs_Vdist` were present in `params`, the third will
  be imputed to agree with the other two.
}
\description{
Fill parameters for 1-compartment model
}
\author{
Caroline Ring
}
