% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_params_2comp.R
\name{fill_params_2comp}
\alias{fill_params_2comp}
\title{Fill parameters for 2-compartment model}
\usage{
fill_params_2comp(params)
}
\arguments{
\item{params}{Named numeric vector of parameters for the 2-compartment model}
}
\value{
A named numeric vector of parameters, with any 2-compartment model
  parameters not present in `params` filled with `NA_real_`. If any two of
  `Fgutabs`, `V1`, and `Fgutabs_V1` were present in `params`, the third will
  be imputed to agree with the other two.
}
\description{
Fill parameters for 2-compartment model
}
\author{
Caroline Ring
}
