% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_params_httk_3comp2.R
\name{fill_params_httk_3comp2}
\alias{fill_params_httk_3comp2}
\title{Fill parameters for 1-compartment model}
\usage{
fill_params_httk_3comp2(params)
}
\arguments{
\item{params}{Named numeric vector of parameters for the 1-compartment model}
}
\value{
A named list of parameters,to be used with `httk`'s
 `3compartment2` model.
}
\description{
This is used to "fill" parameters prior to running `httk::solve_3compartment2()`.
During pre-fitting, the pKa_Donor & pKa_Accept parameters may hold multiple
values and to keep data tidy, each value was split into it's own row. This
function reconvenest these values into the appropriate format and outputs a
list which `httk` can use.
}
\author{
Gilberto Padilla Mercado
}
