% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_methods_fit.R
\name{fit.pk}
\alias{fit.pk}
\title{Fit PK model(s) for a `pk` object}
\usage{
\method{fit}{pk}(obj)
}
\arguments{
\item{obj}{A [pk] object.}
}
\value{
The same [pk] object, with element `fit` added to each `$stat_model`
  element, reflecting the fitted parameters for the corresponding model.
}
\description{
Fit PK model(s) for a `pk` object
}
\author{
Caroline Ring
}
