% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_faceted_methods_fit.R
\name{fit.pk_faceted}
\alias{fit.pk_faceted}
\title{Fitting}
\usage{
\method{fit}{pk_faceted}(obj)
}
\arguments{
\item{obj}{A `pk_faceted` object.}
}
\value{
The same `pk_faceted` object, with list column `pk_object` modified
  by applying [fit.pk()] to each of its elements.
}
\description{
Fit PK model(s) for a `pk_faceted` object
}
\author{
Caroline Ring
}
