% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_method_dispatch.R
\name{fit_sigma.default}
\alias{fit_sigma.default}
\title{Default method for fit_sigma()}
\usage{
\method{fit_sigma}{default}(obj, ...)
}
\arguments{
\item{obj}{An object}

\item{...}{Additional arguments currently.}
}
\value{
An error, when a non-pk object is used for the first argument.
}
\description{
Default method for fit_sigma()
}
