% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_method_dispatch.R
\name{fit_sigma}
\alias{fit_sigma}
\title{fit_sigma()}
\usage{
fit_sigma(obj, ...)
}
\arguments{
\item{obj}{An object.}

\item{...}{Additional arguments.}
}
\value{
A [pk()] object with a list of externally predicted values and optimized
hyper-parameter values and resulting AIC computations.
}
\description{
This is the S3 method generic for fit_sigma()
}
\seealso{
[fit_sigma.pk()] for the method for class [pk()]
}
