% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{flat}
\alias{flat}
\title{Flat model}
\format{
An object of class \code{list} (inherits from \code{pk_model}) of length 10.
}
\usage{
flat
}
\description{
The `pk_model` object defining the flat model.
}
\details{
A `pk_model` object: under the hood, a `list` object with elements
corresponding to the arguments of [pk_model()]. See that function
documentation for the definition of each element.

See [cp_flat()] for the function that predicts blood/plasma concentration for a bolus dose (oral or IV).

See [auc_flat()] for the function that predicts area under the concentration-time curve.

See [tkstats_flat()] for the function that calculates summary toxicokinetic statistics from 1-compartment model parameters.

See [params_flat()] for the function that determines bounds and starting guesses for model parameters, based on the data.
}
\keyword{datasets}
