% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_method_dispatch.R
\name{fold_error}
\alias{fold_error}
\title{Fold error}
\usage{
fold_error(obj, ...)
}
\arguments{
\item{obj}{an object}

\item{...}{Additional arguments currently not in use.}
}
\value{
A data.frame with one row for each `data_group`, `model` and
  `method`. A column contains the fold errors (observed/predicted) of the
  model fitted by the corresponding method. These residuals are
  concentrations in the same units as `obj$data$Conc.Units`; any
  concentration transformations (in `obj$scale$conc`) are *not* applied.
}
\description{
This is the S3 method generic for `fold_error`.
}
\seealso{
[fold_error.pk()] for the `fold_error` method for class [pk()]
}
