% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_method_dispatch.R
\name{fold_errors.default}
\alias{fold_errors.default}
\title{Fold_error default method}
\usage{
\method{fold_errors}{default}(obj, ...)
}
\arguments{
\item{obj}{an object}

\item{...}{Additional arguments currently not in use.}
}
\description{
Fold_error default method
}
