% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_method_dispatch.R
\name{fold_errors}
\alias{fold_errors}
\title{This is the S3 method generic for `fold_errors`.}
\usage{
fold_errors(obj, ...)
}
\arguments{
\item{obj}{an object}

\item{...}{Additional arguments currently not in use.}
}
\description{
`fold_errors()` is an alias for `fold_error()`
}
\seealso{
[fold_error.pk()] for the `fold_error` method for class [pk()]
}
