% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/force_ascii.R
\name{force_ascii}
\alias{force_ascii}
\title{Converting common quotation mark symbols into basic ASCII forms.}
\usage{
force_ascii(x)
}
\arguments{
\item{x}{A character vector that may have some special characters.}
}
\value{
Character vector with converted characters.
}
\description{
`force_ascii()` Takes curly quotes commonly found in chemical names and
converts them to simple ASCII equivalent. This can be important for
cross-compatibility with other databases and systems that use simple
apostrophes `'` and quotes `"` instead of curly quotes or prime symbols.
}
\examples{
try(cat(force_ascii("3,3’,4,4′-tetrachloro-5-biphenylol")))

}
