% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{force_ascii_quotes}
\alias{force_ascii_quotes}
\title{Converting common quotation mark symbols into basic ASCII forms.}
\usage{
force_ascii_quotes(x)
}
\arguments{
\item{x}{A character vector that may have some special characters.}
}
\value{
Character vector with converted characters.
}
\description{
`force_ascii_quotes()` Takes curly quotes commonly found in chemical names and
converts them to simple ASCII equivalent. This can be important for
cross-compatibility with other databases and systems that use simple
apostrophes `'` and quotes `"` instead of curly quotes or prime symbols.
}
\examples{
try(cat(invivoPKfit:::force_ascii_quotes("3,3’,4,4′-tetrachloro-5-biphenylol")))

}
