% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_methods_get_data.R
\name{get_data.pk}
\alias{get_data.pk}
\title{Get data}
\usage{
\method{get_data}{pk}(obj, ...)
}
\arguments{
\item{obj}{A [pk()] object that has been pre-processed}

\item{...}{Additional arguments. Currently not in use.}
}
\value{
A `data.frame`: the `data` element of `obj`
}
\description{
Extract pre-processed data from a [pk()] object
}
\author{
Caroline Ring
}
