% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_method_dispatch.R
\name{get_data}
\alias{get_data}
\title{get_data()}
\usage{
get_data(obj, ...)
}
\arguments{
\item{obj}{An object.}
}
\description{
This is the S3 method generic for get_data()
}
\seealso{
[get_data.pk()] for the method for class [pk()]
}
