% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_methods_get_data_group.R
\name{get_data_group.pk}
\alias{get_data_group.pk}
\title{Get data grouping}
\usage{
\method{get_data_group}{pk}(obj, as_character = FALSE, ...)
}
\arguments{
\item{obj}{An initialized `pk` object.}

\item{as_character}{Logical (Default: `FALSE`). Determines whether to return a
character vector. If set to `FALSE`, a list of expressions containing the `data_group`
variables is returned.}

\item{...}{Additional arguments not currently in use.}
}
\value{
An object of class `call` giving the data grouping as a `dplyr::vars()` specification
}
\description{
Get data grouping
}
