% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_method_dispatch.R
\name{get_data_group}
\alias{get_data_group}
\title{get_data_group()}
\usage{
get_data_group(obj, ...)
}
\arguments{
\item{obj}{An object.}
}
\description{
This is the S3 method generic for get_data_group()
}
\details{
`get_data_group()` is an alias for `data_summary()`
}
\seealso{
[data_summary.pk()] for the method for class [pk()]
}
