% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_methods_get_data_info.R
\name{get_data_info.pk}
\alias{get_data_info.pk}
\title{Get data_info}
\usage{
\method{get_data_info}{pk}(obj, ...)
}
\arguments{
\item{obj}{A [pk()] object that has had `data_info()` run on it}

\item{...}{Additional arguments. Currently not in use.}
}
\value{
A `data.frame`: the `data` element of `obj`
}
\description{
Extract summary data information results from a [pk()] object
}
\author{
Caroline Ring
}
