% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_methods_get_data_original.R
\name{get_data_original.pk}
\alias{get_data_original.pk}
\title{Get data_original}
\usage{
\method{get_data_original}{pk}(obj)
}
\arguments{
\item{obj}{A [pk()] object}
}
\value{
A `data.frame` -- the `data_original` element of `obj`
}
\description{
Get data_original
}
\author{
Caroline Ring
}
