% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_method_dispatch.R
\name{get_data_original}
\alias{get_data_original}
\title{get_data_original()}
\usage{
get_data_original(obj, ...)
}
\arguments{
\item{obj}{An object.}
}
\description{
This is the S3 method generic for get_data_original()
}
\seealso{
[get_data_original.pk()] for the method for class [pk()]
}
