% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_method_dispatch.R
\name{get_data_sigma_group.default}
\alias{get_data_sigma_group.default}
\title{Default method for get_data_sigma_group()}
\usage{
\method{get_data_sigma_group}{default}(obj, ...)
}
\arguments{
\item{obj}{An object}
}
\description{
Default method for get_data_sigma_group()
}
