% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_methods_get_data_sigma_group.R
\name{get_data_sigma_group.pk}
\alias{get_data_sigma_group.pk}
\title{Get data_sigma_group}
\usage{
\method{get_data_sigma_group}{pk}(obj, newdata = NULL)
}
\arguments{
\item{obj}{A [pk()] object}

\item{newdata}{Optional: A `data.frame` with new data for which to get the
`data_sigma_group`s. If NULL (the default), then the groups will be
evaluated for the `obj$data`.}
}
\value{
A `factor` vector giving the error SD group ID for each observation,
  as the interaction of the factors specified in
  `obj$stat_error_model$error_group`.
}
\description{
Get data_sigma_group
}
\author{
Caroline Ring
}
