% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_method_dispatch.R
\name{get_data_sigma_group}
\alias{get_data_sigma_group}
\title{get_data_sigma_group()}
\usage{
get_data_sigma_group(obj, ...)
}
\arguments{
\item{obj}{An object.}

\item{...}{Additional arguments currently not in use.}
}
\value{
A `factor` vector giving the error SD group ID for each observation,
  as the interaction of the factors specified in
  `obj$pk_groups$error_group`.
}
\description{
This is the S3 method generic for get_data_sigma_group()
}
\seealso{
[get_data_sigma_group.pk()] for the method for class [pk()]
}
