% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_method_dispatch.R
\name{get_error_group.default}
\alias{get_error_group.default}
\title{Default method for get_error_group()}
\usage{
\method{get_error_group}{default}(obj, ...)
}
\arguments{
\item{obj}{An object}
}
\description{
Default method for get_error_group()
}
