% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_methods_get_error_group.R
\name{get_error_group.pk}
\alias{get_error_group.pk}
\title{Get error group}
\usage{
\method{get_error_group}{pk}(obj, as_character = FALSE, ...)
}
\arguments{
\item{obj}{A [pk()] object.}

\item{as_character}{Logical (Default: `FALSE`). Determines whether to return a
character vector. If set to `FALSE`, a list of expressions containing the `data_group`
variables is returned.}

\item{...}{Additional arguments. Not in use currently.}
}
\value{
The stat_error_model error grouping
}
\description{
Get error group
}
\author{
Caroline Ring
}
