% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_method_dispatch.R
\name{get_error_group}
\alias{get_error_group}
\title{get_error_group()}
\usage{
get_error_group(obj, ...)
}
\arguments{
\item{obj}{An object.}
}
\description{
This is the S3 method generic for get_error_group()
}
\seealso{
[get_error_group.pk()] for the method for class [pk()]
}
