% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_method_dispatch.R
\name{get_fit.default}
\alias{get_fit.default}
\title{Default method for get_fit()}
\usage{
\method{get_fit}{default}(obj, ...)
}
\arguments{
\item{obj}{An object}
}
\description{
Default method for get_fit()
}
