% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_faceted_methods_get_fit.R,
%   R/pk_methods_get_fit.R
\name{get_fit.pk}
\alias{get_fit.pk}
\title{get_fit}
\usage{
\method{get_fit}{pk}(obj, model = NULL, ...)

\method{get_fit}{pk}(obj, model = NULL, ...)
}
\arguments{
\item{obj}{A [pk] object.}

\item{model}{Optional: Specify one or more of the fitted models for which to
make predictions. If NULL (the default), predictions will be returned for
all of the models in `obj$stat_model`.}
}
\value{
A [tibble::tibble()] grouped by the faceting variables with a new variable ` get_fit ` containing the results of applying [get_fit.pk()] to each item in the list column of [pk()] objects, `obj$pk_object`

A named list of objects of class `optimx`, named for the models in
  `model`. As described in [optimx::optimx()]  If only one model is
  specified, the return value will still be a list, but with only one
  element.
}
\description{
get_fit for pk_faceted objects

Get the [optimx::optimx()] output from a fitted `pk` object
}
\details{
This function returns the object(s) returned by [optimx::optimx()] for the
specified model(s) and method(s), for a fitted `pk` object. See
[optimx::optimx()] for details. Briefly, an `optimx` object is a `data.frame`
with one row for each method used, and variables that give the optimized
values for each parameter, along with several diagnostic variables (e.g. the
objective function value at the optimized parameter values; the number of
function evaluations/iterations; an integer code describing convergence
status). The object will have attributes `details` (providing
any messages returned by the methods) and `npar` (the number of parameters
optimized).
}
\seealso{
Other methods for pk_faceted objects: 
\code{\link{AIC.pk}()},
\code{\link{BIC.pk}()},
\code{\link{check_needed_status.pk}()},
\code{\link{coef.pk}()},
\code{\link{coef_sd.pk}()},
\code{\link{compare_models.pk}()},
\code{\link{data_summary.pk}()},
\code{\link{eval_tkstats.pk}()},
\code{\link{fold_error.pk}()},
\code{\link{get_data.pk}()},
\code{\link{get_data_group.pk}()},
\code{\link{get_data_info.pk}()},
\code{\link{get_data_original.pk}()},
\code{\link{get_data_sigma_group.pk}()},
\code{\link{get_error_group.pk}()},
\code{\link{get_mapping.pk}()},
\code{\link{get_nca.pk}()},
\code{\link{get_prefit.pk}()},
\code{\link{get_scale_conc.pk}()},
\code{\link{get_scale_time.pk}()},
\code{\link{get_settings_data_info.pk}()},
\code{\link{get_settings_optimx.pk}()},
\code{\link{get_settings_preprocess.pk}()},
\code{\link{get_stat_error_model.pk}()},
\code{\link{get_stat_model.pk}()},
\code{\link{get_status.pk}()},
\code{\link{get_tkstats.pk}()},
\code{\link{get_winning_model.pk}()},
\code{\link{is.pk}()},
\code{\link{logLik.pk}()},
\code{\link{nca.pk}()},
\code{\link{plot.pk}()},
\code{\link{predict.pk}()},
\code{\link{preprocess_data.pk_faceted}()},
\code{\link{residuals.pk}()},
\code{\link{rmse.pk}()},
\code{\link{rsq.pk}()},
\code{\link{subtract_pkproto.pk}()},
\code{\link{summary.pk}()}

Other methods for fitted pk objects: 
\code{\link{AIC.pk}()},
\code{\link{BIC.pk}()},
\code{\link{coef.pk}()},
\code{\link{coef_sd.pk}()},
\code{\link{eval_tkstats.pk}()},
\code{\link{get_tkstats.pk}()},
\code{\link{logLik.pk}()},
\code{\link{predict.pk}()},
\code{\link{residuals.pk}()},
\code{\link{rmse.pk}()},
\code{\link{rsq.pk}()}
}
\author{
Caroline Ring
}
\concept{methods for fitted pk objects}
\concept{methods for pk_faceted objects}
