% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_method_dispatch.R
\name{get_fit}
\alias{get_fit}
\title{get_fit()}
\usage{
get_fit(obj, ...)
}
\arguments{
\item{obj}{An object.}

\item{...}{Additional arguments currently not in use.}
}
\description{
This is the S3 method generic for get_fit()
}
\details{
`get_fit()` is an alias for `data_summary()`
}
\seealso{
[data_summary.pk()] for the method for class [pk()]
}
