% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fitted_params.R
\name{get_fitted_params}
\alias{get_fitted_params}
\title{Get a set of fitted model parameters}
\usage{
get_fitted_params(pk_fit_row, model_in)
}
\arguments{
\item{model_in}{The model to evaluate: one of 'flat', '1compartment', or
'2compartment'.}

\item{pk_fit}{A merged table of fitted PK parameters, as produced by [merge_fits()].
Must contain variables `DTXSID`, `Species`, `Analysis_Type`, and
`Studies.Analyzed`. Must also contain variables corresponding to the
parameters of the model specified in argument `model_in`, as required for
the route and media of data in `newdata` (as given by
[get_model_paramnames()]). These variables must be named as
`[param].[model]`. For example, for the 1-compartment model, if `newdata`
contains only IV-dosing data measured in plasma, then `pk_fit` must contain
variables named `kelim.1compartment` and `Vdist.1compartment`.}

\item{fit_log_conc_in}{Whether the fit was done to log-transformed
concentrations: TRUE or FALSE.}

\item{fit_conc_dose_in}{Whether the fit was done to dose-normalized
concentrations: TRUE or FALSE.}

\item{rescale_time_in}{Whether the fit was done after rescaling time from
hours to days, weeks, months, or years, for long studies: TRUE or FALSE.}

\item{method}{The algorithm used to do the fitting: "bobyqa" or "L-BFGS-B."}

\item{DTXSID_in}{The DSSTox Substance ID for which to evaluate}

\item{Species_in}{The species for which to evaluate}

\item{Analysis_Type_in}{The analysis type that produced the fit being
evaluated: one of 'Joint', 'Separate', or 'Pooled'}

\item{Studies.Analyzed_in}{The comma-separated string of studies included in
the fit being analyzed.}
}
\value{
A named list of model parameters for the specified fit, suitable for
  input into the `params` argument of one of the model functions.
}
\description{
Get a set of fitted model parameters for a specific dataset, analysis, and
model
}
\details{
[fit_all()] produces a table of fitted parameters for all chemicals and
species, for a given analysis and model. [merge_fits()] post-processes and
merges these tables for multiple models and fitting conditions. This
facilitates tabular display of fitted parameters. However, it can also be
useful to pull a particular set of fitted parameters from the table, in a
form suitable for input to one of the model functions, e.g. [cp_flat()],
[cp_1comp()], or [cp_2comp()]. This utility function does the necessary data
wrangling.

The user provides a full, merged table of PK parameters as produced by
[merge_fits()]. The user also specifies which fit they want: a unique
combination of model, fitting conditions (log-transformation,
dose-normalization, rescaling time, optimization method used), substance
(DTXSID), species, analysis type (Joint, Separate, or Pooled), and the list
of studies analyzed (only necessary for Separate analyses, since these were
done one study at a time).
}
\author{
Caroline Ring
}
