% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_method_dispatch.R
\name{get_hessian.default}
\alias{get_hessian.default}
\title{Default method for get_hessian()}
\usage{
\method{get_hessian}{default}(obj, ...)
}
\arguments{
\item{obj}{An object}

\item{...}{Additional arguments currently not in use.}
}
\description{
Default method for get_hessian()
}
