% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_method_dispatch.R
\name{get_hessian}
\alias{get_hessian}
\title{get_hessian()}
\usage{
get_hessian(obj, ...)
}
\arguments{
\item{obj}{An object.}

\item{...}{Additional arguments currently not in use.}
}
\value{
A dataframe with one row for each `data_group`, `model` and `method`.
 The remaining column is a `list` column containing the Hessian for each row.
}
\description{
This is the S3 method generic for get_hessian()
}
\seealso{
[hessian.pk()] for the method for class [pk()]
}
