% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_method_dispatch.R
\name{get_mapping.default}
\alias{get_mapping.default}
\title{Default method for get_mapping()}
\usage{
\method{get_mapping}{default}(obj, ...)
}
\arguments{
\item{obj}{An object}
}
\description{
Default method for get_mapping()
}
