% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_methods_get_mapping.R
\name{get_mapping.pk}
\alias{get_mapping.pk}
\title{Get mapping}
\usage{
\method{get_mapping}{pk}(obj)
}
\arguments{
\item{obj}{A [pk()] object}
}
\value{
A list of `quosure`s -- the `mapping` element of `obj`
}
\description{
Get mapping
}
\author{
Caroline Ring
}
