% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_method_dispatch.R
\name{get_mapping}
\alias{get_mapping}
\title{get_mapping()}
\usage{
get_mapping(obj, ...)
}
\arguments{
\item{obj}{An object.}

\item{...}{Additional arguments currently not in use.}
}
\description{
This is the S3 method generic for get_mapping()
}
\seealso{
[get_mapping.pk()] for the method for class [pk()]
}
