% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model_auc_function.R
\name{get_model_auc_function}
\alias{get_model_auc_function}
\title{Get model function}
\usage{
get_model_auc_function(model)
}
\arguments{
\item{model}{The name of a model. One of 'flat', '1compartment', or
'2compartment'.}
}
\value{
The analytic AUC function corresponding to `model`: `auc_flat()` for
  'flat', `auc_1comp()` for '1compartment', `auc_2comp()` for '2compartment'.
  Returned as a function object, not as a character string.
}
\description{
Get the function that predicts area under the concentration-time curve for a
given model name
}
\author{
Caroline Ring
}
