% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model_paramnames.R
\name{get_model_paramnames}
\alias{get_model_paramnames}
\title{Get model parameter names}
\usage{
get_model_paramnames(model)
}
\arguments{
\item{model}{The name of the model. Currently only "flat", "1compartment",
and "2compartment" are implemented.}
}
\value{
A character vector of parameter names for the specified model.
}
\description{
Get parameter names required by a specified model
}
