% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_method_dispatch.R
\name{get_nca.default}
\alias{get_nca.default}
\title{Default method for get_nca()}
\usage{
\method{get_nca}{default}(obj, ...)

\method{get_nca}{default}(obj, ...)
}
\arguments{
\item{obj}{An object}
}
\description{
Default method for get_nca()

Default method for get_nca()
}
