% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_methods_get_nca.R
\name{get_nca.pk}
\alias{get_nca.pk}
\title{Get NCA}
\usage{
\method{get_nca}{pk}(obj)
}
\arguments{
\item{obj}{A [pk()] object that has had `data_info()` run on it}
}
\value{
A `data.frame`: the `data` element of `obj`
}
\description{
Extract non-compartmental analysis results from a [pk()] object
}
\author{
Caroline Ring, Gilberto Padilla Mercado
}
