% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_method_dispatch.R
\name{get_nca}
\alias{get_nca}
\title{get_nca()}
\usage{
get_nca(obj, ...)

get_nca(obj, ...)
}
\arguments{
\item{obj}{An object.}
}
\description{
This is the S3 method generic for get_nca()

This is the S3 method generic for `get_nca()`
}
\details{
`get_nca()` is an alias for `nca`
}
\seealso{
[get_nca.pk()] for the method for class [pk()]

[nca.pk()] for the method for class [pk()]
}
