% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_methods_get_nca_group.R
\name{get_nca_group.pk}
\alias{get_nca_group.pk}
\title{Get nca_group}
\usage{
\method{get_nca_group}{pk}(obj, as_character = FALSE, ...)
}
\arguments{
\item{obj}{A [pk()] object.}

\item{as_character}{Logical (Default: `FALSE`). Determines whether to return a
character vector. If set to `FALSE`, a list of expressions containing the `data_group`
variables is returned.}

\item{...}{Additional arguments. Currently not implemented.}
}
\value{
A named list of the data_info settings
}
\description{
Get nca_group
}
\author{
Caroline Ring, Gilberto Padilla Mercado
}
