% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_method_dispatch.R
\name{get_nca_group}
\alias{get_nca_group}
\title{get_nca_group()}
\usage{
get_nca_group(obj, ...)
}
\arguments{
\item{obj}{An object.}

\item{...}{Additional arguments currently not in use.}
}
\value{
A named list of the data_info settings
}
\description{
This is the S3 method generic for get_nca_group()
}
\seealso{
[get_nca_group.pk()] for the method for class [pk()]
}
