% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_params_2comp.R
\name{get_params_2comp}
\alias{get_params_2comp}
\title{Get parameters for 2-compartment model}
\usage{
get_params_2comp(
  data,
  lower_bound = ggplot2::aes(kelim = 0.5 * log(2)/max(Time_trans), k12 = 0.5 *
    log(2)/max(Time_trans), k21 = 0.5 * log(2)/max(Time_trans), V1 = 0.01, Fgutabs = 0,
    kgutabs = 0.5 * log(2)/max(Time_trans), Fgutabs_V1 = 0.01, Rblood2plasma = 0.01),
  upper_bound = ggplot2::aes(kelim = 2 * log(2)/min(Time_trans[Time_trans > 0]), k12 = 2
    * log(2)/min(Time_trans[Time_trans > 0]), k21 = 2 * log(2)/min(Time_trans[Time_trans
    > 0]), V1 = 100, Fgutabs = 1, kgutabs = 2 * log(2)/min(Time_trans[Time_trans > 0]),
    Fgutabs_V1 = 100, Rblood2plasma = 100),
  param_units = ggplot2::aes(kelim = paste0("1/", unique(Time_trans.Units)), V1 =
    paste0("(", unique(Dose.Units), ")", "/", "(", unique(Conc.Units), ")"), k21 =
    paste0("1/", unique(Time_trans.Units)), k12 = paste0("1/", unique(Time_trans.Units)),
    Fgutabs = "unitless fraction", kgutabs = paste0("1/", unique(Time_trans.Units)),
    Fgutabs_V1 = paste0("(", unique(Conc.Units), ")", "/", "(", unique(Dose.Units), ")"),
    Rblood2plasma = "unitless ratio")
)
}
\arguments{
\item{data}{The data set to be fitted (e.g. the result of [preprocess_data()])}

\item{lower_bound}{A mapping specified using a call to [ggplot2::aes()],
giving the lower bounds for each variable, as expressions which may include
variables in `data`.}

\item{upper_bound}{A mapping specified using a call to [ggplot2::aes()],
giving the upper bounds for each variable, as expressions which may include
variables in `data`.}

\item{param_units}{A mapping specified using a call to [ggplot2::aes()],
giving the units for each variable, as expressions which may include
variables in `data`.}
}
\value{
A `data.frame`with the following variables:
- `param_name`: Character: Names of the model parameters
- `param_units`: Character: Units of the model parameters
- `optimize_param`: TRUE if each parameter is to be estimated from the data; FALSE otherwise
- `use_param`: TRUE if each parameter is to be used in evaluating the model; FALSE otherwise
-`lower_bounds`: Numeric: The lower bounds for each parameter
- `upper_bounds`: Numeric: The upper bounds for each parameter
- `start`: Numeric: The starting guesses for each parameter
}
\description{
Get parameters for 2-compartment model and determine whether each is to be
estimated from the data
}
\details{
The full set of model parameters for the 1-compartment model includes `V1`,
`kelim`, `k12`, `k21`, `kgutabs`, and `Fgutabs`.

## IV data, no oral data

If IV dosing data are available, but no oral dosing data are available, then
only the parameters `V1`, `kelim`, `k12`, and `k21` will be estimated from
the data. The parameters `kgutabs` and `Fgutabs` cannot be estimated from IV
data alone.

## Oral data, no IV data

If oral dosing data are available, but no IV dosing data are available, then
the parameters `kelim`, `k12`, `k21`, and `kgutabs` will be estimated from
the data. However, the parameters `Fgutabs` and `V1` cannot be identified
separately. From oral data alone, only the ratio `Fgutabs/V1` can be
identified. This ratio is represented by a single parameter named
`Fgutabs_V1`. `Fgutabs` and `V1` will not be optimized, but `Fgutabs_V1` will
be optimized, along with `kelim`, `k12`, `k21`, and `kgutabs`.

## Oral data and IV data

If both oral and IV dosing data are available, then `V1`, `kelim`, `k12`,
`k21`, `kgutabs`, and `Fgutabs` will all be estimated from the data.

# Blood and plasma data

If both blood and plasma data are available, then `Rblood2plasma` will be estimated from the data.

# Only one of blood or plasma data

If only one of blood or plasma data are available, then `Rblood2plasma` will be
held constant at 1, not estimated from the data.
}
\seealso{
Other 2-compartment model functions: 
\code{\link{auc_2comp}()},
\code{\link{cp_2comp_dt}()},
\code{\link{cp_2comp}()},
\code{\link{tkstats_2comp}()},
\code{\link{transformed_params_2comp}()}

Other get_params functions: 
\code{\link{get_params_1comp}()},
\code{\link{get_params_flat}()}

Other built-in model functions: 
\code{\link{auc_1comp}()},
\code{\link{auc_2comp}()},
\code{\link{auc_flat}()},
\code{\link{cp_1comp}()},
\code{\link{cp_2comp_dt}()},
\code{\link{cp_2comp}()},
\code{\link{cp_flat}()},
\code{\link{get_params_1comp}()},
\code{\link{get_params_flat}()},
\code{\link{tkstats_2comp}()},
\code{\link{transformed_params_2comp}()}
}
\author{
Caroline Ring
}
\concept{2-compartment model functions}
\concept{built-in model functions}
\concept{get_params functions}
