% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_params_httk_3comp2.R
\name{get_params_httk_3comp2}
\alias{get_params_httk_3comp2}
\title{Get parameters to fit `httk`'s `3compartment2` PBPK model}
\usage{
get_params_httk_3comp2(
  data,
  this_species,
  this_chemical,
  lower_bound = ggplot2::aes(BW = NA, Caco2.Pab = NA, Caco2.Pab.dist = NA, Clint = 0,
    Clint.dist = NA, Clmetabolismc = NA, Funbound.plasma = 0, Funbound.plasma.dist = NA,
    Funbound.plasma.adjustment = NA, Fabsgut = NA, Fhep.assay.correction = NA, hematocrit
    = NA, Kgut2pu = NA, Krbc2pu = NA, kgutabs = NA, Kliver2pu = NA, Krest2pu = NA,
    Kblood2air = NA, liver.density = NA, logHenry = NA, million.cells.per.gliver = NA, MW
    = NA, Pow = NA, MA = NA, Qcardiacc = NA, Qgfrc = NA, Qgutf = NA, Qliverf = NA, Qalvc
    = NA, Rblood2plasma = NA, 
     Vgutc = NA, Vliverc = NA, Vrestc = NA),
  upper_bound = ggplot2::aes(BW = NA, Caco2.Pab = NA, Caco2.Pab.dist = NA, Clint = 1000,
    Clint.dist = NA, Clmetabolismc = NA, Funbound.plasma = 1, Funbound.plasma.dist = NA,
    Funbound.plasma.adjustment = NA, Fabsgut = NA, Fhep.assay.correction = NA, hematocrit
    = NA, Kgut2pu = NA, Krbc2pu = NA, kgutabs = NA, Kliver2pu = NA, Krest2pu = NA,
    Kblood2air = NA, liver.density = NA, logHenry = NA, million.cells.per.gliver = NA, MW
    = NA, Pow = NA, MA = NA, Qcardiacc = NA, Qgfrc = NA, Qgutf = NA, Qliverf = NA, Qalvc
    = NA, Rblood2plasma = NA, 
     Vgutc = NA, Vliverc = NA, Vrestc = NA),
  param_units = ggplot2::aes(BW = "kg", Caco2.Pab = "1E-6 cm/s", Caco2.Pab.dist_m =
    "1E-6 cm/s", Caco2.Pab.dist_u = "1E-6 cm/s", Caco2.Pab.dist_l = "1E-6 cm/s", Clint =
    "uL/min/10^6 hepatocytes", Clint.dist_m = "uL/min/10^6 hepatocytes", Clint.dist_l =
    "uL/min/10^6 hepatocytes", Clint.dist_u = "uL/min/10^6 hepatocytes", Clmetabolismc =
    "L/h/kg BW", Funbound.plasma = "unitless fraction", Funbound.plasma.dist_m =
    "unitless fraction", Funbound.plasma.dist_l = "unitless fraction",
    Funbound.plasma.dist_u = "unitless fraction", 
     Funbound.plasma.adjustment =
    "unitless coefficient", Fabsgut = "fraction", Fhep.assay.correction = "fraction",
    hematocrit = "percent volume RBCs in blood", Kgut2pu = "unitless ratio", Krbc2pu =
    "unitless ratio", kgutabs = "unitless ratio", Kliver2pu = "unitless ratio", Krest2pu
    = "unitless ratio", Kblood2air = "unitless ratio", liver.density = "g/cm^3", logHenry
    = "log10(atmosphers*m^3/mole)", million.cells.per.gliver = "cells/g liver", MW =
    "g/mol", Pow = "octanol:water partition coefficinet", 
     MA =
    "phospholipid:water distribution coefficient", Qcardiacc = "L/h/kg BW^(3/4)", Qgfrc =
    "fraction", Qgutf = "fraction", Qliverf = "fraction", Qalvc = "fraction",
    Rblood2plasma = "unitless ratio", Vgutc = "L/kg BW", Vliverc = "L/kg BW", Vrestc =
    "L/kg BW"),
  restrictive = TRUE
)
}
\arguments{
\item{data}{The data set to be fitted (e.g. the result of [preprocess_data()])}

\item{this_species}{A character vector naming the species for calculations in `httk`.}

\item{this_chemical}{A character vector naming the chemical for calculations in `httk`.}

\item{lower_bound}{A mapping specified using a call to [ggplot2::aes()],
giving the lower bounds for each variable, as expressions which may include
variables in `data`.}

\item{upper_bound}{A mapping specified using a call to [ggplot2::aes()],
giving the upper bounds for each variable, as expressions which may include
variables in `data`.}

\item{param_units}{A mapping specified using a call to [ggplot2::aes()],
giving the units for each variable, as expressions which may include
variables in `data`.}

\item{restrictive}{A logical value (TRUE or FALSE. Default: FALSE) that says whether the
assumption is that the clearance is restrictive or non-restrictive}
}
\value{
A vector of blood or plasma concentration values  corresponding
 to `time`.
}
\description{
Get parameters to fit the `3compartment2`
model from the `httk` package (Wambaugh, Schacht, and Ring. 2025).
}
\section{Required parameters}{

These are given by the \link[httk]{parameterize_3comp2} function in `httk`.
Furthermore, they are transformed to a vector during hte prefitting process.
}

\seealso{
Other built-in model functions: 
\code{\link{auc_1comp}()},
\code{\link{auc_1comp_cl}()},
\code{\link{auc_1comp_rad}()},
\code{\link{auc_2comp}()},
\code{\link{auc_flat}()},
\code{\link{cp_1comp}()},
\code{\link{cp_1comp_cl}()},
\code{\link{cp_1comp_rad}()},
\code{\link{cp_2comp}()},
\code{\link{cp_2comp_dt}()},
\code{\link{cp_flat}()},
\code{\link{cp_httk_3comp2}()},
\code{\link{get_params_1comp}()},
\code{\link{get_params_1comp_cl}()},
\code{\link{get_params_1comp_fup}()},
\code{\link{get_params_1comp_rad}()},
\code{\link{get_params_2comp}()},
\code{\link{get_params_flat}()},
\code{\link{get_starts_1comp}()},
\code{\link{get_starts_1comp_cl}()},
\code{\link{get_starts_1comp_rad}()},
\code{\link{get_starts_2comp}()},
\code{\link{get_starts_flat}()},
\code{\link{get_starts_httk_3comp2}()},
\code{\link{tkstats_2comp}()},
\code{\link{transformed_params_2comp}()}

Other httk model functions: 
\code{\link{cp_httk_3comp2}()},
\code{\link{get_starts_httk_3comp2}()}

Other model concentration functions: 
\code{\link{cp_1comp}()},
\code{\link{cp_1comp_cl}()},
\code{\link{cp_1comp_rad}()},
\code{\link{cp_2comp}()},
\code{\link{cp_flat}()},
\code{\link{cp_httk_3comp2}()}
}
\author{
Gilberto Padilla Mercado
}
\concept{built-in model functions}
\concept{httk model functions}
\concept{model concentration functions}
