% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_peak.R
\name{get_peak}
\alias{get_peak}
\title{Helper function to find peak}
\usage{
get_peak(x, y, ...)
}
\arguments{
\item{x}{A numeric vector of `x` data}

\item{y}{A numeric vector of `y` data}

\item{...}{Optional: Additional arguments which will be passed to
[stats::approx()] (other than `x`, `y`, and `xout`).}
}
\value{
A list with two named numeric scalar components, `x` and `y`,
  containing the x- and y-values at the peak.
}
\description{
Finds x- and y-value at peak y value.
}
\details{
This is a helper function called by [get_starts()] to estimate the end of the
absorption phase for oral TK data.

If there is more than one unique `x` value where both `x` and corresponding
`y` are finite, this function calls [stats::approx()], then uses
[base::which.max()] to locate the maximum interpolated `y`-value. In effect,
this takes the average `y` value for each unique `x` value, then finds the
maximum.

If there is only one unique `x` value where both `x` and corresponding `y`
are finite, this function returns that `x` value, along with the average of
the finite `y` values corresponding to it.

If there are no unique `x` values where both `x` and corresponding `y` are
finite, this function returns `NA_real_` for the peak `x` and `y` values.
}
\author{
Caroline Ring
}
