% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_methods_get_prefit.R
\name{get_prefit.pk}
\alias{get_prefit.pk}
\title{Get prefit}
\usage{
\method{get_prefit}{pk}(obj, model = NULL, ...)
}
\arguments{
\item{obj}{A [pk()] object that has had `data_info()` run on it}

\item{model}{A vector with the names of model(s) that should be considered in output.}

\item{...}{Additional arguments. Currently not in use.}
}
\value{
A list of `data.frame`s: one for each model in `obj$stat_model`
}
\description{
Extract pre-fitting results from a [pk()] object
}
\author{
Caroline Ring
}
