% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_method_dispatch.R
\name{get_prefit}
\alias{get_prefit}
\title{get_prefit()}
\usage{
get_prefit(obj, ...)
}
\arguments{
\item{obj}{An object.}
}
\description{
This is the S3 method generic for get_prefit()
}
\seealso{
[get_prefit.pk()] for the method for class [pk()]
}
