% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_method_dispatch.R
\name{get_scale_conc.default}
\alias{get_scale_conc.default}
\title{Default method for get_scale_conc()}
\usage{
\method{get_scale_conc}{default}(obj, ...)
}
\arguments{
\item{obj}{An object}
}
\description{
Default method for get_scale_conc()
}
