% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_methods_get_scale_conc.R
\name{get_scale_conc.pk}
\alias{get_scale_conc.pk}
\title{Get scale_conc}
\usage{
\method{get_scale_conc}{pk}(obj, ...)
}
\arguments{
\item{obj}{A [pk()] object}

\item{...}{Additional arguments not currently in use.}
}
\value{
A `list`: `obj$scales$conc`
}
\description{
Extract concentration scale/transformation instructions from a [pk()] object
}
\author{
Caroline Ring
}
