% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_method_dispatch.R
\name{get_scale_conc}
\alias{get_scale_conc}
\title{get_scale_conc()}
\usage{
get_scale_conc(obj, ...)
}
\arguments{
\item{obj}{An object.}
}
\description{
This is the S3 method generic for get_scale_conc()
}
\seealso{
[get_scale_conc.pk()] for the method for class [pk()]
}
