% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_method_dispatch.R
\name{get_scale_time.default}
\alias{get_scale_time.default}
\title{Default method for get_scale_time()}
\usage{
\method{get_scale_time}{default}(obj, ...)
}
\arguments{
\item{obj}{An object}

\item{...}{Additional arguments currently not in use.}
}
\description{
Default method for get_scale_time()
}
