% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_method_dispatch.R
\name{get_settings_data_info.default}
\alias{get_settings_data_info.default}
\title{Default method for get_settings_data_info()}
\usage{
\method{get_settings_data_info}{default}(obj, ...)
}
\arguments{
\item{obj}{An object}

\item{...}{Additional arguments currently not in use.}
}
\description{
Default method for get_settings_data_info()
}
