% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_methods_get_settings_data_info.R
\name{get_settings_data_info.pk}
\alias{get_settings_data_info.pk}
\title{Get settings_data_info}
\usage{
\method{get_settings_data_info}{pk}(obj, ...)
}
\arguments{
\item{obj}{A [pk()] object}

\item{...}{Additional arguments. Currently not implemented.}
}
\value{
A named list of the data_info settings
}
\description{
Get settings_data_info
}
\author{
Caroline Ring
}
