% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_method_dispatch.R
\name{get_settings_data_info}
\alias{get_settings_data_info}
\title{get_settings_data_info()}
\usage{
get_settings_data_info(obj, ...)
}
\arguments{
\item{obj}{An object.}
}
\description{
This is the S3 method generic for get_settings_data_info()
}
\seealso{
[get_settings_data_info.pk()] for the method for class [pk()]
}
