% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_method_dispatch.R
\name{get_settings_optimx.default}
\alias{get_settings_optimx.default}
\title{Default method for get_settings_optimx()}
\usage{
\method{get_settings_optimx}{default}(obj, ...)
}
\arguments{
\item{obj}{An object}

\item{...}{Additional arguments currently not in use.}
}
\description{
Default method for get_settings_optimx()
}
