% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_methods_get_settings_optimx.R
\name{get_settings_optimx.pk}
\alias{get_settings_optimx.pk}
\title{Get settings_optimx}
\usage{
\method{get_settings_optimx}{pk}(obj)
}
\arguments{
\item{obj}{A [pk()] object}
}
\value{
A named list of the optimx settings
}
\description{
Get settings_optimx
}
\author{
Caroline Ring
}
