% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_method_dispatch.R
\name{get_settings_optimx}
\alias{get_settings_optimx}
\title{get_settings_optimx()}
\usage{
get_settings_optimx(obj, ...)
}
\arguments{
\item{obj}{An object.}

\item{...}{Additional arguments currently not in use.}
}
\description{
This is the S3 method generic for get_settings_optimx()
}
\seealso{
[get_settings_optimx.pk()] for the method for class [pk()]
}
