% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_method_dispatch.R
\name{get_settings_preprocess.default}
\alias{get_settings_preprocess.default}
\title{Default method for get_settings_preprocess()}
\usage{
\method{get_settings_preprocess}{default}(obj, ...)
}
\arguments{
\item{obj}{An object}
}
\description{
Default method for get_settings_preprocess()
}
