% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_methods_get_settings_preprocess.R
\name{get_settings_preprocess.pk}
\alias{get_settings_preprocess.pk}
\title{Get settings_preprocess}
\usage{
\method{get_settings_preprocess}{pk}(obj, ...)
}
\arguments{
\item{obj}{A [pk()] object}

\item{...}{Additional arguments. Currently not in use.}
}
\value{
A named list of the preprocessing settings
}
\description{
Get settings_preprocess
}
\author{
Caroline Ring
}
