% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_method_dispatch.R
\name{get_settings_preprocess}
\alias{get_settings_preprocess}
\title{get_settings_preprocess()}
\usage{
get_settings_preprocess(obj, ...)
}
\arguments{
\item{obj}{An object.}

\item{...}{Additional arguments currently not in use.}
}
\value{
A named list of the preprocessing settings
}
\description{
This is the S3 method generic for get_settings_preprocess()
}
\seealso{
[get_settings_preprocess.pk()] for the method for class [pk()]
}
