% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_starts_1comp_cl.R
\name{get_starts_1comp_cl}
\alias{get_starts_1comp_cl}
\title{Get starting values for 1-compartment model with specific clearance}
\usage{
get_starts_1comp_cl(data, par_DF, restrictive)
}
\arguments{
\item{data}{The data set to be fitted (e.g. the result of [preprocess_data()])}

\item{par_DF}{A `data.frame` with the following variables
\itemize{
\item `param_name`: Character: Names of the model parameters
\item `param_units`: Character: Units of the model parameters
\item `optimize_param`: TRUE if each parameter is to be estimated from the data; FALSE otherwise
\item `use_param`: TRUE if each parameter is to be used in evaluating the model; FALSE otherwise
\item`lower_bounds`: Numeric: The lower bounds for each parameter
\item `upper_bounds`: Numeric: The upper bounds for each parameter
}}

\item{restrictive}{A boolean value determinining whether to assume restrictive
or non-restrictive clearance when getting starting values.}
}
\value{
The same `data.frame` as `par_DF`, with an additional variable
 `starts` containing the derived starting value for each parameter. If a
 parameter cannot be estimated from the available data, then its starting value
 will be `NA_real_`
}
\description{
Derive starting values for 1-compartment model parameters from available data
}
\details{
This function is called internally by [get_params_1comp_cl()] and should
generally not be called directly by the user.

The full set of model parameters for the 1-compartment model includes `Vdist`,
`kelim`, `kgutabs`, `Fgutabs`, and `Rblood2plasma`.
However, in this version of the one-compartment model,
we use the liver, alveolar glomerular flow rates (`Q_totli`, `Q_alv`, and `Q_gfr`,
respectively) provided by [httk].
Other parameters provided by [httk] include:
`Fup`, `Clint`, `Kblood2air`, `Rblood2plasma`, `kgutabs`, and `Fgutabs`.

`Vdist` is calculated from estimated total clearance and `kelim`, which is
calculated from the data.

The numerical optimizer requires starting guesses for the value of each
parameter to be estimated from the data.

These are intended to be *very* rough starting guesses, so the algorithm here
is extremely naive. This function is not itself intended to produce valid
estimates for any of the model parameters, and it is highly unlikely to do so.
}
\section{Estimation of `kelim`}{


First, data are filtered to exclude any non-detects.

Then, data are split by route of administration, into an IV data set and an oral data
set. (It is possible that either IV or oral data may not be
available for a chemical.)
If IV data exist, then only IV data are used to derive starting estimates for
`kelim`, even if oral data also exist.

If only oral data exist, then the oral data are used to derive a starting
estimate for `kelim`.

Whichever data set is used (IV or oral), the starting value for `kelim` is
derived by assuming that the range of observed time values in the data set
spans two elimination half-lives. This implies that the elimination half-life
is equal to the midpoint of observed time values, and that the starting value
for the elimination time constant `kelim` is therefore `log(2)` divided by the
midpoint of observed time values.

Of course, this assumption is unlikely to be correct. However, we hope that it
will yield a starting guess for `kelim` that is at least on the right order of
magnitude.
}

\section{Starting value for `Vdist`}{


Using a calculated value for total clearance, `Cltot`, `Vdist` is estimated
by dividing this by the estimation of `kelim`.
}

\seealso{
Other 1-compartment model functions: 
\code{\link{auc_1comp}()},
\code{\link{auc_1comp_cl}()},
\code{\link{auc_1comp_rad}()},
\code{\link{cp_1comp}()},
\code{\link{cp_1comp_cl}()},
\code{\link{get_params_1comp}()},
\code{\link{get_params_1comp_cl}()},
\code{\link{get_params_1comp_fup}()},
\code{\link{get_starts_1comp}()}

Other get_starts functions: 
\code{\link{get_starts_1comp}()},
\code{\link{get_starts_1comp_rad}()},
\code{\link{get_starts_2comp}()},
\code{\link{get_starts_flat}()},
\code{\link{get_starts_httk_3comp2}()}

Other built-in model functions: 
\code{\link{auc_1comp}()},
\code{\link{auc_1comp_cl}()},
\code{\link{auc_1comp_rad}()},
\code{\link{auc_2comp}()},
\code{\link{auc_flat}()},
\code{\link{cp_1comp}()},
\code{\link{cp_1comp_cl}()},
\code{\link{cp_1comp_rad}()},
\code{\link{cp_2comp}()},
\code{\link{cp_2comp_dt}()},
\code{\link{cp_flat}()},
\code{\link{cp_httk_3comp2}()},
\code{\link{get_params_1comp}()},
\code{\link{get_params_1comp_cl}()},
\code{\link{get_params_1comp_fup}()},
\code{\link{get_params_1comp_rad}()},
\code{\link{get_params_2comp}()},
\code{\link{get_params_flat}()},
\code{\link{get_params_httk_3comp2}()},
\code{\link{get_starts_1comp}()},
\code{\link{get_starts_1comp_rad}()},
\code{\link{get_starts_2comp}()},
\code{\link{get_starts_flat}()},
\code{\link{get_starts_httk_3comp2}()},
\code{\link{tkstats_2comp}()},
\code{\link{transformed_params_2comp}()}
}
\author{
Caroline Ring
}
\concept{1-compartment model functions}
\concept{built-in model functions}
\concept{get_starts functions}
