% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_starts_1comp_fup.R
\name{get_starts_1comp_fup}
\alias{get_starts_1comp_fup}
\title{Get starting values for 1-compartment model with specific clearance}
\usage{
get_starts_1comp_fup(data, par_DF, restrictive = TRUE)
}
\arguments{
\item{data}{The data set to be fitted (e.g. the result of [preprocess_data()])}

\item{par_DF}{A `data.frame` with the following variables
\itemize{
\item `param_name`: Character: Names of the model parameters
\item `param_units`: Character: Units of the model parameters
\item `optimize_param`: TRUE if each parameter is to be estimated from the data; FALSE otherwise
\item `use_param`: TRUE if each parameter is to be used in evaluating the model; FALSE otherwise
\item`lower_bounds`: Numeric: The lower bounds for each parameter
\item `upper_bounds`: Numeric: The upper bounds for each parameter
}}

\item{restrictive}{Optional boolean value which determines whether the function
assumes restrictive (TRUE) or nonrestrictive (FALSE, default) clearance.}
}
\value{
The same `data.frame` as `par_DF`, with an additional variable
 `starts` containing the derived starting value for each parameter. If a
 parameter cannot be estimated from the available data, then its starting value
 will be `NA_real_`
}
\description{
Derive starting values for 1-compartment model parameters from available data
}
\details{
This function is called internally by [get_params_1comp()] and should
generally not be called directly by the user.

The full set of model parameters for the 1-compartment model includes `Vdist`,
`kelim`, `kgutabs`, `Fgutabs`, and `Rblood2plasma`. Whether each one can be
estimated from the data depends on what routes of administration are included
in the data. However, in this version of the one-compartment model,
we use the liver and glomerular flow rates (`Q_totli` and `Q_gfr`, respectively)
provided by [httk] as well as intrisic clearance `Clint` and fraction unbound
in plasma, `Fup`. These starting values are established and then are used
to calculate the other parameters in the 1-compartment model.

The numerical optimizer requires starting guesses for the value of each
parameter to be estimated from the data. Default starting guesses are derived from the available data.

These are intended to be *very* rough starting guesses, so the algorithm here
is extremely naive. This function is not itself intended to produce valid
estimates for any of the model parameters, and it is highly unlikely to do so.

The derivation process is as follows.

First, data are filtered to exclude any non-detects.

Then, data are split by route of administration, into an IV data set and an oral data
set. (It is possible that either IV or oral data may not be
available for a chemical.)

# Starting value for `kelim`

If IV data exist, then only IV data are used to derive starting estimates for
`kelim`, even if oral data also exist.

If only oral data exist, then the oral data are used to derive a starting
estimate for `kelim`.

Whichever data set is used (IV or oral), the starting value for `kelim` is
derived by assuming that the range of observed time values in the data set
spans two elimination half-lives. This implies that the elimination half-life
is equal to the midpoint of observed time values, and that the starting value
for the elimination time constant `kelim` is therefore `log(2)` divided by the
midpoint of observed time values.

Of course, this assumption is unlikely to be correct. However, we hope that it
will yield a starting guess for `kelim` that is at least on the right order of
magnitude.

# Starting value for `Vdist`

Using a calculated value for total clearance, `Cl_tot`, `Vdist` is estimated
by dividing this by the estimation of `kelim`.

# Starting value for `kgutabs`

If oral data exist (whether or not IV data also exist), then the oral data
are used to derive a starting value for `kgutabs`.

First, concentrations are dose-normalized by dividing them by their corresponding
doses. Then the normalized concentrations are log10-transformed.

The time of peak concentration (`tmax`), and the median (normalized,
log-transformed) peak concentration (`Cmax_log10`), are identified using [get_peak()].

As a very rough guess,`tmax` is assumed to
occur at one absorption half-life. Under this assumption, `kgutabs` is equal
to `log(2)/tmax`, and this is taken as the starting value.

# Starting value for `Fgutabs_Vdist`

If any oral data exist (whether or not IV data also exist), then the oral data
are used to derive a starting value for `Fgutabs_Vdist`.

If the kinetics obey a one-compartment model, then if concentrations are
dose-normalized, log-transformed, and plotted vs. time, then at late time
points (after concentration has peaked), the concentration vs. time
relationship will approach a straight line with slope `-kelim`.

If this straight line is extrapolated back to time 0, then the resulting
intercept (call it `A`), expressed on the natural scale, is equal to
`Fgutabs_Vdist * kgutabs/(kgutabs-kelim)`. See
https://www.boomer.org/c/p4/c09/c0902.php .

Roughly, we approximate `A` on the log10 scale by extrapolating back from the peak along a
straight line with slope `-kelim`, using the previously-derived starting
value for `kelim`. So `log10(A) = Cmax_log10 + kelim*tmax`.

Using the previously-derived starting values for `kgutabs` and `kelim`, then,
the starting value for `Fgutabs_Vdist` can be derived as `A * (kgutabs-kelim)/kgutabs`.

# Starting value for `Fgutabs`

If both oral and IV data exist, then the derived starting values for `Vdist`
(from the IV data) and `Fgutabs_Vdist` (from the oral data) are multiplied to
yield a derived starting value for `Fgutabs`.
}
\section{Starting value for `Rblood2plasma`}{


The starting value for `Rblood2plasma` is set to the value given by [httk::parameterize_gas_pbtk()].
}

\seealso{
Other 1-compartment model functions: 
\code{\link{auc_1comp}()},
\code{\link{auc_1comp_cl}()},
\code{\link{auc_1comp_rad}()},
\code{\link{cp_1comp}()},
\code{\link{cp_1comp_cl}()},
\code{\link{get_params_1comp}()},
\code{\link{get_params_1comp_cl}()},
\code{\link{get_params_1comp_fup}()},
\code{\link{get_starts_1comp}()},
\code{\link{get_starts_1comp_cl}()}

Other get_starts functions: 
\code{\link{get_starts_1comp}()},
\code{\link{get_starts_1comp_cl}()},
\code{\link{get_starts_1comp_rad}()},
\code{\link{get_starts_2comp}()},
\code{\link{get_starts_flat}()}

Other built-in model functions: 
\code{\link{auc_1comp}()},
\code{\link{auc_1comp_cl}()},
\code{\link{auc_1comp_rad}()},
\code{\link{auc_2comp}()},
\code{\link{auc_flat}()},
\code{\link{cp_1comp}()},
\code{\link{cp_1comp_cl}()},
\code{\link{cp_1comp_rad}()},
\code{\link{cp_2comp}()},
\code{\link{cp_2comp_dt}()},
\code{\link{cp_flat}()},
\code{\link{cp_httk_3comp2}()},
\code{\link{get_params_1comp}()},
\code{\link{get_params_1comp_cl}()},
\code{\link{get_params_1comp_fup}()},
\code{\link{get_params_1comp_rad}()},
\code{\link{get_params_2comp}()},
\code{\link{get_params_flat}()},
\code{\link{get_starts_1comp}()},
\code{\link{get_starts_1comp_cl}()},
\code{\link{get_starts_1comp_rad}()},
\code{\link{get_starts_2comp}()},
\code{\link{get_starts_flat}()},
\code{\link{tkstats_2comp}()},
\code{\link{transformed_params_2comp}()}
}
\author{
Caroline Ring
}
\concept{1-compartment model functions}
\concept{built-in model functions}
\concept{get_starts functions}
