% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_starts_1comp_rad.R
\name{get_starts_1comp_rad}
\alias{get_starts_1comp_rad}
\title{Get starting values for 1-compartment model with specific clearance}
\usage{
get_starts_1comp_rad(data, par_DF, restrictive)
}
\arguments{
\item{data}{The data set to be fitted (e.g. the result of [preprocess_data()])}

\item{par_DF}{A `data.frame` with the following variables
\itemize{
\item `param_name`: Character: Names of the model parameters
\item `param_units`: Character: Units of the model parameters
\item `optimize_param`: TRUE if each parameter is to be estimated from the data; FALSE otherwise
\item `use_param`: TRUE if each parameter is to be used in evaluating the model; FALSE otherwise
\item`lower_bounds`: Numeric: The lower bounds for each parameter
\item `upper_bounds`: Numeric: The upper bounds for each parameter
}}

\item{restrictive}{A boolean value determinining whether to assume restrictive
or non-restrictive clearance when getting starting values.}
}
\value{
The same `data.frame` as `par_DF`, with an additional variable
 `starts` containing the derived starting value for each parameter. If a
 parameter cannot be estimated from the available data, then its starting value
 will be `NA_real_`
}
\description{
Derive starting values for 1-compartment model parameters from available data
}
\details{
This function is called internally by [get_params_1comp()] and should
generally not be called directly by the user.

The full set of model parameters for the 1-compartment model includes `Vdist`,
`kelim`, `kgutabs`, `Fgutabs`, and `Rblood2plasma`. Whether each one can be
estimated from the data depends on what routes of administration are included
in the data. However, in this version of the one-compartment model,
we use the liver and glomerular flow rates (`Q_totli` and `Q_gfr`, respectively)
provided by [httk] as well as intrisic clearance `Clint` and fraction unbound
in plasma, `Fup`. These starting values are established and then are used
to calculate the other parameters in the 1-compartment model.

The numerical optimizer requires starting guesses for the value of each
parameter to be estimated from the data. Default starting guesses are derived from the available data.

These are intended to be *very* rough starting guesses, so the algorithm here
is extremely naive. This function is not itself intended to produce valid
estimates for any of the model parameters, and it is highly unlikely to do so.

The derivation process is as follows.

First, data are filtered to exclude any non-detects.

Then, data are split by route of administration, into an IV data set and an oral data
set. (It is possible that either IV or oral data may not be
available for a chemical.)
}
\section{Starting value for `Frec`}{


The empirical value of `Frec` is attainable when the experimental collection
of excreta goes on until there is no detectable compound. In this case,
the values will be equal to the maximum cumulative amount recovered divided
by `Dose` or `Dose` * `Fgutabs` for intravenous and oral administration, respectively.

If all `excreta` data are 'detectable', the assumption that we can necessarily
obtain an empirical estimate of `Frec` may not hold. While in this case still,
the maximum is used, the model fitting steps will help realize a better estimate
for this fraction recovered.
}

\section{Starting value for `Vdist`}{


Using a calculated value for total clearance, `Cltot`, `Vdist` is estimated
by dividing this by the estimation of `kelim`.
}

\seealso{
Other 1-compartment radiation model functions: 
\code{\link{cp_1comp_rad}()},
\code{\link{get_params_1comp_rad}()}

Other get_starts functions: 
\code{\link{get_starts_1comp}()},
\code{\link{get_starts_1comp_cl}()},
\code{\link{get_starts_1comp_fup}()},
\code{\link{get_starts_2comp}()},
\code{\link{get_starts_flat}()},
\code{\link{get_starts_httk_3comp2}()}

Other built-in model functions: 
\code{\link{auc_1comp}()},
\code{\link{auc_1comp_cl}()},
\code{\link{auc_1comp_rad}()},
\code{\link{auc_2comp}()},
\code{\link{auc_flat}()},
\code{\link{cp_1comp}()},
\code{\link{cp_1comp_cl}()},
\code{\link{cp_1comp_rad}()},
\code{\link{cp_2comp}()},
\code{\link{cp_2comp_dt}()},
\code{\link{cp_flat}()},
\code{\link{cp_httk_3comp2}()},
\code{\link{get_params_1comp}()},
\code{\link{get_params_1comp_cl}()},
\code{\link{get_params_1comp_fup}()},
\code{\link{get_params_1comp_rad}()},
\code{\link{get_params_2comp}()},
\code{\link{get_params_flat}()},
\code{\link{get_params_httk_3comp2}()},
\code{\link{get_starts_1comp}()},
\code{\link{get_starts_1comp_cl}()},
\code{\link{get_starts_1comp_fup}()},
\code{\link{get_starts_2comp}()},
\code{\link{get_starts_flat}()},
\code{\link{get_starts_httk_3comp2}()},
\code{\link{tkstats_2comp}()},
\code{\link{transformed_params_2comp}()}
}
\author{
Gilberto Padilla Mercado, Caroline Ring
}
\concept{1-compartment radiation model functions}
\concept{built-in model functions}
\concept{get_starts functions}
