% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_methods_get_stat_error_model.R
\name{get_stat_error_model.pk}
\alias{get_stat_error_model.pk}
\title{Get stat_error_model}
\usage{
\method{get_stat_error_model}{pk}(obj, ...)
}
\arguments{
\item{obj}{A [pk()] object}

\item{...}{Additional arguments.}
}
\value{
A named list of the stat_error_model settings
}
\description{
Get stat_error_model
}
\author{
Caroline Ring
}
