% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_method_dispatch.R
\name{get_stat_error_model}
\alias{get_stat_error_model}
\title{get_stat_error_model()}
\usage{
get_stat_error_model(obj, ...)
}
\arguments{
\item{obj}{An object.}
}
\description{
This is the S3 method generic for get_stat_error_model()
}
\seealso{
[get_stat_error_model.pk()] for the method for class [pk()]
}
