% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_methods_get_stat_model.R
\name{get_stat_model.pk}
\alias{get_stat_model.pk}
\title{Get stat_model}
\usage{
\method{get_stat_model}{pk}(obj, ...)
}
\arguments{
\item{obj}{A [pk()] object}

\item{...}{Additional arguments. Currently not in use.}
}
\value{
A `list` -- the `stat_model` element of `obj`
}
\description{
Get stat_model
}
\author{
Caroline Ring
}
