% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_method_dispatch.R
\name{get_stat_model}
\alias{get_stat_model}
\title{get_stat_model()}
\usage{
get_stat_model(obj, ...)
}
\arguments{
\item{obj}{An object.}

\item{...}{Additional arguments currently not in use.}
}
\value{
A `list` -- the `stat_model` element of `obj`
}
\description{
This is the S3 method generic for get_stat_model()
}
\seealso{
[get_stat_model.pk()] for the method for class [pk()]
}
