% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_faceted_methods_get_status.R,
%   R/pk_methods_get_status.R
\name{get_status.pk}
\alias{get_status.pk}
\title{get_status}
\usage{
\method{get_status}{pk}(obj)

\method{get_status}{pk}(obj)
}
\arguments{
\item{obj}{A `pk` object}
}
\value{
A [tibble::tibble()] grouped by the faceting variables with a new variable ` get_status ` containing the results of applying [get_status.pk()] to each item in the list column of [pk()] objects, `obj$pk_object`

The status of the `pk` object as an integer.
}
\description{
get_status for pk_faceted objects

Check status of a `pk` object
}
\details{
`pk` objects have integer statuses reflecting what stage of the analysis
process they are at.

1. Object has been initialized
2. Data pre-processing complete
3. Model pre-fitting complete 4
. Model fitting complete

If a `pk` object of status 2 or greater has its instructions modified with
`+`, then its status will be reset to 1, indicating that any analysis results
contained in the object are now outdated and all steps of the analysis need
to be re-run.

This function allows the user to check the status of a `pk` object.

A message will be printed listing the analysis steps that have been completed
for this `pk` object, and the integer status will be returned.
}
\seealso{
Other methods for pk_faceted objects: 
\code{\link{AIC.pk}()},
\code{\link{BIC.pk}()},
\code{\link{check_needed_status.pk}()},
\code{\link{coef.pk}()},
\code{\link{coef_sd.pk}()},
\code{\link{compare_models.pk}()},
\code{\link{data_summary.pk}()},
\code{\link{eval_tkstats.pk}()},
\code{\link{fold_error.pk}()},
\code{\link{get_data.pk}()},
\code{\link{get_data_group.pk}()},
\code{\link{get_data_info.pk}()},
\code{\link{get_data_original.pk}()},
\code{\link{get_data_sigma_group.pk}()},
\code{\link{get_error_group.pk}()},
\code{\link{get_fit.pk}()},
\code{\link{get_mapping.pk}()},
\code{\link{get_nca.pk}()},
\code{\link{get_prefit.pk}()},
\code{\link{get_scale_conc.pk}()},
\code{\link{get_scale_time.pk}()},
\code{\link{get_settings_data_info.pk}()},
\code{\link{get_settings_optimx.pk}()},
\code{\link{get_settings_preprocess.pk}()},
\code{\link{get_stat_error_model.pk}()},
\code{\link{get_stat_model.pk}()},
\code{\link{get_tkstats.pk}()},
\code{\link{get_winning_model.pk}()},
\code{\link{is.pk}()},
\code{\link{logLik.pk}()},
\code{\link{nca.pk}()},
\code{\link{plot.pk}()},
\code{\link{predict.pk}()},
\code{\link{preprocess_data.pk_faceted}()},
\code{\link{residuals.pk}()},
\code{\link{rmse.pk}()},
\code{\link{rsq.pk}()},
\code{\link{subtract_pkproto.pk}()},
\code{\link{summary.pk}()}
}
\author{
Caroline Ring
}
\concept{methods for pk_faceted objects}
