% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_method_dispatch.R
\name{get_status}
\alias{get_status}
\title{Get status}
\usage{
get_status(obj, ...)
}
\arguments{
\item{obj}{an object}

\item{...}{Additional arguments currently not in use.}
}
\value{
The status of the `pk` object as an integer.
}
\description{
This is the S3 method generic.
}
\seealso{
[get_status.pk()] for the `get_status` method for class [pk()]
}
