% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_method_dispatch.R
\name{get_tkstats.default}
\alias{get_tkstats.default}
\title{Default method for get_tkstats()}
\usage{
\method{get_tkstats}{default}(obj, ...)
}
\arguments{
\item{obj}{an object}

\item{...}{Additional arguments currently not in use.}
}
\value{
An error, when a non-pk object is used for the first argument.
}
\description{
Default method for get_tkstats()
}
