% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_tkstats.R
\name{get_tkstats}
\alias{get_tkstats}
\title{Get non-compartmental TK statistics from observed data}
\usage{
get_tkstats(
  cvt_pre,
  group_by = c("DTXSID", "Species", "Route", "Media"),
  dose_norm = TRUE
)
}
\arguments{
\item{cvt_pre}{A `data.table` of concentration vs. time data, preprocessed as
with [preprocess_data()]. Must contain variables Time, Conc_Dose, DTXSID,
Species, Route, Media, Subject, Study_ID, and Series_ID.}

\item{group_by}{A character vector of variables to in `cvt_pre` to group data
by. NCA stats will be computed for each group defined by a unique
combination of the values of these variables. Default is `c("DTXSID",
"Species", "Route", "Media")`. You could also use `c("DTXSID", "Species",
"Route", "Media", "Dose")` to check dose-specific parameters.}

\item{dose_norm}{Logical: TRUE (the default) means to do NCA on dose-normalized
concentrations (`Conc_Dose` in `cvt_pre`), and FALSE means to do NCA on
non-dose-normalized concentrations (`Conc` in `cvt_pre`).}
}
\value{
A `data.table` of non-compartmental statistics, with variables
  DTXSID, Species, Route, Media, tmax (time of peak concentration),
  Cmax_1mgkg (peak concentration at 1 mg/kg single bolus dose),
  AUC_tlast_1mgkg (AUC at last observed time point for 1 mg/kg single bolus
  dose), AUC_inf_1mgkg (AUC extrapolated to infinite time for 1 mg/kg single
  bolus dose), AUMC_inf_1mgkg (area under first moment curve extrapolated to
  infinite time for 1 mg/kg single bolus dose), MRT (maximum residence time,
  only for IV administration), halflife (terminal half-life, only for IV
  administration), Clearance (clearance rate), and Vss (volume of
  distribution at steady state, only for IV administration),
  Clearance_Fgutabs (clearance rate divided by oral bioavailability, only for
  oral administration), and MTT (maximum transit time, only for oral
  administration).
}
\description{
Get non-compartmental TK statistics from observed data
}
\author{
Caroline Ring
}
