% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_method_dispatch.R
\name{get_winning_model.default}
\alias{get_winning_model.default}
\title{Default method for get_winning_model()}
\usage{
\method{get_winning_model}{default}(obj, ...)
}
\arguments{
\item{obj}{An object}
}
\description{
Default method for get_winning_model()
}
