% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_faceted_methods_get_winning_model.R,
%   R/pk_methods_get_winning_model.R
\name{get_winning_model.pk}
\alias{get_winning_model.pk}
\title{get_winning_model}
\usage{
\method{get_winning_model}{pk}(obj, newdata = NULL, method = NULL, criterion = "AIC", ...)

\method{get_winning_model}{pk}(obj, newdata = NULL, method = NULL, criterion = "AIC", ...)
}
\arguments{
\item{obj}{A [pk()] object}

\item{newdata}{Optional: A `data.frame` containing new data to plot. Must
contain at least variables `Chemical`, `Species`, `Route`, `Media`, `Dose`,
`Time`, `Time.Units`, `Conc`, `Detect`, `Conc_SD`.  Default `NULL`, to use
the data in `obj$data`.}

\item{method}{Character: One or more of the [optimx::optimx()] methods used in
fitting. The winning model will be determined for each of these methods.
Default `NULL` to get the winning model for each method in
`obj$settings_optimx$method`.}

\item{criterion}{The name of a criterion function to use for model comparison.
Default "AIC". Must be the name of a function that (as for `AIC`) accepts
arguments `obj`, `newdata`, `method` and `model` (may accept other
arguments, specified in `...`) and returns output as for `AIC`: a named list
of numeric vectors (named for each of the model names in `model`), where
each vector has elements named for each of the method names in `method`,
containing the criterion value calculated for that model fitted using that
method.}

\item{...}{Optional: Other arguments to `criterion` function.}
}
\value{
A [tibble::tibble()] grouped by the faceting variables with a new variable ` get_winning_model ` containing the results of applying [get_winning_model.pk()] to each item in the list column of [pk()] objects, `obj$pk_object`

A named character array with one element for each item in `method`,
 giving the name of the winning model. The return value has attribute
 `criterion` giving the name of the criterion function used to compare
 models. (If the return value is named `x`, you can access this attribute
 using `attr(x, 'criterion')`.
}
\description{
get_winning_model for pk_faceted objects

Get winning model for a fitted `pk` object
}
\details{
Get the winning model (i.e. the model with the lowest value of the criterion
specified in `criterion`) for a fitted `pk` object, for a specified method,
and optionally for a specified new dataset.
}
\seealso{
Other methods for pk_faceted objects: 
\code{\link{AIC.pk}()},
\code{\link{BIC.pk}()},
\code{\link{check_needed_status.pk}()},
\code{\link{coef.pk}()},
\code{\link{coef_sd.pk}()},
\code{\link{compare_models.pk}()},
\code{\link{data_summary.pk}()},
\code{\link{eval_tkstats.pk}()},
\code{\link{fold_error.pk}()},
\code{\link{get_data.pk}()},
\code{\link{get_data_group.pk}()},
\code{\link{get_data_info.pk}()},
\code{\link{get_data_original.pk}()},
\code{\link{get_data_sigma_group.pk}()},
\code{\link{get_error_group.pk}()},
\code{\link{get_fit.pk}()},
\code{\link{get_mapping.pk}()},
\code{\link{get_nca.pk}()},
\code{\link{get_prefit.pk}()},
\code{\link{get_scale_conc.pk}()},
\code{\link{get_scale_time.pk}()},
\code{\link{get_settings_data_info.pk}()},
\code{\link{get_settings_optimx.pk}()},
\code{\link{get_settings_preprocess.pk}()},
\code{\link{get_stat_error_model.pk}()},
\code{\link{get_stat_model.pk}()},
\code{\link{get_status.pk}()},
\code{\link{get_tkstats.pk}()},
\code{\link{is.pk}()},
\code{\link{logLik.pk}()},
\code{\link{nca.pk}()},
\code{\link{plot.pk}()},
\code{\link{predict.pk}()},
\code{\link{preprocess_data.pk_faceted}()},
\code{\link{residuals.pk}()},
\code{\link{rmse.pk}()},
\code{\link{rsq.pk}()},
\code{\link{subtract_pkproto.pk}()},
\code{\link{summary.pk}()}
}
\author{
Caroline Ring
}
\concept{methods for pk_faceted objects}
