% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_method_dispatch.R
\name{hessian.default}
\alias{hessian.default}
\title{Default method for hessian()}
\usage{
\method{hessian}{default}(obj, ...)
}
\arguments{
\item{obj}{An object}

\item{...}{Additional arguments currently not in use.}
}
\description{
Default method for hessian()
}
