% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invivoPKfit-package.R
\docType{package}
\name{invivoPKfit}
\alias{invivoPKfit-package}
\alias{invivoPKfit}
\title{invivoPKfit: A package to fit model parameters to in vivo data}
\description{
The invivoPKfit package takes in vivo concentration-time data from TNO, RTI,
and NHEERL, and fits parameters of 1-compartment and 2-compartment models
for each chemical.
}
\author{
\strong{Maintainer}: John Wambaugh \email{wambaugh.john@epa.gov} (\href{https://orcid.org/0000-0002-4024-534X}{ORCID})

Authors:
\itemize{
  \item Caroline Ring \email{Ring.Caroline@epa.gov} (\href{https://orcid.org/0000-0002-0463-1251}{ORCID})
  \item Gilberto Padilla Mercado (\href{https://orcid.org/0000-0001-5423-1646}{ORCID})
  \item Chris Cook
}

}
\keyword{internal}
