% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_methods_is.R
\name{is.pk_faceted}
\alias{is.pk_faceted}
\title{Check whether an object is of class `pk_faceted`}
\usage{
is.pk_faceted(obj)
}
\arguments{
\item{obj}{The object whose class is to be tested}
}
\value{
TRUE if the object inherits from class `pk`, FALSE if it does not
}
\description{
Check whether an object is of class `pk_faceted`
}
\author{
Caroline Ring
}
