% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_model.R
\name{is.pk_model}
\alias{is.pk_model}
\title{Checks whether object is of class `pk_model`}
\usage{
is.pk_model(obj)
}
\arguments{
\item{obj}{An object.}
}
\value{
Logical. Whether `obj` is a `pk_model`.
}
\description{
Checks whether object is of class `pk_model`
}
