% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkproto_methods.R
\name{is.pk_scales}
\alias{is.pk_scales}
\title{Is an object class `pk_scales`?}
\usage{
is.pk_scales(obj)
}
\arguments{
\item{obj}{An object}
}
\value{
TRUE if `obj` inherits from class `pk_scales`; FALSE if not
}
\description{
Is an object class `pk_scales`?
}
\author{
Caroline Ring
}
