% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkproto_methods.R
\name{is.pkproto}
\alias{is.pkproto}
\title{Is an object pkproto?}
\usage{
is.pkproto(obj)
}
\arguments{
\item{obj}{An object}
}
\value{
TRUE if `obj` inherits from class `pkproto`; FALSE if not
}
\description{
Is an object pkproto?
}
\author{
Caroline Ring
}
