% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglikelihood_datatable.R
\name{log.likelihood}
\alias{log.likelihood}
\title{The log-likelihood function used for fitting}
\usage{
\method{log}{likelihood}(params, DT, modelfun, model)
}
\arguments{
\item{params}{A named list of parameter values}

\item{DT}{Concentration-time data for a given chemical}

\item{modelfun}{"analytic" to use the analytic model solution, "full" to use
the full ODE model}

\item{model}{The model to fit, either "1compartment" or "2compartment" (other
 models not implemented for now)

 @return A log-likelihood value for the data given the parameter values in params}
}
\description{
The log-likelihood function used for fitting
}
