% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_methods_logLik.R
\name{logLik.pk}
\alias{logLik.pk}
\title{Log-likelihood}
\usage{
\method{logLik}{pk}(
  obj,
  newdata = NULL,
  model = NULL,
  method = NULL,
  negative = FALSE,
  force_finite = FALSE
)
}
\arguments{
\item{obj}{A `pk` object}

\item{newdata}{Optional: A `data.frame` with new data for which to compute
log-likelihood. If NULL (the default), then log-likelihoods will be
computed for the data in `obj$data`. `newdata` is required to contain at
least the following variables: `Time`, `Dose`, `Route`,`Media`, `Conc`,
`Detect`, `Conc_SD`, `N_Subjects`. Before log-likelihood is calculated,
`Time` will be transformed according to the transformation in
`obj$scales$time` and `Conc` will be transformed according to the
transformation in `obj$scales$conc`.}

\item{model}{Optional: Specify one or more of the fitted models for which to
calculate log-likelihood. If NULL (the default), log-likelihoods will be
returned for all of the models in `obj$stat_model`.}

\item{method}{Optional: Specify one or more of the [optimx::optimx()] methods
for which to calculate log-likelihoods. If NULL (the default),
log-likelihoods will be returned for all of the models in
`obj$optimx_settings$method`.}
}
\value{
A named list of numeric vectors. There is one list element named for
  each model in `obj`'s [stat_model()] element, i.e. each PK model that was
  fitted to the data. Each list element is a numeric vector with as many
  elements as there were [optimx::optimx()] methods (specified in
  [settings_optimx()]). The vector names are the method names.  Each vector
  element contains the log-likelihood of the model fitted by the
  corresponding method, calculated for the data in `newdata` if any has been
  specified, or for the data in `obj$data` if `newdata` is `NULL`.
}
\description{
Extract log-likelihood(s) from a fitted `pk` object
}
\details{
For details on how the log-likelihood is calculated, see [log_likelihood()].
}
\author{
Caroline Ring
}
