% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_units.R
\name{midpt_log10}
\alias{midpt_log10}
\title{Log10-scaled midpoint}
\usage{
midpt_log10(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
The log10-scaled midpoint, calculated as `log10(mean(range(x, na.rm = TRUE))`
}
\description{
Log10-scaled midpoint
}
