% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{model_httk_3comp2}
\alias{model_httk_3comp2}
\title{3-compartment `httk` model for fitting `Funbound.plasma` and `Clint`}
\format{
An object of class \code{list} (inherits from \code{pk_model}) of length 10.
}
\usage{
model_httk_3comp2
}
\description{
The `pk_model` object defining the "3compartment2" model from `httk`.
}
\details{
A `pk_model` object: under the hood, a `list` object with named elements
corresponding to the arguments of [pk_model()]. Note that this model is currently
in development and fits can be unstable due to errors arising during ODE
solving. See that function documentation for the definition of each element.

See [cp_httk_3comp2_cl()] for the function that predicts blood/plasma concentration for a bolus dose (oral or IV).

See [params_3comp2_cl()] for the function that determines bounds and starting guesses for model parameters, based on the data.
}
\keyword{datasets}
