% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_faceted_methods_nca.R, R/pk_methods_nca.R
\name{nca.pk}
\alias{nca.pk}
\title{nca}
\usage{
\method{nca}{pk}(obj, newdata = NULL, nca_group = NULL, exclude = TRUE, dose_norm = FALSE)

\method{nca}{pk}(obj, newdata = NULL, nca_group = NULL, exclude = TRUE, dose_norm = FALSE)
}
\arguments{
\item{obj}{A [pk()] model object. Must be fitted, or the function will exit
with an error.}

\item{newdata}{Optional: A `data.frame` containing new data for which to
compute the TK stats. Must contain at least variables `Chemical`,
`Species`, `Route`, `Dose`, `Conc`, `Dose.Units`, `Conc.Units`, either
`Time_trans.Units` or `Time.Units`, and any other variables named in
`tk_grouping`. Default `NULL`, to use the data in `get_data(obj)`.}

\item{nca_group}{A list of variables provided using a `dplyr::vars()` call.
The data (either `newdata` or `obj$data`) will be grouped according to the
unique combinations of these variables. For each unique combination of
these variables in the data, a set of TK statistics will be computed. The
default is `NULL`, to use the same data grouping that was set in
[stat_nca()] for the `pk` object. However, you may specify a different data
grouping if you wish.}

\item{exclude}{Logical: `TRUE` to group the data for NCA after removing any
observations in the data marked for exclusion (if there is a variable
`exclude` in the data, an observation is marked for exclusion when `exclude
%in% TRUE`). `FALSE` to include all observations when grouping the data for
NCA, regardless of exclusion status. Default `TRUE`.}

\item{dose_norm}{Logical: `TRUE` to perform NCA after dose-normalizing
concentrations. `FALSE` (default) to perform NCA on un-transformed
concentrations.}
}
\value{
A [tibble::tibble()] grouped by the faceting variables with a new variable ` nca ` containing the results of applying [nca.pk()] to each item in the list column of [pk()] objects, `obj$pk_object`

A `data.frame` with variables including all the grouping variables in
  `nca_group`, `nca_group_id`; `design` (the auto-detected study design for
  this group); `param_name` (the name of the NCA parameter); `param_value`
  (the NCA parameter value); `param_sd_z` (standard deviation of the
  estimated NCA parameter value, if avaialble); `param_units` (the units of
  the NCA parameter, derived from the units of the data).
}
\description{
nca for pk_faceted objects

Non-compartmental analysis for a `pk` object
}
\details{
Perform non-compartmental analysis of data in a `pk` object (or optionally,
new data), using data groupings defined by `get_nca_group()` for the `pk`
object (or optionally, new groupings). If you provide both `newdata` and
`nca_group`, then everything in the `pk` object will be ignored and you will
simply be doing NCA *de novo* (which may be what you want).
}
\seealso{
Other methods for pk_faceted objects: 
\code{\link{AIC.pk}()},
\code{\link{BIC.pk}()},
\code{\link{check_needed_status.pk}()},
\code{\link{coef.pk}()},
\code{\link{coef_sd.pk}()},
\code{\link{compare_models.pk}()},
\code{\link{data_summary.pk}()},
\code{\link{eval_tkstats.pk}()},
\code{\link{fold_error.pk}()},
\code{\link{get_data.pk}()},
\code{\link{get_data_group.pk}()},
\code{\link{get_data_info.pk}()},
\code{\link{get_data_original.pk}()},
\code{\link{get_data_sigma_group.pk}()},
\code{\link{get_error_group.pk}()},
\code{\link{get_fit.pk}()},
\code{\link{get_mapping.pk}()},
\code{\link{get_nca.pk}()},
\code{\link{get_prefit.pk}()},
\code{\link{get_scale_conc.pk}()},
\code{\link{get_scale_time.pk}()},
\code{\link{get_settings_data_info.pk}()},
\code{\link{get_settings_optimx.pk}()},
\code{\link{get_settings_preprocess.pk}()},
\code{\link{get_stat_error_model.pk}()},
\code{\link{get_stat_model.pk}()},
\code{\link{get_status.pk}()},
\code{\link{get_tkstats.pk}()},
\code{\link{get_winning_model.pk}()},
\code{\link{is.pk}()},
\code{\link{logLik.pk}()},
\code{\link{plot.pk}()},
\code{\link{predict.pk}()},
\code{\link{preprocess_data.pk_faceted}()},
\code{\link{residuals.pk}()},
\code{\link{rmse.pk}()},
\code{\link{rsq.pk}()},
\code{\link{subtract_pkproto.pk}()},
\code{\link{summary.pk}()}
}
\author{
Caroline Ring
}
\concept{methods for pk_faceted objects}
