% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{paste_fraction}
\alias{paste_fraction}
\title{Creates a string representation of a fraction}
\usage{
paste_fraction(x, y)
}
\arguments{
\item{x}{Numerator.}

\item{y}{Denominator.}
}
\value{
A character vector.
}
\description{
Parentheses will enclose the numerator or denominator if there is already
a division symbol "/" in the expression (if it is an expression or character vector).
}
