% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_methods_add_pkproto.R
\name{pk_add.default}
\alias{pk_add.default}
\title{Add pkproto object default method}
\usage{
\method{pk_add}{default}(object, pk_obj, objectname)
}
\description{
Add pkproto object default method
}
