% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_methods_add_pkproto.R
\name{pk_add.pk_facet_by}
\alias{pk_add.pk_facet_by}
\title{Add facet_by()}
\usage{
\method{pk_add}{pk_facet_by}(object, pk_obj, objectname)
}
\arguments{
\item{object}{The `pk_facet_by` object to be added.}

\item{pk_obj}{The [pk()] object to which the `pk_facet_by` object will be added.}

\item{objectname}{The name of the `pk_facet_by` object.}
}
\description{
Add facet_by()
}
