% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_methods_add_pkproto.R
\name{pk_add.pk_nca_group}
\alias{pk_add.pk_nca_group}
\title{Add a `pk_nca_group` object.}
\usage{
\method{pk_add}{pk_nca_group}(pkproto_obj, pk_obj, objectname)
}
\arguments{
\item{pkproto_obj}{The `pk_nca_group` object to be added.}

\item{pk_obj}{The `pk` object to which the `pk_nca_group` object will be added.}

\item{objectname}{The name of the `pk_nca_group` object.}
}
\value{
The `pk` object, modified by the `pk_nca_group` object.
}
\description{
Add a `pk_nca_group` object.
}
\author{
Caroline Ring, Gilberto Padilla Mercado
}
