% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_methods_add_pkproto.R
\name{pk_add.pk_scales}
\alias{pk_add.pk_scales}
\title{Add a `pk_scales` object to a `pk` object.}
\usage{
\method{pk_add}{pk_scales}(pkproto_obj, pk_obj, objectname)
}
\arguments{
\item{pkproto_obj}{The `pk_scales` object to be added.}

\item{pk_obj}{The `pk` object to which the `pk_scales` object will be added.}

\item{objectname}{The name of the `pk_scales` object.}
}
\value{
The `pk` object, modified by the `pk_scales` object.
}
\description{
Add a `pk_scales` object to a `pk` object.
}
\author{
Caroline Ring
}
