% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_methods_add_pkproto.R
\name{pk_add.pk_sd_group}
\alias{pk_add.pk_sd_group}
\title{Add sd_group}
\usage{
\method{pk_add}{pk_sd_group}(pkproto_obj, pk_obj, objectname)
}
\arguments{
\item{pkproto_obj}{The `pk_sd_group` object to be added.}

\item{pk_obj}{The [pk()] object to which the `pk_sd_group` object will be added.}

\item{objectname}{The name of the `pk_sd_group` object.}
}
\value{
The [pk()] object, with the added `pk_sd_group` in the `groups` sub-list.
}
\description{
Add sd_group
}
\author{
Gilberto Padilla Mercado
}
