% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_methods_add_pkproto.R
\name{pk_add.pk_stat_model}
\alias{pk_add.pk_stat_model}
\title{Add a `pk_stat_model` object.}
\usage{
\method{pk_add}{pk_stat_model}(pkproto_obj, pk_obj, objectname)
}
\arguments{
\item{pkproto_obj}{The `pk_stat_model` object to be added.}

\item{pk_obj}{The `pk` object to which the `pk_stat_model` object will be added.}

\item{objectname}{The name of the `pk_stat_model` object.}
}
\value{
The `pk` object, modified by adding the `stat_model`.
}
\description{
Add a `pk_stat_model` object.
}
\author{
Caroline Ring
}
