% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_methods_add_pkproto.R
\name{pk_add}
\alias{pk_add}
\title{Add a `pkproto` object to a `pk` object}
\usage{
pk_add(pkproto_obj, pk_obj, objectname)
}
\arguments{
\item{pkproto_obj}{The `pkproto` object to be added}

\item{pk_obj}{The `pk` object to which the `pkproto` object is to be added}

\item{objectname}{The object name}
}
\description{
This is the S3 generic method.
}
\seealso{
[pk_add.pk_scales()] for the method for adding `pk_scales` objects
  (from [scale_conc()] and [scale_time()]); [pk_add.pk_settings_preprocess()]
  for the method for adding `pk_settings_preprocess` objects (from
  [settings_preprocess()]); [pk_add.pk_settings_data_info()] for the method
  for adding `pk_settings_data_info` objects (from [settings_data_info()]);
  [pk_add.pk_settings_optimx()] for the method for adding
  `pk_settings_optimx` objects (from [settings_optimx()]);
  [pk_add.pk_stat_model()] for the method for adding `pk_stat_model` objects
  (from `stat_model()`)
}
