% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_methods_add_pkproto.R
\name{pk_add.uneval}
\alias{pk_add.uneval}
\title{Add an `uneval` object}
\usage{
\method{pk_add}{uneval}(object, pk_obj, objectname)
}
\arguments{
\item{object}{The `uneval` (mapping) object to be added.}

\item{pk_obj}{The [pk()] object to which the `uneval` object will be added.}

\item{objectname}{The name of the `uneval` object.}
}
\value{
The [pk()] object, modified by adding the new mapping.
}
\description{
Add an object created by [ggplot2::aes()]
}
\details{
This function adds a new variable mapping (created by [ggplot2::aes()]),
which has class `uneval`, to an existing [pk()] object.

The new mapping will completely replace any existing mapping.
}
\author{
Caroline Ring
}
