% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_methods_add_pkproto.R
\name{pk_subtract.default}
\alias{pk_subtract.default}
\title{Subtract pkproto object default method}
\usage{
\method{pk_subtract}{default}(object, pk_obj, objectname)
}
\description{
Subtract pkproto object default method
}
