% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_methods_subtract_pkproto.R
\name{pk_subtract.pk_stat_model}
\alias{pk_subtract.pk_stat_model}
\title{Subtract a `pk_stat_model` object.}
\usage{
\method{pk_subtract}{pk_stat_model}(pkproto_obj, pk_obj, objectname)
}
\arguments{
\item{pkproto_obj}{The `pk_stat_model` object to be subtracted.}

\item{pk_obj}{The `pk` object from which the `pk_stat_model` object will be subtracted.}

\item{objectname}{The name of the `pk_stat_model` object.}
}
\value{
The `pk` object, modified by subtracting the `stat_model`.
}
\description{
Subtract a `pk_stat_model` object.
}
\author{
Caroline Ring
}
