% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkdataset_nheerlcleaned.R
\docType{data}
\name{pkdataset_nheerlcleaned}
\alias{pkdataset_nheerlcleaned}
\title{Toxicokinetic data from the "Concentration vs. Time Database"}
\format{
A data table with 2454 rows and 19 variables:
}
\source{
\url{https://github.com/USEPA/CompTox-PK-CvTdb}
}
\usage{
pkdataset_nheerlcleaned
}
\description{
A dataset containing experimental time-course data of chemical compound
concentrations in body fluids and tissues
}
\keyword{datasets}
