% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_faceted_methods_plot.R, R/pk_methods_plot.R
\name{plot.pk}
\alias{plot.pk}
\title{plot}
\usage{
\method{plot}{pk}(
  obj,
  newdata = NULL,
  model = NULL,
  method = NULL,
  exclude = TRUE,
  use_scale_conc = FALSE,
  plot_data_aes = NULL,
  plot_fit_aes = NULL,
  facet_fun = NULL,
  facet_fun_args = NULL,
  n_interp = 10,
  ...
)

\method{plot}{pk}(
  obj,
  newdata = NULL,
  model = NULL,
  method = NULL,
  exclude = TRUE,
  use_scale_conc = FALSE,
  plot_data_aes = NULL,
  plot_fit_aes = NULL,
  facet_fun = NULL,
  facet_fun_args = NULL,
  n_interp = 10,
  ...
)
}
\arguments{
\item{obj}{A [pk()] object}

\item{newdata}{Optional: A `data.frame` containing new data to plot. Must
contain at least variables `Chemical`, `Species`, `Route`, `Media`, `Dose`,
`Time`, `Time.Units`, `Conc`, `Detect`, `Conc_SD`.  Default `NULL`, to use
the data in `obj$data`.}

\item{model}{Character: One or more of the models fitted. Curve fits will be
plotted for these models. Default `NULL` to plot fits for all models in
`obj$stat_model`.}

\item{method}{Character: One or more of the [optimx::optimx()] methods used.
Default `NULL` to plot fits for all methods in `obj$settings_optimx$method`.}

\item{use_scale_conc}{Possible values: `TRUE`, `FALSE`, or a named list with
elements `dose_norm` and `log10_trans` which themselves should be either
`TRUE` or `FALSE`.  If `use_scale_conc = FALSE` (the default for this
function), then the data and fits will be plotted without any
dose-normalization or log-transformation. If `use_scale_conc = TRUE` , then
the concentration scaling/transformations in `obj` will be applied to the
y-axis (concentration axis). If `use_scale_conc = list(dose_norm = ...,
log10_trans = ...)`, then the specified dose normalization and/or
log10-transformation will be applied to the y-axis (concentration axis) of
the plots.}

\item{plot_data_aes}{Optional: Aesthetic mapping for the plot layer that
visualizes the data. Default `NULL`, in which case a default mapping will be
used based on the value of `use_scale_conc`.}

\item{plot_fit_aes}{Optional: Aesthetic mapping for the plot layer that
visualizes the fitted curves. Default `NULL`, in which case a default mapping will be
used based on the value of `use_scale_conc`.}

\item{facet_fun}{Default `"facet_grid"`. Optional: The name of the `ggplot2` faceting function to use:
[ggplot2::facet_grid()], [ggplot2::facet_wrap()], or `'none'` to do no
faceting. Default `NULL`, in which case a default faceting will be
applied based on the value of `use_scale_conc`.}

\item{facet_fun_args}{A named list of arguments to the faceting function in
 `facet_fun` (if any). Default:
```
list(rows = ggplot2::vars(Route),
cols= ggplot2::vars(Media),
scales = "free_y",
labeller = "label_both")
```}

\item{n_interp}{For plotting: the number of time points to interpolate between
each observed time point. Default 10.}
}
\value{
A [tibble::tibble()] grouped by the faceting variables with a new variable ` plot ` containing the results of applying [plot.pk()] to each item in the list column of [pk()] objects, `obj$pk_object`

A [ggplot2::ggplot()]-class plot object.
}
\description{
plot for pk_faceted objects

Plot data and model fits from a [pk()] object.
}
\details{
If the [pk()] object has not been fitted, then only the data will be plotted
(because no curve fits exist).
}
\seealso{
Other methods for pk_faceted objects: 
\code{\link{AIC.pk}()},
\code{\link{BIC.pk}()},
\code{\link{check_needed_status.pk}()},
\code{\link{coef.pk}()},
\code{\link{coef_sd.pk}()},
\code{\link{compare_models.pk}()},
\code{\link{data_summary.pk}()},
\code{\link{eval_tkstats.pk}()},
\code{\link{fold_error.pk}()},
\code{\link{get_data.pk}()},
\code{\link{get_data_group.pk}()},
\code{\link{get_data_info.pk}()},
\code{\link{get_data_original.pk}()},
\code{\link{get_data_sigma_group.pk}()},
\code{\link{get_error_group.pk}()},
\code{\link{get_fit.pk}()},
\code{\link{get_mapping.pk}()},
\code{\link{get_nca.pk}()},
\code{\link{get_prefit.pk}()},
\code{\link{get_scale_conc.pk}()},
\code{\link{get_scale_time.pk}()},
\code{\link{get_settings_data_info.pk}()},
\code{\link{get_settings_optimx.pk}()},
\code{\link{get_settings_preprocess.pk}()},
\code{\link{get_stat_error_model.pk}()},
\code{\link{get_stat_model.pk}()},
\code{\link{get_status.pk}()},
\code{\link{get_tkstats.pk}()},
\code{\link{get_winning_model.pk}()},
\code{\link{is.pk}()},
\code{\link{logLik.pk}()},
\code{\link{nca.pk}()},
\code{\link{predict.pk}()},
\code{\link{preprocess_data.pk_faceted}()},
\code{\link{residuals.pk}()},
\code{\link{rmse.pk}()},
\code{\link{rsq.pk}()},
\code{\link{subtract_pkproto.pk}()},
\code{\link{summary.pk}()}
}
\author{
Caroline Ring
}
\concept{methods for pk_faceted objects}
