% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_methods.R
\name{plot_data.pk}
\alias{plot_data.pk}
\title{Plot concentration vs. time data.}
\usage{
plot_data.pk(
  obj,
  newdata = NULL,
  plot_dose_norm = TRUE,
  plot_log10_conc = FALSE
)
}
\arguments{
\item{obj}{A [pk()] object with concentration-time data in element `obj$data`.}

\item{newdata}{Optional: A new set of concentration vs. time data to plot,
different from `obj$data`. Default `NULL` to plot the data in `obj$data`.}
}
\value{
A [ggplot2::ggplot()] plot object.
}
\description{
Plot concentration vs. time data.
}
\author{
Caroline Ring
}
