% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_methods_add_pkproto.R
\name{+.pk}
\alias{+.pk}
\title{Add a `pkproto` object to a `pk` object}
\usage{
\method{+}{pk}(e1, e2)
}
\arguments{
\item{e1}{A `pk` pbject}

\item{e2}{A `pkproto` object}
}
\value{
The `pk` object, modified by adding the `pkproto` object
}
\description{
Add a `pkproto` object to a `pk` object
}
\details{
Note that `e1 + e2` is equivalent to

```
`+`(e1, e2)
```
}
\author{
Caroline Ring
}
