% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_faceted_methods_add_pkproto.R
\name{+.pk_faceted}
\alias{+.pk_faceted}
\title{Add a `pkproto` object to a `pk_faceted` object}
\usage{
\method{+}{pk_faceted}(e1, e2)
}
\arguments{
\item{e1}{A `pk_faceted` pbject}

\item{e2}{A `pkproto` object}
}
\value{
The `pk` object, modified by adding the `pkproto` object
}
\description{
Add a `pkproto` object to a `pk_faceted` object
}
\details{
Note that `e1 + e2` is equivalent to

```
`+`(e1, e2)
```
}
\author{
Caroline Ring
}
