% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{post_name_value}
\alias{post_name_value}
\title{Creates a [cli::cli_fmt()] output for the pattern "name(x), length = "}
\usage{
post_name_value(x, extra = "")
}
\arguments{
\item{x}{A named vector with a single value for each element.}

\item{extra}{A character vector that should be printed between each name and value.}
}
\value{
A cli::cli_fmt output list.
}
\description{
Creates a [cli::cli_fmt()] output for the pattern "name(x), length = "
}
\author{
Gilberto Padilla Mercado
}
