% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_method_dispatch.R
\name{preprocess_data}
\alias{preprocess_data}
\title{Preprocess data generic}
\usage{
preprocess_data(x, ...)
}
\description{
Preprocess data generic
}
\seealso{
[preprocess_data.pk()] for the `preprocess_data` method for class [pk()]
}
