% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_methods.R
\name{print.pk}
\alias{print.pk}
\title{Print a PK object}
\usage{
\method{print}{pk}(obj)
}
\arguments{
\item{obj}{A `pk` object}
}
\value{
Invisibly: The `pk` object with added elements containing the optimization results
}
\description{
Print a
}
\details{
A `pk` object is just a list of data and fitting options. In order to
actually perform the optimization and fit the model, you need to call one of
the methods to do that -- including [print.pk()], [summarize.pk()],
[fit.pk()]. If you just type in a set of instructions like `pk(data =
my_data) + stat_model(model = c(model_flat, model_1comp, model_2comp)` and hit
Enter/Return, then by default R will call the [print.pk()] method. (This is
true no matter what you type at the R command line and hit enter -- R will
call the appropriate `print` method for an object of that class, or
[print.default()] if it can't find a class-specific print method.) Therefore,
[print_pk()] does the following:

- Pre-processes the data
- Does initial data checking and summary (e.g., number of observations by route, media, detect/nondetect)
- Determines parameters to be optimized for each specified model, based on the data
- Checks data to see whether the selected parameters may be identifiable (e.g., are there more observations than there are parameters to be estimated?)
- Determines bounds and starting values for each parameter to be optimized
- Performs optimization to estimate parameter values and uncertainties
- Adds the optimization results to the `pk` object
- Prints the `pk` object
- Returns the `pk` object invisibly
}
\author{
Caroline Ring
}
