% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_columns.R
\name{rename_columns}
\alias{rename_columns}
\title{Helper function to rename columns and coerce data to usable format}
\usage{
rename_columns(
  data.set,
  compound.col,
  cas.col,
  reference.col,
  species.col,
  species.default,
  species.weight.col,
  species.weight.units.col,
  species.weight.units.default,
  dose.col,
  time.col,
  time.units.col,
  time.units.default,
  media.col,
  media.units.col,
  media.units.default,
  value.col,
  units.col,
  units.default,
  route.col,
  route.default,
  source.col,
  source.default,
  loq.col,
  loq.default,
  subject.col,
  subject.default,
  info.col,
  info.default
)
}
\arguments{
\item{data.set}{A table of concentration-time data.}

\item{compound.col}{Column in data.set to serve as compound column}

\item{cas.col}{Column in data.set to serve as CAS column}

\item{reference.col}{Column in data.set to serve as reference column}

\item{species.col}{Column in data.set to serve as species column}

\item{species.default}{Default value to fill newly created species column, should data.set lack a column equivalent to species}

\item{species.weight.col}{Column in data.set to serve as species.weight column}

\item{species.weight.units.col}{Column in data.set to serve as species.weight.units column}

\item{species.weight.units.default}{Default value to fill newly created species.weight.units column, should data.set lack a column equivalent to species.weight.units}

\item{dose.col}{Column in data.set to serve as dose column}

\item{time.col}{Column in data.set to serve as time column}

\item{time.units.col}{Column in data.set to serve as time.units column}

\item{time.units.default}{Default value to fill newly created time.units column, should data.set lack a column equivalent to time.units}

\item{media.col}{Column in data.set to serve as media column}

\item{media.units.col}{Column in data.set to serve as media.units column}

\item{media.units.default}{Default value to fill newly created media.units coloumn, should data.set lack a column equivalent to media.units}

\item{value.col}{Column in data.set to serve as value column}

\item{units.col}{Column in data.set to serve as units column}

\item{units.default}{Default value to fill newly created units column, should data.set lack a column equivalent to units}

\item{route.col}{Column in data.set to serve as route column}

\item{route.default}{Default value to fill newly created route column, should data.set lack a column equivalent to route}

\item{source.col}{Column in data.set to serve as source column}

\item{source.default}{Default value to fill newly created source column, should data.set lack a column equivalent to source}

\item{loq.col}{Column in data.set to serve as loq column}

\item{loq.default}{Default value to fill newly created loq column, should data.set lack a column equivalent to loq}

\item{subject.col}{Column in data.set to serve as subject column}

\item{subject.default}{Default value to fill newly created subject column, should data.set lack a column equivalent to subject}

\item{info.col}{Column in data.set to serve as info column}

\item{info.default}{Default value to fill newly created info column, should data.set lack a column equivalent to info}
}
\value{
A data.frame with corrected column names
}
\description{
Helper function to rename columns and coerce data to usable format
}
\author{
Christopher Cook, John Wambaugh
}
