% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_method_dispatch.R
\name{rmse.default}
\alias{rmse.default}
\title{Root mean squared error (RMSE) default method}
\usage{
\method{rmse}{default}(obj, ...)
}
\arguments{
\item{obj}{an object}

\item{...}{Additional arguments.}
}
\value{
An error, when a non-pk object is used for the first argument.
}
\description{
Root mean squared error (RMSE) default method
}
