% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_method_dispatch.R
\name{rmse}
\alias{rmse}
\title{Root mean squared error (RMSE)}
\usage{
rmse(obj, ...)
}
\arguments{
\item{obj}{the pk object}

\item{...}{Additional arguments currently not in use.}
}
\value{
A `data.frame` with calculated RMSE as the final column. There is one row per
  each model in `obj`'s [stat_model()] element, i.e. each PK model that was
  fitted to the data, each [optimx::optimx()] methods (specified in
  [settings_optimx()]), `rmse_group` specified.
}
\description{
This is the S3 method generic for `rmse`.
}
\seealso{
[rmse.pk()] for the `rmse` method for class [pk()]
}
