% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_method_twofold_test.R
\name{rowwise_calc_percentages}
\alias{rowwise_calc_percentages}
\title{Helper function for calculating percentages of count data, by row}
\usage{
rowwise_calc_percentages(data, group_cols = NULL)
}
\arguments{
\item{data}{A data.frame that contains columns of count data and
possibly columns of group names.}

\item{group_cols}{String or numeric indices for the columns which contain
grouping variables.}
}
\value{
data.frame with rowwise totals and percentages
}
\description{
This function takes totals and calculates rowise percentages across columns
Expects columns for each percentage, can specify a vector of "grouping" column names
}
