% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_method_dispatch.R
\name{rsq}
\alias{rsq}
\title{rsq()}
\usage{
rsq(obj, ...)
}
\arguments{
\item{obj}{An object.}

\item{...}{Additional arguments currently not in use.}
}
\description{
This is the S3 method generic for rsq()
}
\seealso{
[rsq.pk()] for the method for class [pk()]
}
