% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkproto_scale_data.R
\name{scale_time}
\alias{scale_time}
\title{Scale times}
\usage{
scale_time(new_units = "identity", ...)
}
\arguments{
\item{new_units}{New units to use for time. Default is `"identity"` (leave
time in the original units). Another useful option is `"auto"`, to
automatically select new time units based on the time of the last detected
observation. You may also specify any time units understood by
`lubridate::duration()`, i.e., `"seconds"`, `"hours"`, `"days"`, `"weeks"`,
`"months"`, `"years"`, `"milliseconds"`, `"microseconds"`, `"nanoseconds"`,
and/or `"picoseconds"`. You may only specify one new unit (e.g., `new_units
= c("days", "weeks")` is not valid).}

\item{...}{Other arguments (not currently used)}
}
\value{
An object of class `pk_scales`: A named list with two elements `name
  = "time"` (denoting the variable to be scaled) and `value = list("new_units" =
  new_units, ...)` (denoting the arguments supplied to [scale_time()]). See
  [pk_add.pk_scales()].
}
\description{
Transform time data
}
