% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkproto_settings.R
\name{settings_data_info}
\alias{settings_data_info}
\title{Data info settings}
\usage{
settings_data_info(
  nca_group = dplyr::vars(Chemical, Species, Reference, Route, Media, Dose),
  ...
)
}
\arguments{
\item{nca_group}{A set of variables. Data will be split into groups according
to unique combinations of these variables, and non-compartmental analysis
will be performed separately on each group. Default `dplyr::vars(Chemical,
Species, Reference, Route, Media, Dose)`.}

\item{...}{Other arguments (currently ignored).}
}
\value{
An object of class `c(pkproto, pk_settings_data_info)`
}
\description{
Data info settings
}
