% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkproto_settings_optimx.R
\name{settings_optimx}
\alias{settings_optimx}
\title{`optimx` optimizer settings}
\usage{
settings_optimx(
  method = c("bobyqa", "L-BFGS-B"),
  hessian = FALSE,
  control = list(kkt = FALSE, maxit = 1e+07),
  ...
)
}
\arguments{
\item{method}{The name(s) of optimization methods to be used. See
[optimx::opm()] for options. Default is `"bobyqa"` and `"L-BFGS-B"`.}

\item{hessian}{Whether to compute the Hessian at the final set of parameters; as in [optimx::opm()].}

\item{control}{A list of control parameters for the optimizer; see [optimx::opm()] for options and details.}

\item{...}{Additional arguments not currently implemented.}
}
\value{
An object of class `pk_settings`.
}
\description{
`optimx` optimizer settings
}
\author{
Caroline Ring
}
