% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkproto_stat_preprocess_groups.R
\name{stat_loq_group}
\alias{stat_loq_group}
\title{LOQ Group}
\usage{
stat_loq_group(...)
}
\arguments{
\item{...}{A set of unquoted variables whose unique combinations define
a group with which to impute missing LOQ values (using the
minimum non-missing LOQ value in the group multiplied by `calc_loq_factor`).
Default is `Chemical, Species, Reference, Media`.}
}
\value{
A list of expressions. This is added to the `pk` object.
}
\description{
Defines the grouping variables for LOQ imputation in [do_preprocess.pk()].
}
\author{
Gilberto Padilla Mercado
}
