% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkproto_stat_nca_group.R
\name{stat_nca_group}
\alias{stat_nca_group}
\title{NCA group}
\usage{
stat_nca_group(...)
}
\arguments{
\item{...}{A set of variables. Data will be split into groups according
to unique combinations of these variables, and non-compartmental analysis
will be performed separately on each group. Default `Chemical,
Species, Reference, Route, Media, Dose`.}
}
\value{
An object of class `c(pkproto, pk_settings_data_info)`
}
\description{
NCA group
}
