% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkproto_stat_preprocess_groups.R
\name{stat_sd_group}
\alias{stat_sd_group}
\title{SD Group}
\usage{
stat_sd_group(...)
}
\arguments{
\item{...}{A set of unquoted variables whose unique combinations define
a group with which to impute missing SD values (using the
minimum non-missing SD value in the group).
Default is `Chemical, Species, Reference, Media`.}
}
\value{
A list of expressions. This is added to the `pk` object.
}
\description{
Defines the grouping to calculate standard deviation of data in [do_preprocess.pk()].
}
\author{
Gilberto Padilla Mercado
}
