% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{status_data_info}
\alias{status_data_info}
\title{Status ID for data summary info}
\format{
An object of class \code{integer} of length 1.
}
\usage{
status_data_info
}
\description{
An integer status that denotes [data_info()] has been completed.
}
\keyword{datasets}
