% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{status_fit}
\alias{status_fit}
\title{Status ID for fitting}
\format{
An object of class \code{integer} of length 1.
}
\usage{
status_fit
}
\description{
An integer status that denotes [fit()] has been completed.
}
\keyword{datasets}
