% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{status_init}
\alias{status_init}
\title{Status ID for initialization}
\format{
An object of class \code{integer} of length 1.
}
\usage{
status_init
}
\description{
An integer status that denotes a [pk()] object has been initialized.
}
\keyword{datasets}
