% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{status_prefit}
\alias{status_prefit}
\title{Status ID for pre-fitting}
\format{
An object of class \code{integer} of length 1.
}
\usage{
status_prefit
}
\description{
An integer status that denotes [prefit()] has been completed.
}
\keyword{datasets}
