% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{status_preprocess}
\alias{status_preprocess}
\title{Status ID for preprocessing}
\format{
An object of class \code{integer} of length 1.
}
\usage{
status_preprocess
}
\description{
An integer status that denotes [preprocess_data()] has been completed.
}
\keyword{datasets}
