% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{str_replace_lst}
\alias{str_replace_lst}
\title{Wrapper for [stringr::str_replace_all()] which replaces whole words.}
\usage{
str_replace_lst(.x, .lst)
}
\arguments{
\item{.x}{A vector to match against.}

\item{.lst}{A named vector in the form `old_value = new_value`.}
}
\value{
A character vector with values substituted in.
}
\description{
Wrapper for [stringr::str_replace_all()] which replaces whole words.
}
