% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{string2num}
\alias{string2num}
\title{Converting comma-delimited string representations of numeric vectors to numerics}
\usage{
string2num(x)
}
\arguments{
\item{x}{A character vector that represents numeric value(s)}
}
\value{
A numeric vector of values or `NA_real_`
}
\description{
Converting comma-delimited string representations of numeric vectors to numerics
}
\examples{
char2num(" ")
char2num("22.2")
char2num("12.3,4.1,10.1")

}
