% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_methods_subtract_pkproto.R
\name{subtract_pk}
\alias{subtract_pk}
\title{Subtract various `pkproto` objects from a `pk` object}
\usage{
subtract_pk(pk_obj, object, objectname)
}
\arguments{
\item{pk_obj}{The `pk` object}

\item{object}{The `pkproto` object to be subtracted}

\item{objectname}{The name of the `pkproto` object to be subtracted}
}
\value{
The `pk` object modified by the subtraction.
}
\description{
Subtract various `pkproto` objects from a `pk` object
}
