% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_faceted_methods_summary.R,
%   R/pk_methods_summary.R
\name{summary.pk}
\alias{summary.pk}
\title{summary}
\usage{
\method{summary}{pk}(obj)

\method{summary}{pk}(obj)
}
\arguments{
\item{obj}{A [pk] object.}
}
\value{
A [tibble::tibble()] grouped by the faceting variables with a new variable ` summary ` containing the results of applying [summary.pk()] to each item in the list column of [pk()] objects, `obj$pk_object`

A `data.frame` consisting of a summary table of fitting options and results.
}
\description{
summary for pk_faceted objects

This summary includes summary information about the data; about any data
transformations applied; about the models being fitted; about the error model
being applied; and any fitting results, if the `pk` object has been fitted.
It also includes TK quantities calculated from the fitted model parameters,
e.g. halflife; clearance; tmax; Cmax; AUC; Css.
}
\seealso{
Other methods for pk_faceted objects: 
\code{\link{AIC.pk}()},
\code{\link{BIC.pk}()},
\code{\link{check_needed_status.pk}()},
\code{\link{coef.pk}()},
\code{\link{coef_sd.pk}()},
\code{\link{compare_models.pk}()},
\code{\link{data_summary.pk}()},
\code{\link{eval_tkstats.pk}()},
\code{\link{fold_error.pk}()},
\code{\link{get_data.pk}()},
\code{\link{get_data_group.pk}()},
\code{\link{get_data_info.pk}()},
\code{\link{get_data_original.pk}()},
\code{\link{get_data_sigma_group.pk}()},
\code{\link{get_error_group.pk}()},
\code{\link{get_fit.pk}()},
\code{\link{get_mapping.pk}()},
\code{\link{get_nca.pk}()},
\code{\link{get_prefit.pk}()},
\code{\link{get_scale_conc.pk}()},
\code{\link{get_scale_time.pk}()},
\code{\link{get_settings_data_info.pk}()},
\code{\link{get_settings_optimx.pk}()},
\code{\link{get_settings_preprocess.pk}()},
\code{\link{get_stat_error_model.pk}()},
\code{\link{get_stat_model.pk}()},
\code{\link{get_status.pk}()},
\code{\link{get_tkstats.pk}()},
\code{\link{get_winning_model.pk}()},
\code{\link{is.pk}()},
\code{\link{logLik.pk}()},
\code{\link{nca.pk}()},
\code{\link{plot.pk}()},
\code{\link{predict.pk}()},
\code{\link{preprocess_data.pk_faceted}()},
\code{\link{residuals.pk}()},
\code{\link{rmse.pk}()},
\code{\link{rsq.pk}()},
\code{\link{subtract_pkproto.pk}()}
}
\author{
Caroline Ring
}
\concept{methods for pk_faceted objects}
