% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{time_conversions}
\alias{time_conversions}
\title{Time conversion table}
\format{
An object of class \code{data.frame} with 121 rows and 3 columns.
}
\usage{
time_conversions
}
\description{
A `data.frame` that has the converted units from "time_units"
}
\keyword{datasets}
