% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{time_units}
\alias{time_units}
\title{Allowable time units}
\format{
An object of class \code{character} of length 11.
}
\usage{
time_units
}
\description{
A `character` vector of allowable units for time variables.
}
\details{
These are the time units understood by [lubridate::period()] and [lubridate::duration()].
}
\keyword{datasets}
