% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformed_params_2comp.R
\name{transformed_params_2comp}
\alias{transformed_params_2comp}
\title{Transformed parameters for 2-compartment model}
\usage{
transformed_params_2comp(params, ...)
}
\arguments{
\item{params}{A named numeric vector of parameters for the 2-compartment
model. Any missing parameters will be filled with `NA_real_`.}

\item{...}{Additional arguments (not currently used).}
}
\value{
A named numeric vector of transformed parameters with elements
 "alpha", "beta", "A_iv_unit", "B_iv_unit", "A_oral_unit", "B_oral_unit".
}
\description{
Transformed parameters for 2-compartment model
}
\seealso{
Other built-in model functions: 
\code{\link{auc_1comp}()},
\code{\link{auc_1comp_cl}()},
\code{\link{auc_1comp_rad}()},
\code{\link{auc_2comp}()},
\code{\link{auc_flat}()},
\code{\link{cp_1comp}()},
\code{\link{cp_1comp_cl}()},
\code{\link{cp_1comp_rad}()},
\code{\link{cp_2comp}()},
\code{\link{cp_2comp_dt}()},
\code{\link{cp_flat}()},
\code{\link{cp_httk_3comp2}()},
\code{\link{get_params_1comp}()},
\code{\link{get_params_1comp_cl}()},
\code{\link{get_params_1comp_fup}()},
\code{\link{get_params_1comp_rad}()},
\code{\link{get_params_2comp}()},
\code{\link{get_params_flat}()},
\code{\link{get_params_httk_3comp2}()},
\code{\link{get_starts_1comp}()},
\code{\link{get_starts_1comp_cl}()},
\code{\link{get_starts_1comp_fup}()},
\code{\link{get_starts_1comp_rad}()},
\code{\link{get_starts_2comp}()},
\code{\link{get_starts_flat}()},
\code{\link{get_starts_httk_3comp2}()},
\code{\link{tkstats_2comp}()}

Other 2-compartment model functions: 
\code{\link{auc_2comp}()},
\code{\link{cp_2comp}()},
\code{\link{cp_2comp_dt}()},
\code{\link{get_params_2comp}()},
\code{\link{get_starts_2comp}()},
\code{\link{tkstats_2comp}()}
}
\author{
Caroline Ring, Gilberto Padilla Mercado, John Wambaugh
}
\concept{2-compartment model functions}
\concept{built-in model functions}
