% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_method_dispatch.R
\name{twofold_test.default}
\alias{twofold_test.default}
\title{Default method for twofold_test()}
\usage{
\method{twofold_test}{default}(obj, ...)
}
\arguments{
\item{obj}{An object}

\item{...}{Additional arguments currently not in use.}
}
\description{
Default method for twofold_test()
}
