% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_method_twofold_test.R
\name{twofold_test.pk}
\alias{twofold_test.pk}
\title{Evaluate whether data and predictions are within two-fold of mean or concentration, respectively}
\usage{
\method{twofold_test}{pk}(obj, newdata = NULL, suppress_messages = FALSE)
}
\arguments{
\item{obj}{A pk object.}

\item{newdata}{A data.frame with new data. Uses the same column names as obj$data}

\item{suppress_messages}{Logical, default: FALSE.
Should there be no printed message output?}
}
\value{
A list of data frames.
}
\description{
Evaluate whether data and predictions are within two-fold of mean or concentration, respectively
}
\author{
Gilberto Padilla Mercado
}
