% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_model.R
\name{adjust_model_name}
\alias{adjust_model_name}
\title{Sets the `name` element for models to the `pk_model` object name in the environment}
\usage{
adjust_model_name(model)
}
\arguments{
\item{model}{A `pk_model` object.}
}
\value{
an object of class `pk_model` with `name` matching it's name in the environment.
}
\description{
When creating new `pk_model` objects, the name of the 'base' model is kept.
Please use this function to 'reset' the name of the new `pk_model` object.
}
\author{
Gilberto Padilla Mercado
}
