% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_method_dispatch.R
\name{AFE.default}
\alias{AFE.default}
\title{Default method for AFE()}
\usage{
\method{AFE}{default}(obj, ...)
}
\arguments{
\item{obj}{An object}

\item{...}{Additional arguments currently not in use.}
}
\description{
Default method for AFE()
}
