% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_methods_AIC.R
\name{AIC.pk}
\alias{AIC.pk}
\title{Akaike information criterion}
\usage{
\method{AIC}{pk}(
  obj,
  newdata = NULL,
  model = NULL,
  method = NULL,
  exclude = TRUE,
  drop_obs = TRUE,
  k = 2
)
}
\arguments{
\item{obj}{A `pk` object}

\item{newdata}{Optional: A `data.frame` with new data for which to compute
log-likelihood. If NULL (the default), then log-likelihoods will be
computed for the data in `obj$data`. `newdata` is required to contain at
least the following variables: `Time`, `Time.Units`, `Dose`,
`Route`,`Media`, `Conc`, `Detect`, `N_Subjects`. Before log-likelihood is
calculated, `Time` will be transformed according to the transformation in
`obj$scales$time` and `Conc` will be transformed according to the
transformation in `obj$scales$conc`.}

\item{model}{Optional: Specify one or more of the fitted models for which to
calculate log-likelihood. If NULL (the default), log-likelihoods will be
returned for all of the models in `obj$stat_model`.}

\item{method}{Optional: Specify one or more of the [optimx::optimx()] methods
for which to make predictions and calculate AICs. If NULL (the default),
log-likelihoods will be returned for all of the models in
`obj$settings_optimx$method`.}

\item{exclude}{Logical: `TRUE` to compute the AIC after removing any
observations in the data marked for exclusion (if there is a variable
`exclude` in the data, an observation is marked for exclusion when `exclude
%in% TRUE`). `FALSE` to include all observations, regardless of exclusion
status. Default `TRUE`.}

\item{k}{Default 2. The `k` parameter in the log-likelihood formula (see
Details).}
}
\value{
A named list of numeric vectors. There is one list element named for
  each model in `obj`'s [stat_model()] element, i.e. each PK model that was
  fitted to the data. Each list element is a numeric vector with as many
  elements as there were [optimx::optimx()] methods (specified in
  [settings_optimx()]). The vector names are the method names.  Each vector
  element contains the AIC of the model fitted by the corresponding method,
  using the data in `newdata`.
}
\description{
Get the Akaike information criterion (AIC) for a fitted `pk` object
}
\details{
The AIC is calculated from the log-likelihood (LL) as follows:
\deqn{\textrm{AIC} = -2\textrm{LL} + k n_{par}}

where \eqn{n_{par}} is the number of parameters in the fitted model, and
\eqn{k = 2} for the standard AIC.
}
\seealso{
Other fit evaluation metrics: 
\code{\link{BIC.pk}()},
\code{\link{logLik.pk}()},
\code{\link{rmse.pk}()},
\code{\link{rsq.pk}()}

Other log likelihood functions: 
\code{\link{BIC.pk}()},
\code{\link{logLik.pk}()}

Other methods for fitted pk objects: 
\code{\link{BIC.pk}()},
\code{\link{coef.pk}()},
\code{\link{coef_sd.pk}()},
\code{\link{eval_tkstats.pk}()},
\code{\link{get_fit.pk}()},
\code{\link{get_tkstats.pk}()},
\code{\link{logLik.pk}()},
\code{\link{predict.pk}()},
\code{\link{residuals.pk}()},
\code{\link{rmse.pk}()},
\code{\link{rsq.pk}()}
}
\author{
Caroline Ring
}
\concept{fit evaluation metrics}
\concept{log likelihood functions}
\concept{methods for fitted pk objects}
