% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_method_dispatch.R
\name{check_required_status.default}
\alias{check_required_status.default}
\title{Default method for checking required status}
\usage{
\method{check_required_status}{default}(obj, ...)
}
\arguments{
\item{obj}{An object.}

\item{...}{Additional arguments currently not in use.}
}
\value{
An error, when a non-pk object is used for the first argument.
}
\description{
Default method for checking required status
}
