% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_method_dispatch.R
\name{coef_sd.default}
\alias{coef_sd.default}
\title{Coefficient standard deviation default}
\usage{
\method{coef_sd}{default}(object, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Additional arguments currently not in use.}
}
\value{
An error, when a non-pk object is used for the first argument.
}
\description{
Coefficient standard deviation default
}
