% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_method_dispatch.R
\name{coef_sd}
\alias{coef_sd}
\title{Coefficient standard deviations}
\usage{
coef_sd(obj, ...)
}
\arguments{
\item{obj}{An object}
}
\description{
This is the S3 method generic for `coef_sd`.
}
\seealso{
[coef_sd.pk()] for the `coef_sd` method for class [pk()]
}
