% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_method_dispatch.R
\name{compare_models.default}
\alias{compare_models.default}
\title{Default method for compare_models()}
\usage{
\method{compare_models}{default}(obj, ...)
}
\arguments{
\item{obj}{an object}

\item{...}{Additional arguments currently not in use.}
}
\description{
Default method for compare_models()
}
