% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_faceted_methods_compare_models.R,
%   R/pk_methods_compare_models.R
\name{compare_models.pk}
\alias{compare_models.pk}
\title{compare_models}
\usage{
\method{compare_models}{pk}(
  obj,
  newdata = NULL,
  model = NULL,
  method = NULL,
  criterion = "AIC",
  ...
)

\method{compare_models}{pk}(
  obj,
  newdata = NULL,
  model = NULL,
  method = NULL,
  criterion = "AIC",
  ...
)
}
\arguments{
\item{obj}{A [pk()] model object. Must be fitted, or the function will exit
with an error.}

\item{newdata}{Optional: A `data.frame` containing new data for which to
compute the TK stats. Must contain at least variables `Chemical`,
`Species`, `Route`, `Media`, `Dose`, and any other variables named in
`tk_grouping`. Default `NULL`, to use the data in `obj$data`.}

\item{model}{Character: One or more of the models fitted. Default `NULL` to
return TK stats for all models.}

\item{method}{Character: One or more of the [optimx::optimx()] methods used.
Default `NULL` to return TK stats for all methods.}

\item{criterion}{The name of a criterion function to use for model
comparison. Default "AIC". Must be the name of a function that (as for
`AIC`) accepts arguments `obj`, `newdata`, `method` and `model` (may accept
other arguments, specified in `...`) and returns output as for `AIC`: a
named list of numeric vectors (named for each of the model names in
`model`), where each vector has elements named for each of the method names
in `method`, containing the criterion value calculated for that model
fitted using that method.}

\item{...}{Optional: Other arguments to `criterion` function.}

\item{tk_group}{A list of variables provided using a `dplyr::vars()` call.
The data (either `newdata` or `obj$data`) will be grouped according to the
unique combinations of these variables. For each unique combination of
these variables in the data, a set of TK statistics will be computed. The
default is `obj$data_settings$nca_group`, to derive TK statistics for the
same groups of data as non-compartmental analysis statistics. With the
default, you can directly compare e.g. a model-predicted AUC_inf to the
corresponding NCA-estimated AUC_inf. However, you may specify a different
data grouping if you wish. Each group should have a unique combination of
`Chemical`, `Species`, `Route`, `Media`, and `Dose`.}
}
\value{
A [tibble::tibble()] grouped by the faceting variables with a new variable ` compare_models ` containing the results of applying [compare_models.pk()] to each item in the list column of [pk()] objects, `obj$pk_object`

A `data.frame` with variables
- `model`: The name of each model
- `method`: The name of each method
- A variable named for `criterion` (e.g. if `criterion = "AIC"` then the result will have a
  variable named `AIC`): The criterion value for each model/method
}
\description{
compare_models for pk_faceted objects

Perform model comparison for a fitted [pk()] object.
}
\details{
Models are compared according to the goodness-of-fit criterion named in
"criterion", and the name of the winning model is returned.
}
\seealso{
Other methods for pk_faceted objects: 
\code{\link{AIC.pk}()},
\code{\link{BIC.pk}()},
\code{\link{check_needed_status.pk}()},
\code{\link{coef.pk}()},
\code{\link{coef_sd.pk}()},
\code{\link{data_summary.pk}()},
\code{\link{eval_tkstats.pk}()},
\code{\link{fold_error.pk}()},
\code{\link{get_data.pk}()},
\code{\link{get_data_group.pk}()},
\code{\link{get_data_info.pk}()},
\code{\link{get_data_original.pk}()},
\code{\link{get_data_sigma_group.pk}()},
\code{\link{get_error_group.pk}()},
\code{\link{get_fit.pk}()},
\code{\link{get_mapping.pk}()},
\code{\link{get_nca.pk}()},
\code{\link{get_prefit.pk}()},
\code{\link{get_scale_conc.pk}()},
\code{\link{get_scale_time.pk}()},
\code{\link{get_settings_data_info.pk}()},
\code{\link{get_settings_optimx.pk}()},
\code{\link{get_settings_preprocess.pk}()},
\code{\link{get_stat_error_model.pk}()},
\code{\link{get_stat_model.pk}()},
\code{\link{get_status.pk}()},
\code{\link{get_tkstats.pk}()},
\code{\link{get_winning_model.pk}()},
\code{\link{is.pk}()},
\code{\link{logLik.pk}()},
\code{\link{nca.pk}()},
\code{\link{plot.pk}()},
\code{\link{predict.pk}()},
\code{\link{preprocess_data.pk_faceted}()},
\code{\link{residuals.pk}()},
\code{\link{rmse.pk}()},
\code{\link{rsq.pk}()},
\code{\link{subtract_pkproto.pk}()},
\code{\link{summary.pk}()}
}
\author{
Caroline Ring
}
\concept{methods for pk_faceted objects}
