% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_method_dispatch.R
\name{compare_models}
\alias{compare_models}
\title{Model comparison}
\usage{
compare_models(obj, ...)
}
\arguments{
\item{obj}{An object.}

\item{...}{Additional arguments currently not in use.}
}
\description{
This is the S3 method generic for compare_models()
}
\seealso{
[compare_models.pk()] for the method for class [pk()]
}
