% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cvt}
\alias{cvt}
\title{CvTdb data}
\format{
An object of class \code{data.frame} with 14152 rows and 56 columns.
}
\usage{
cvt
}
\description{
Concentration vs. time data from CvTdb
}
\details{
This is concentration vs. time data from CvTdb, most recently downloaded as
of the date in [`cvt_date`].

These data have been filtered to retain only oral and intravenous
administration, and only measurements in blood and plasma. They have also
been filtered to retain only observations where the same chemical was both
administered and measured in blood/plasma (i.e., excluding observations where
a metabolite was measured).

A `data.frame` with 14805 rows and 105 variables.
}
\keyword{datasets}
