% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cvtdb_original}
\alias{cvtdb_original}
\title{SQL query result (current)}
\format{
An object of class \code{data.frame} with 42443 rows and 56 columns.
}
\usage{
cvtdb_original
}
\description{
This is the raw SQL query result.
}
\details{
A data.frame similar to [`cvt`] and [`cvt_2.0.0`], but to create those objects
some values are changed to normalized values for use with invivoPKfit specifically
and to include experiments that may not pass filtering due to data coding
details, but are reasonable to include for analysis.
}
\keyword{datasets}
