% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_method_dispatch.R
\name{do_prefit}
\alias{do_prefit}
\title{Prefitting}
\usage{
do_prefit(obj, ...)
}
\arguments{
\item{obj}{the pk object}

\item{...}{Additional arguments currently not in use.}
}
\description{
Prefitting
}
\seealso{
[do_prefit.pk()] for the `do_prefit` method for class [pk()]
}
