% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_method_dispatch.R
\name{do_preprocess}
\alias{do_preprocess}
\title{Preprocess data generic}
\usage{
do_preprocess(obj, ...)
}
\arguments{
\item{obj}{the pk object.}

\item{...}{Additional arguments currently not in use.}
}
\description{
Preprocess data generic
}
\seealso{
[do_preprocess.pk()] for the `do_preprocess` method for class [pk()]
}
