% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_method_dispatch.R
\name{do_preprocess}
\alias{do_preprocess}
\title{Preprocess data generic}
\usage{
do_preprocess(obj, ...)
}
\arguments{
\item{obj}{the pk object.}

\item{...}{Additional arguments currently not in use.}
}
\value{
The same `pk` object, with added elements `data` (containing the
  cleaned, gap-filled data) and `data_info` (containing summary information
  about the data, e.g. number of observations by route, media,
  detect/nondetect; empirical tmax, time of peak concentration for oral data;
  number of observations before and after empirical tmax)
}
\description{
Preprocess data generic
}
\seealso{
[do_preprocess.pk()] for the `do_preprocess` method for class [pk()]
}
