% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_method_dispatch.R
\name{eval_tkstats.default}
\alias{eval_tkstats.default}
\title{Default method for eval_tkstats()}
\usage{
\method{eval_tkstats}{default}(obj, ...)
}
\arguments{
\item{obj}{An object}
}
\description{
Default method for eval_tkstats()
}
