% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_method_dispatch.R
\name{eval_tkstats}
\alias{eval_tkstats}
\title{eval_tkstats()}
\usage{
eval_tkstats(obj, ...)
}
\arguments{
\item{obj}{An object.}

\item{...}{Additional arguments currently not in use.}
}
\value{
A `data.frame` with one  row for each "winning" model in
  `model` from [get_winning_model()]. The `data.frame` will have the variables
  returned by the `tkstats_fun` for its corresponding model. (For the
  built-in models `model_flat`, `model_1comp`, and `model_2comp`, these
  variables are `param_name` and `param_value`.) Additionally, there will be
  a variable `method` denoting the [optimx::optimx()] method used to optimize
  the set of model parameters used to derive each set of TK statistics.
}
\description{
This is the S3 method generic for eval_tkstats()
}
\seealso{
[eval_tkstats.pk()] for the method for class [pk()]
}
