% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_method_dispatch.R
\name{eval_tkstats}
\alias{eval_tkstats}
\title{eval_tkstats()}
\usage{
eval_tkstats(obj, ...)
}
\arguments{
\item{obj}{An object.}
}
\description{
This is the S3 method generic for eval_tkstats()
}
\details{
`eval_tkstats()` is an alias for `data_summary()`
}
\seealso{
[data_summary.pk()] for the method for class [pk()]
}
