% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_method_dispatch.R
\name{fold_error.default}
\alias{fold_error.default}
\title{fold_error default method}
\usage{
\method{fold_error}{default}(obj, ...)
}
\arguments{
\item{obj}{an object}

\item{...}{Additional arguments currently not in use.}
}
\description{
fold_error default method
}
