% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_methods_fold_error.R
\name{fold_errors.pk}
\alias{fold_errors.pk}
\title{Fold errors}
\usage{
\method{fold_errors}{pk}(obj, newdata = NULL, model = NULL, method = NULL)
}
\arguments{
\item{obj}{A `pk` object}

\item{newdata}{Optional: A `data.frame` with new data for which to compute
fold errors. If NULL (the default), then fold errors will be computed for
the data in `obj$data`. `newdata` is required to contain at least the
following variables: `Time`, `Dose`, `Route`, and `Media`. `Time` will be
transformed according to the transformation in `obj$scales$time` before
fold errors are calculated.}

\item{model}{Optional: Specify one or more of the fitted models for which to
make predictions and calculate fold errors. If NULL (the default), fold errors will be
returned for all of the models in `obj$stat_model`.}

\item{method}{Optional: Specify one or more of the [optimx::optimx()] methods
for which to make predictions and calculate RMSEs. If NULL (the default),
fold errors will be returned for all of the models in
`obj$optimx_settings$method`.}
}
\value{
A named list of numeric matrices. There is one list element named for
  each model in `obj`'s [stat_model()] element, i.e. each PK model that was
  fitted to the data. Each list element is a matrix with the same number of
  rows as the data in `obj$data` (corresponding to the rows in `obj$data`),
  and as many columns as there were [optimx::optimx()] methods (specified in
  [settings_optimx()]). The column names are the method names.  Each column
  contains the fold errors (observed/predicted) of the model fitted by the
  corresponding method. These residuals are concentrations in the same units
  as `obj$data$Conc.Units`; any concentration transformations (in
  `obj$scale$conc`) are *not* applied.
}
\description{
Calculate fold errors for a fitted `pk` object.
}
\author{
Caroline Ring
}
