% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_method_dispatch.R
\name{fold_errors}
\alias{fold_errors}
\title{Fold error}
\usage{
fold_errors(x)
}
\description{
This is the S3 method generic for `fold_errors`.
}
\seealso{
[fold_errors.pk()] for the `fold_errors` method for class [pk()]
}
