% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_methods_get_data_group.R
\name{get_data_group.pk}
\alias{get_data_group.pk}
\title{Get data groupingh}
\usage{
\method{get_data_group}{pk}(obj)
}
\arguments{
\item{obj}{An initialized `pk` object}
}
\value{
An object of class `call` giving the data grouping as a `dplyr::vars()` specification
}
\description{
Get data groupingh
}
