% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_method_dispatch.R
\name{get_data_summary}
\alias{get_data_summary}
\title{get_data_summary()}
\usage{
get_data_summary(obj, ...)
}
\arguments{
\item{obj}{An object.}

\item{...}{Additional arguments currently not in use.}
}
\description{
This is the S3 method generic for get_data_summary()
}
\details{
`get_data_summary()` is an alias for `data_summary()`
}
\seealso{
[data_summary.pk()] for the method for class [pk()]
}
