% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_methods_get_error_group.R
\name{get_error_group.pk}
\alias{get_error_group.pk}
\title{Get error group}
\usage{
\method{get_error_group}{pk}(obj)
}
\arguments{
\item{obj}{A [pk()] object}
}
\value{
The stat_error_model error grouping
}
\description{
Get error group
}
\author{
Caroline Ring
}
