% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_method_dispatch.R
\name{get_fit}
\alias{get_fit}
\title{get_fit()}
\usage{
get_fit(obj, ...)
}
\arguments{
\item{obj}{An object.}

\item{...}{Additional arguments currently not in use.}
}
\value{
A named list of objects of class `optimx`, named for the models in
  `model`. As described in [optimx::optimx()]  If only one model is
  specified, the return value will still be a list, but with only one
  element.
}
\description{
This is the S3 method generic for get_fit()
}
\seealso{
[get_fit.pk()] for the method for class [pk()]
}
