% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_method_dispatch.R
\name{get_hessian}
\alias{get_hessian}
\title{get_hessian()}
\usage{
get_hessian(obj, ...)
}
\arguments{
\item{obj}{An object.}

\item{...}{Additional arguments currently not in use.}
}
\description{
This is the S3 method generic for hessian()
}
\seealso{
[hessian.pk()] for the method for class [pk()]
}
