% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_method_dispatch.R
\name{get_nca}
\alias{get_nca}
\title{get_nca()}
\usage{
get_nca(obj, ...)
}
\arguments{
\item{obj}{An object.}

\item{...}{Additional arguments currently not in use.}
}
\description{
This is the S3 method generic for get_nca()
}
\seealso{
[get_nca.pk()] for the method for class [pk()]
}
