% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_methods_get_prefit.R
\name{get_prefit.pk}
\alias{get_prefit.pk}
\title{Get prefit}
\usage{
\method{get_prefit}{pk}(obj, ...)
}
\arguments{
\item{obj}{A [pk()] object that has had `do_prefit()` run on it}

\item{...}{Additional arguments. Currently not in use.}
}
\value{
A list of `data.frame`s in the object's `prefit` element.
}
\description{
Extract pre-fitting results from a [pk()] object
}
\author{
Caroline Ring
}
