% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_methods_get_scale_time.R
\name{get_scale_time.pk}
\alias{get_scale_time.pk}
\title{Get scale_time}
\usage{
\method{get_scale_time}{pk}(obj, ...)
}
\arguments{
\item{obj}{A [pk()] object}

\item{...}{Additional arguments not in use.}
}
\value{
A `list`: `obj$scales$time`
}
\description{
Extract time scale/transformation instructions from a [pk()] object
}
\author{
Caroline Ring
}
