% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_method_dispatch.R
\name{get_scale_time}
\alias{get_scale_time}
\title{get_scale_time()}
\usage{
get_scale_time(obj, ...)
}
\arguments{
\item{obj}{An object.}

\item{...}{Additional arguments currently not in use.}
}
\value{
A `list`: `obj$scales$time`
}
\description{
This is the S3 method generic for get_scale_time()
}
\seealso{
[get_scale_time.pk()] for the method for class [pk()]
}
