% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_method_dispatch.R
\name{get_status.default}
\alias{get_status.default}
\title{Default method for getting status}
\usage{
\method{get_status}{default}(obj, ...)
}
\arguments{
\item{obj}{an object}

\item{...}{Additional arguments currently not in use.}
}
\description{
Default method for getting status
}
