% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_method_dispatch.R
\name{get_tkstats}
\alias{get_tkstats}
\title{Get TK stats}
\usage{
get_tkstats(obj, ...)
}
\arguments{
\item{obj}{an object}

\item{...}{Additional arguments currently not in use.}
}
\value{
A data.frame with one row for each `data_group`, `model` and `method`
  with the variables in the `data.frame` returned by the `tkstats_fun` for
  its corresponding model.
  (For the built-in models `model_flat`, `model_1comp`, and `model_2comp`, these
  variables are `param_name` and `param_value`.)
}
\description{
This is the S3 method generic for get_tkstats(0)
}
\seealso{
[get_tkstats.pk()] for the method for class [pk()]
}
