% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_method_dispatch.R
\name{get_tkstats}
\alias{get_tkstats}
\title{Get TK stats}
\usage{
get_tkstats(obj, ...)
}
\arguments{
\item{obj}{An object.}
}
\description{
This is the S3 method generic for get_tkstats(0)
}
\seealso{
[get_tkstats.pk()] for the method for class [pk()]
}
