% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_method_dispatch.R
\name{get_winning_model}
\alias{get_winning_model}
\title{get_winning_model()}
\usage{
get_winning_model(obj, ...)
}
\arguments{
\item{obj}{An object.}

\item{...}{Additional arguments currently not in use.}
}
\value{
A data.frame with one row for each `data_group`, `model` and `method` and
 The return value has attribute `criterion` giving the name of the criterion function used to compare
 models.
}
\description{
This is the S3 method generic for get_winning_model()
}
\seealso{
[get_winning_model.pk()] for the method for class [pk()]
}
