% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hessian_funs.R
\name{hess_sd1}
\alias{hess_sd1}
\title{Inverse diagonal, method 1}
\usage{
hess_sd1(m)
}
\arguments{
\item{m}{A square numeric matrix.}
}
\value{
A numeric vector of length `n`.
}
\description{
Get square root of diagonal of inverse matrix, first method
}
\details{
Invert a square numeric matrix `m` of size \eqn{n \times n} using [solve()],
then take the square root of the diagonal.
}
\author{
Caroline Ring
}
