% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{model_httk_gas_pbtk}
\alias{model_httk_gas_pbtk}
\title{Gas pbtk `httk` model}
\format{
An object of class \code{list} (inherits from \code{pk_model}) of length 12.
}
\usage{
model_httk_gas_pbtk
}
\description{
The `pk_model` object defining the "gas_pbtk" model from `httk`, with
recalculations of other "constant" parameters that depend on optimized parameters.
}
\details{
A `pk_model` object: under the hood, a `list` object with named elements
corresponding to the arguments of [pk_model()]. The default set of parameters
to optimize are `Clint` and `Funbound.plasma`, but users can also fit partition coefficients.
Note that fitting may present instability during ODE solving step.
See that function documentation for the definition of each element.

See [cp_httk_gas_pbtk()] for the function that predicts blood/plasma concentration for a bolus dose (oral or IV).

See [get_params_httk_gas_pbtk()] for the function that determines bounds and starting guesses for model parameters, based on the data.
}
\keyword{datasets}
