% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_methods.R
\name{nca.pk}
\alias{nca.pk}
\title{Non-compartmental analysis}
\usage{
nca.pk(obj, newdata = NULL, dose_norm = TRUE)
}
\arguments{
\item{obj}{A `pk` object containing concentration-time data in the element `obj$data`.}

\item{dose_norm}{Whether to dose-normalize data before performing NCA. Default TRUE.}
}
\value{
A `data.frame` with variables as listed in Details.
}
\description{
Do non-compartmental analysis of a `pk` object.
}
\details{
This function calls [PK::nca()] to calculate the following quantities:

For intravenous (IV) bolus administration:

* `nca.iv.AUC_tlast`: AUC (area under concentration-time curve) evaluated at the last reported time point
* `nca.iv.AUC_inf`: AUC evaluated at time t = \eqn{\infty}
* `nca.iv.AUMC_inf`: AUMC (area under the first moment curve, i.e. under the AUC vs. time curve) evaluated at time \eqn{t = \infty}
* `nca.iv.MRT`: MRT (mean residence time)
* `nca.iv.halflife`: Half-life
* `nca.iv.Clearance`: Clearance
* Vss: apparent volume of distribution at steady-state.

For oral bolus administration:

* `nca.oral.AUC_tlast`: AUC (area under concentration-time curve) evaluated at the last reported time point
* `nca.oral.AUC_inf`: AUC evaluated at time \eqn{t = \infty}
* `nca.oral.AUMC_inf`: AUMC (area under the first moment curve, i.e. under the AUC vs. time curve) evaluated at time \eqn{t = \infty}
* `nca.oral.MTT`: MTT (mean transit time), the sum of MRT and mean absorption time (MAT)

If `obj$data` only contains data for one of those routes, the NCA quantities
for the "missing" route will be returned, but filled with `NA_real_.`

Additionally, for oral bolus administration, the following quantities are
estimated using [get_peak()]:

* `nca.oral.tmax`: the time at which peak concentration occurs
* `nca.oral.Cmax`: the peak concentration

If `dose_norm == TRUE`, then all data are dose-normalized before computing NCA
quantities. This means that all dose-dependent NCA quantities reflect an
estimate for a 1 mg/kg dose.

If `dose_norm == FALSE`, then all quantities are calculated separately for each dose.
}
