% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_method_dispatch.R
\name{nca}
\alias{nca}
\title{nca()}
\usage{
nca(obj, ...)
}
\arguments{
\item{obj}{An object.}
}
\description{
This is the S3 method generic for nca()
}
\seealso{
[nca.pk()] for the method for class [pk()]
}
