% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{num2string}
\alias{num2string}
\title{Converts numeric vector to comma-delimited string representation}
\usage{
num2string(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A string with comma-delimited values of `x` or an 'empty' string.
}
\description{
Converts numeric vector to comma-delimited string representation
}
\examples{
num2string(c(1,2.3))
num2string(NA_real_)


}
