% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_methods_add_pkproto.R
\name{pk_add.default}
\alias{pk_add.default}
\title{Add pkproto object default method}
\usage{
\method{pk_add}{default}(pkproto_obj, pk_obj, objectname)
}
\arguments{
\item{pkproto_obj}{The `pkproto` object to be added}

\item{pk_obj}{The `pk` object to which the `pkproto` object is to be added}

\item{objectname}{The object name}
}
\value{
The `pk` object modified by the addition.
}
\description{
Add pkproto object default method
}
