% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_methods_add_pkproto.R
\name{pk_add.pk_facet_data}
\alias{pk_add.pk_facet_data}
\title{Add facet_data()}
\usage{
\method{pk_add}{pk_facet_data}(pkproto_obj, pk_obj, objectname)
}
\arguments{
\item{pkproto_obj}{The `pk_facet_data` object to be added.}

\item{pk_obj}{The [pk()] object to which the `pk_facet_data` object will be added.}

\item{objectname}{The name of the `pk_facet_data` object.}
}
\value{
The [pk()] object, with the added `pk_facet_data`.
}
\description{
Add facet_data()
}
\author{
Caroline Ring
}
