% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_methods_add_pkproto.R
\name{pk_subtract.default}
\alias{pk_subtract.default}
\title{Subtract pkproto object default method}
\usage{
\method{pk_subtract}{default}(pkproto_obj, pk_obj, objectname)
}
\arguments{
\item{pkproto_obj}{The `pkproto` object to be subtracted}

\item{pk_obj}{The `pk` object to which the `pkproto` object is to be subtracted}

\item{objectname}{The object name}
}
\value{
The `pk` object modified by the addition.
}
\description{
Subtract pkproto object default method
}
