% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_methods_plot.R
\name{plot.pk}
\alias{plot.pk}
\title{Plot a [pk()] object.}
\usage{
\method{plot}{pk}(obj, newdata = NULL, model = NULL, method = NULL)
}
\arguments{
\item{obj}{A [pk()] object}

\item{newdata}{Optional: A `data.frame` containing new data to plot. Must
contain at least variables `Chemical`, `Species`, `Route`, `Media`, `Dose`,
`Time`, `Time.Units`, `Conc`, `Detect`, `Conc_SD`.  Default `NULL`, to use
the data in `obj$data`.}

\item{model}{Character: One or more of the models fitted. Curve fits will be
plotted for these models. Default `NULL` to plot fits for all models in
`obj$stat_model`.}

\item{method}{Character: One or more of the [optimx::optimx()] methods used.
Default `NULL` to plot fits for all methods in
`obj$optimx_settings$method`.}
}
\description{
Plot data and model fits from a [pk()] object.
}
\details{
If the [pk()] object has not been fitted, then only the data will be plotted
(because no curve fits exist).
}
