% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_conctime.R
\name{plot_conctime}
\alias{plot_conctime}
\title{plot_conctime}
\usage{
plot_conctime(PK.fit.table, data.set, model,
  mean.type = "Fitted geometric mean")
}
\arguments{
\item{PK.fit.table}{Table of parameter estimates from \code(\link(fit_all))}

\item{data.set}{A data.table of concentration vs. time data for a given
chemical}

\item{model}{Analytic model to evaluate. Currently only "1compartment" or
 "2compartment" are implemented.
#param mean.type (Defaults to fitted geometric mean")}
}
\value{
None
}
\description{
Creates a PDF of the fits from \code{\link{fit_all}} and data together
}
\author{
{Caroline Ring, John Wambaugh}
}
