% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_faceted_methods_predict.R,
%   R/pk_methods_predict.R
\name{predict.pk}
\alias{predict.pk}
\title{predict}
\usage{
\method{predict}{pk}(
  obj,
  newdata = NULL,
  model = NULL,
  method = NULL,
  type = "conc",
  exclude = TRUE,
  use_scale_conc = FALSE,
  ...
)

\method{predict}{pk}(
  obj,
  newdata = NULL,
  model = NULL,
  method = NULL,
  type = "conc",
  exclude = TRUE,
  use_scale_conc = FALSE,
  ...
)
}
\arguments{
\item{obj}{A [pk] object.}

\item{newdata}{Optional: A `data.frame` with new data for which to make
predictions. If NULL (the default), then predictions will be made for the
data in `obj$data`. `newdata` is required to contain at least the following
variables: `Time`, `Time.Units`, `Dose`, `Route`, and `Media`. If variable
`Time_trans` is not present, then `Time` will be transformed according to
the transformation in `obj$scales$time` before making predictions;
otherwise, `Time_trans` will be used to make predictions.}

\item{model}{Optional: Specify one or more of the fitted models for which to
make predictions. If NULL (the default), predictions will be returned for
all of the models in `obj$stat_model`.}

\item{method}{Optional: Specify one or more of the [optimx::optimx()] methods
for which to make predictions. If NULL (the default), predictions will be
returned for all of the models in `obj$settings_optimx$method`.}

\item{type}{Either `"conc"` (the default) or `"auc"`. `type = "conc"`
predicts concentrations; `type = "auc"` predicts area under the
concentration-time curve (AUC).}

\item{exclude}{Logical: `TRUE` to return `NA_real_` for any observations in
the data marked for exclusion (if there is a variable `exclude` in the
data, an observation is marked for exclusion when `exclude %in% TRUE`).
`FALSE` to return the prediction for each observation, regardless of
exclusion. Default `TRUE`.}

\item{use_scale_conc}{Possible values: `TRUE`, `FALSE`, or a named list with
elements `dose_norm` and `log10_trans` which themselves should be either
`TRUE` or `FALSE`. If `use_scale_conc = TRUE`, then the concentration
scaling/transformations in `obj` will be applied to both predicted and
observed concentrations before the log-likelihood is computed. If
`use_scale_conc = FALSE` (the default for this function), then no
concentration scaling or transformation will be applied before the
log-likelihood is computed. If `use_scale_conc = list(dose_norm = ...,
log10_trans = ...)`, then the specified dose normalization and/or
log10-transformation will be applied.}
}
\value{
A [tibble::tibble()] grouped by the faceting variables with a new variable ` predict ` containing the results of applying [predict.pk()] to each item in the list column of [pk()] objects, `obj$pk_object`

A named list of numeric matrixes. There is one list element named for
  each model in `obj`'s [stat_model()] element, *i.e.*, each PK model that
  was fitted to the data. Each list element is a matrix with the same number
  of rows as the data in `obj$data` (corresponding to the rows in
  `obj$data`), and as many columns as there were [optimx::optimx()] methods
  (specified in [settings_optimx()]). The column names are the method names.
  Each column contains the predictions of the model fitted by the
  corresponding method. If `use_scale_conc %in% FALSE`, these predictions are
  un-transformed concentrations in the same units as `obj$data$Conc.Units`.
  If `use_scale_conc %in% TRUE`, the predictions are transformed
  concentrations in the same units as `obj$data$Conc_trans.Units`.
}
\description{
predict for pk_faceted objects

Extract predictions from a fitted `pk` object.
}
\seealso{
Other methods for pk_faceted objects: 
\code{\link{AIC.pk}()},
\code{\link{BIC.pk}()},
\code{\link{check_needed_status.pk}()},
\code{\link{coef.pk}()},
\code{\link{coef_sd.pk}()},
\code{\link{compare_models.pk}()},
\code{\link{data_summary.pk}()},
\code{\link{eval_tkstats.pk}()},
\code{\link{fold_error.pk}()},
\code{\link{get_data.pk}()},
\code{\link{get_data_group.pk}()},
\code{\link{get_data_info.pk}()},
\code{\link{get_data_original.pk}()},
\code{\link{get_data_sigma_group.pk}()},
\code{\link{get_error_group.pk}()},
\code{\link{get_fit.pk}()},
\code{\link{get_mapping.pk}()},
\code{\link{get_nca.pk}()},
\code{\link{get_prefit.pk}()},
\code{\link{get_scale_conc.pk}()},
\code{\link{get_scale_time.pk}()},
\code{\link{get_settings_data_info.pk}()},
\code{\link{get_settings_optimx.pk}()},
\code{\link{get_settings_preprocess.pk}()},
\code{\link{get_stat_error_model.pk}()},
\code{\link{get_stat_model.pk}()},
\code{\link{get_status.pk}()},
\code{\link{get_tkstats.pk}()},
\code{\link{get_winning_model.pk}()},
\code{\link{is.pk}()},
\code{\link{logLik.pk}()},
\code{\link{nca.pk}()},
\code{\link{plot.pk}()},
\code{\link{preprocess_data.pk_faceted}()},
\code{\link{residuals.pk}()},
\code{\link{rmse.pk}()},
\code{\link{rsq.pk}()},
\code{\link{subtract_pkproto.pk}()},
\code{\link{summary.pk}()}

Other methods for fitted pk objects: 
\code{\link{AIC.pk}()},
\code{\link{BIC.pk}()},
\code{\link{coef.pk}()},
\code{\link{coef_sd.pk}()},
\code{\link{eval_tkstats.pk}()},
\code{\link{get_fit.pk}()},
\code{\link{get_tkstats.pk}()},
\code{\link{logLik.pk}()},
\code{\link{residuals.pk}()},
\code{\link{rmse.pk}()},
\code{\link{rsq.pk}()}
}
\author{
Caroline Ring
}
\concept{methods for fitted pk objects}
\concept{methods for pk_faceted objects}
