% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_methods_predict.R
\name{predict.pk}
\alias{predict.pk}
\title{Get predictions}
\usage{
\method{predict}{pk}(
  obj,
  newdata = NULL,
  model = NULL,
  method = NULL,
  type = "conc",
  exclude = TRUE,
  scale_conc = FALSE,
  ...
)
}
\arguments{
\item{obj}{A [pk] object.}

\item{newdata}{Optional: A `data.frame` with new data for which to make
predictions. If NULL (the default), then predictions will be made for the
data in `obj$data`. `newdata` is required to contain at least the following
variables: `Time`, `Dose`, `Route`, and `Media`. `Time` will be transformed
according to the transformation in `obj$scales$time` before predictions are
made.}

\item{model}{Optional: Specify one or more of the fitted models for which to
make predictions. If NULL (the default), predictions will be returned for
all of the models in `obj$stat_model`.}

\item{method}{Optional: Specify one or more of the [optimx::optimx()] methods
for which to make predictions. If NULL (the default), predictions will be
returned for all of the models in `obj$settings_optimx$method`.}

\item{type}{Either `"conc"` (the default) or `"auc"`. `type = "conc"`
predicts concentrations; `type = "auc"` predicts area under the
concentration-time curve (AUC).}

\item{exclude}{Logical: `TRUE` to return `NA_real_` for any observations in
the data marked for exclusion (if there is a variable `exclude` in the
data, an observation is marked for exclusion when `exclude %in% TRUE`).
`FALSE` to return the prediction for each observation, regardless of
exclusion. Default `TRUE`.}

\item{scale_conc}{Logical: `TRUE` to apply the concentration scaling defined
in `obj$scales$conc` to the predictions before returning them. `FALSE` to
return the un-transformed predictions. Default `FALSE`.}
}
\value{
A named list of numeric matrixes. There is one list element named for
  each model in `obj`'s [stat_model()] element, *i.e.*, each PK model that
  was fitted to the data. Each list element is a matrix with the same number
  of rows as the data in `obj$data` (corresponding to the rows in
  `obj$data`), and as many columns as there were [optimx::optimx()] methods
  (specified in [settings_optimx()]). The column names are the method names.
  Each column contains the predictions of the model fitted by the
  corresponding method. If `scale_conc %in% FALSE`, these predictions are
  un-transformed concentrations in the same units as `obj$data$Conc.Units`. If
  `scale_conc %in% TRUE`, the predictions are transformed concentrations in
  the same units as `obj$data$Conc_trans.Units`.
}
\description{
Extract predictions from a fitted `pk` object.
}
\author{
Caroline Ring
}
