% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_methods_print.R
\name{print.pk}
\alias{print.pk}
\title{Print a user-friendly version of a `pk` object}
\usage{
\method{print}{pk}(x, ...)
}
\arguments{
\item{x}{A pk object.}

\item{...}{Additional arguments. Currently not in use.}
}
\value{
Summary output
}
\description{
Prints a clear summary of `pk` object and returns it invisibly.
}
\author{
Gilberto Padilla Mercado
}
