% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_methods_residuals.R
\name{residuals.pk}
\alias{residuals.pk}
\title{Get residuals}
\usage{
\method{residuals}{pk}(
  obj,
  newdata = NULL,
  model = NULL,
  method = NULL,
  exclude = TRUE,
  scale_conc = FALSE,
  ...
)
}
\arguments{
\item{obj}{A `pk` object}

\item{newdata}{Optional: A `data.frame` with new data for which to make
predictions and compute residuals. If NULL (the default), then residuals
will be computed for the data in `obj$data`. `newdata` is required to
contain at least the following variables: `Time`, `Dose`, `Route`, `Media`,
`Conc`, `Detect`, `N_Subjects`, `Conc_SD`. `Time` will be transformed
according to the transformation in `obj$scales$time`. Residuals will be
calculated on un-transformed concentration data (i.e., ignoring any
scalings or transformations in `obj$scales$conc`).}

\item{model}{Optional: Specify one or more of the fitted models for which to
make predictions and calculate residuals. If NULL (the default), residuals
will be returned for all of the models in `obj$stat_model`.}

\item{method}{Optional: Specify one or more of the [optimx::optimx()] methods
for which to make predictions and calculate residuals. If NULL (the
default), residuals will be returned for all of the models in
`obj$optimx_settings$method`.}

\item{exclude}{Logical: `TRUE` to return `NA_real_` for any observations in
the data marked for exclusion (if there is a variable `exclude` in the
data, an observation is marked for exclusion when `exclude %in% TRUE`).
`FALSE` to return the residual for each observation, regardless of
exclusion. Default `TRUE`.}

\item{scale_conc}{Logical: `TRUE` to apply the concentration
transformation(s) defined in `obj$scales$conc` to both predictions and
observations before calculating residuals. `FALSE` to calculate residuals
for un-transformed predictions and observations. Default `FALSE`.}
}
\value{
A named list of numeric matrices. There is one list element named for
  each model in `obj`'s [stat_model()] element, i.e. each PK model that was
  fitted to the data. Each list element is a matrix with the same number of
  rows as the data in `obj$data` (corresponding to the rows in `obj$data`),
  and as many columns as there were [optimx::optimx()] methods (specified in
  [settings_optimx()]). The column names are the method names.  Each column
  contains the residuals (observed - predicted) of the model fitted by the
  corresponding method.  If `scale_conc %in% FALSE`, these residuals are in
  the same units as `obj$data$Conc.Units`. If `scale_conc %in% TRUE`, the
  residuals are in the same units as `obj$data$Conc_trans.Units`.
}
\description{
Extract residuals from a fitted `pk` object.
}
\details{
Residuals are `obs - pred` in general, where `obs` is the observed
concentration value and `pred` is the predicted concentration value.

For non-detect observations, residual is zero if `pred` is also below the
LOQ. Otherwise, the residual is the difference between the LOQ and `pred`.
}
\author{
Caroline Ring
}
