% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkproto_scale_data.R
\name{scale_conc}
\alias{scale_conc}
\title{Scale concentrations}
\usage{
scale_conc(dose_norm = FALSE, log10_trans = FALSE, ...)
}
\arguments{
\item{dose_norm}{Logical: Whether to normalize observed concentrations (and
observed concentration standard deviations and limits of quantification) by
dividing them by the corresponding dose. Default `FALSE`.}

\item{log10_trans}{Logical: Whether to apply a `log10()` transformation to
observed concentrations (and limits of quantification), after any dose
normalization is applied. Default `FALSE`.}

\item{...}{Other arguments (not currently used)}
}
\value{
An object of class `pk_scales`: A named list with two elements. The
 first element `name = "conc"` denotes the variable to be scaled. The second
 element `value` is itself a named list of the arguments supplied to
 [scale_conc()]). This object is usually added to an existing [pk()] object
 using `+`. See [pk_add.pk_scales()].
}
\description{
Scale concentrations
}
\author{
Caroline Ring
}
