% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_model.R
\name{set_params_optimize}
\alias{set_params_optimize}
\title{Set model parameters to optimize}
\usage{
set_params_optimize(model, params = "default")
}
\arguments{
\item{model}{An object of class `pk_model`.}

\item{params}{A character vector with any of the parameters in `model$params`,
or the name of any of the parameter groups in `model$param_groups`.}
}
\value{
An object of class `pk_model` with set/updated `param_fun_args` to
include `pars_to_optimize` argument specifying the parameters to be optimized.
}
\description{
Set model parameters to optimize
}
\seealso{
Other pk_model modifiers: 
\code{\link{adjust_model_name}()},
\code{\link{set_params_starts}()},
\code{\link{toggle_clearance_mode}()}
}
\author{
Gilberto Padilla Mercado
}
\concept{pk_model modifiers}
