% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkproto_stat_model.R
\name{stat_model}
\alias{stat_model}
\title{Add PK model(s) to be fitted}
\usage{
stat_model(model = c("flat", "1comp", "2comp"), ...)
}
\arguments{
\item{model}{The name(s) of models to be fitted. These should be the names of objects of class `pk_model`. Built-in options are `"flat"`,
`"1comp"`, and `"2comp"`. You may add your own model by using [pk_model()].}
}
\description{
Add PK model(s) to be fitted
}
