% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_methods_summary.R
\name{summary.pk}
\alias{summary.pk}
\title{Print summary of a `pk` object}
\usage{
\method{summary}{pk}(obj, ...)
}
\arguments{
\item{obj}{A [pk] object.}

\item{...}{Additional arguments. Currently not in use.}
}
\value{
A list of `data.frame`s consisting of a summary table of fitting options and results.
}
\description{
This summary includes summary information about the data; about any data
transformations applied; about the models being fitted; about the error model
being applied; and any fitting results, if the `pk` object has been fitted.
It also includes TK quantities calculated from the fitted model parameters,
e.g. halflife; clearance; tmax; Cmax; AUC; Css.
}
\author{
Caroline Ring
}
