% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_model.R
\name{toggle_clearance_mode}
\alias{toggle_clearance_mode}
\title{Switch between model parameters to optimize}
\usage{
toggle_clearance_mode(model)
}
\arguments{
\item{model}{An object of class `pk_model`.}
}
\value{
An object of class `pk_model` with set/updated `param_fun_args` to
include `pars_to_optimize` argument specifying the parameters to be optimized.
}
\description{
Switch between model parameters to optimize
}
\author{
Gilberto Padilla Mercado
}
