% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_method_dispatch.R
\name{twofold_test}
\alias{twofold_test}
\title{twofold_test()}
\usage{
twofold_test(obj, ...)
}
\arguments{
\item{obj}{An object.}

\item{...}{Additional arguments currently not in use.}
}
\description{
This is the S3 method generic for twofold_test()
}
\seealso{
[twofold_test.pk()] for the method for class [pk()]
}
